### 2025-09-19 #9022 @editorialbot
[PRE REVIEW]: JustRelax.jl: A Julia package for geodynamic modeling with matrix-free solvers
**Submitting author:** @albert-de-montserrat (<a href="http://orcid.org/0000-0003-1694-3735">Albert de Montserrat</a>)
**Repository:** https://github.com/PTsolvers/JustRelax.jl
**Branch with paper.md** (empty if default branch): adm/JOSS
**Version:** v0.4.2
