### 2025-09-23 d63ca8f
update `init_particles` call to newest JP
