+++ miniapps/subduction/3D/Subduction3D_MPI.jl
@@ -50,7 +50,9 @@ function main3D(x_global, y_global, z_global, li, origin, phases_GMG, igg; nx = 
 
     # Initialize particles -------------------------------
     nxcell, max_xcell, min_xcell = 150, 175, 125
-    particles = init_particles(backend_JP, nxcell, max_xcell, min_xcell, xvi, di, ni)
+       particles = init_particles(
+        backend_JP, nxcell, max_xcell, min_xcell, xvi...
+    )
     subgrid_arrays = SubgridDiffusionCellArrays(particles)
     # velocity grids
     grid_vx, grid_vy, grid_vz = velocity_grids(xci, xvi, di)
