### 2025-09-23 1aa02ee
add checkpointing function that can deal with only stokes arrays (thermal optional); fix bugs in tests
