+++ test/Project.toml
@@ -2,6 +2,7 @@
 CellArrays = "d35fcfd7-7af4-4c67-b1aa-d78070614af4"
 GeoParams = "e018b62d-d9de-4a26-8697-af89c310ae38"
 GeophysicalModelGenerator = "3700c31b-fa53-48a6-808a-ef22d5a84742"
+JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 JustPIC = "10dc771f-8528-4cd9-9d3b-b21b2e693339"
 MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
 ParallelStencil = "94395366-693c-11ea-3b26-d9b7aac5d958"
