+++ miniapps/subduction/2D/Subduction2D_restart.jl
@@ -76,55 +76,50 @@ function main(li, origin, phases_GMG, igg; nx = 16, ny = 16, figdir = "figs2D", 
     dt = 10.0e3 * 3600 * 24 * 365 # diffusive CFL timestep limiter
     # ----------------------------------------------------
 
-    # Initialize particles -------------------------------
-    nxcell = 40
-    max_xcell = 60
-    min_xcell = 20
-    particles = init_particles(
-        backend_JP, nxcell, max_xcell, min_xcell, xvi...
-    )
-    subgrid_arrays = SubgridDiffusionCellArrays(particles)
-    # velocity grids
-    grid_vxi = velocity_grids(xci, xvi, di)
-    # material phase & temperature
-    pPhases, pT = init_cell_arrays(particles, Val(2))
+    # Load particles -------------------------------
+    data            = load(joinpath(figdir, "checkpoint", "particles.jld2"))
+    particles       = TA(backend)(Float64, data["particles"])
+    pPhases         = TA(backend)(Float64, data["phases"])
+    phase_ratios    = TA(backend)(Float64, data["phase_ratios"])
+    particle_args   = TA(backend).(Float64, data["particle_args"])
+    particle_args_reduced = TA(backend).(Float64, data["particle_args_reduced"])
+
+    subgrid_arrays  = SubgridDiffusionCellArrays(particles)
+    # velocity staggered grids
+    grid_vxi        = velocity_grids(xci, xvi, di)
 
     # particle fields for the stress rotation
     pτ = StressParticles(particles)
-    particle_args = (pT, pPhases, unwrap(pτ)...)
-    particle_args_reduced = (pT, unwrap(pτ)...)
+    pT = init_cell_arrays(particles, Val(1))
 
-    # Assign particles phases anomaly
-    phases_device = PTArray(backend)(phases_GMG)
-    phase_ratios = phase_ratios = PhaseRatios(backend_JP, length(rheology), ni)
-    init_phases!(pPhases, phases_device, particles, xvi)
+    # Update phase ratios based on loaded particles
     update_phase_ratios!(phase_ratios, particles, xci, xvi, pPhases)
     # ----------------------------------------------------
 
     # STOKES ---------------------------------------------
-    # Allocate arrays needed for every Stokes problem
-    stokes = StokesArrays(backend, ni)
+    # Load Stokes arrays
+    !isfile(joinpath(figdir, "checkpoint", "checkpoint.jld2")) && error("Checkpoint file not found. Please check the path.")
+    stokes_cpu, thermal_cpu, t, dt = load_checkpoint_jld2(joinpath(figdir, "checkpoint"))
+    stokes = PTArray(backend_JR, stokes_cpu)
     pt_stokes = PTStokesCoeffs(li, di; ϵ_abs = 1.0e-4, ϵ_rel = 1.0e-4, Re = 1.0e0, r = 0.7, CFL = 0.9 / √2.1) # Re=3π, r=0.7
     # ----------------------------------------------------
 
     # TEMPERATURE PROFILE --------------------------------
     Ttop = 20 + 273
     Tbot = maximum(T_GMG)
-    thermal = ThermalArrays(backend, ni)
-    @views thermal.T[2:(end - 1), :] .= PTArray(backend)(T_GMG)
+    # Load thermal arrays
+    thermal = PTArray(backend_JR, thermal_cpu)
+
     thermal_bc = TemperatureBoundaryConditions(;
         no_flux = (left = true, right = true, top = false, bot = false),
     )
     thermal_bcs!(thermal, thermal_bc)
-    @views thermal.T[:, end] .= Ttop
-    @views thermal.T[:, 1] .= Tbot
[…]
