+++ docs/make.jl
@@ -110,6 +110,10 @@ makedocs(;
                 "Rheology" => "man/plume3D/rheology.md",
                 "Setting up the model" => "man/plume3D/plume3D.md",
             ],
+            "Checkpointing/Restart" => Any[
+                "Checkpointing" => "man/checkpointing.md",
+                "Restart" => "man/restart.md",
+            ]
         ],
         "List of functions" => "man/listfunctions.md",
         "References" => Any[
