### 2025-09-24 a54c9f3
update phase_ratio_arrays functions (CPU) and miniapp
