+++ src/phases/PhaseRatios.jl
@@ -20,7 +20,7 @@ phase_arrays = (phase_1, phase_2)
 update_phase_ratios!(phase_ratios, phase_arrays, xci, xvi)
 ```
 """
-function update_phase_ratios!(
+function update_phase_ratios_2D!(
         phase_ratios::JustPIC.PhaseRatios{B, T}, phase_arrays::NTuple{N, AbstractMatrix}, xci, xvi
     ) where {B, T <: AbstractMatrix, N}
 
@@ -56,8 +56,8 @@ phase_arrays = (phase_1, phase_2)
 update_phase_ratios!(phase_ratios, phase_arrays, xci, xvi)
 ```
 """
-function update_phase_ratios!(
-        phase_ratios::JustPIC.PhaseRatios{B, T}, phase_arrays::NTuple{N, AbstractArray}, xci, xvi
+function update_phase_ratios_3D!(
+        phase_ratios::JustPIC.PhaseRatios{B, T}, phase_arrays::NTuple{N, AbstractMatrix}, xci, xvi
     ) where {B, T <: AbstractArray, N}
 
     phase_ratios_center_from_arrays!(phase_ratios, phase_arrays)
@@ -89,24 +89,25 @@ end
     ) where {N}
 
     values = ntuple(i -> phase_arrays[i][I...], Val(N))
-    total = sum(values)
 
-    # Normalize (handle case where total might be zero)
-    if total > eps(typeof(total))
-        normalized_values = clamp.(values ./ total, 0.0, 1.0)
-        # Clean up very small values (round to zero if < 1e-5)
-        cleaned_values = ntuple(i -> normalized_values[i] < 1e-5 ? zero(eltype(normalized_values)) : normalized_values[i], Val(N))
-        # Renormalize to ensure sum = 1
-        final_total = sum(cleaned_values)
-        if final_total > eps(typeof(final_total))
-            normalized_values = cleaned_values ./ final_total
-        else
-            normalized_values = ntuple(_ -> one(eltype(values)) / N, Val(N))
-        end
-    else
-        # If all values are zero, distribute equally
-        normalized_values = ntuple(_ -> one(eltype(values)) / N, Val(N))
+    total = zero(eltype(values[1]))
+    for i in 1:N
+        total += values[i]
     end
+
+    # Normalize
+    normalized_values = ntuple(i -> values[i] / total, Val(N))
+    # Manual clamping
+    normalized_values = ntuple(i -> min(max(normalized_values[i], 0.0), 1.0), Val(N))
+    # Clean up very small values (round to zero if < 1e-5)
+    cleaned_values = ntuple(i -> normalized_values[i] < 1.0e-5 ? zero(eltype(normalized_values)) : normalized_values[i], Val(N))
+    # Renormalize to ensure sum = 1
+    final_total = zero(eltype(values[1]))  # Fixed: was 'value[1]'
+    for i in 1:N
+        final_total += cleaned_values[i]
+    end
+
+    normalized_values = ntuple(i -> cleaned_values[i] / final_total, Val(N))
 
     # Update phase ratios array
     for k in 1:N
@@ -116,6 +117,9 @@ end
     return nothing
 end
 
+## ============================================================================
+## VERTEX VALUES
+## ============================================================================
 
 function phase_ratios_vertex_from_arrays!(
         phase_ratios::JustPIC.PhaseRatios, phase_arrays::NTuple{N, AbstractArray}, xvi::NTuple{ND}, xci::NTuple{ND}
@@ -141,6 +145,7 @@ end
         total_weight = zero(T)
         w_vals = ntuple(_ -> zero(T), Val(N))
         x_v, y_v = cell_vertex
+
         for offset₁ in -1:0, offset₂ in -1:0
             i_cell = I[1] + offset₁
[…]
