+++ src/ext/CUDA/2D.jl
@@ -26,7 +26,7 @@ import JustRelax:
 
 import JustRelax: normal_stress, shear_stress, shear_vorticity, unwrap
 
-import JustPIC._2D: numphases, nphases, PhaseRatios, update_phase_ratios!, compute_dx, face_offset
+import JustPIC._2D: numphases, nphases#, PhaseRatios, update_phase_ratios!, compute_dx, face_offset
 
 __init__() = @init_parallel_stencil(CUDA, Float64, 2)
 
@@ -474,16 +474,12 @@ end
 
 # Phase ratios with arrays
 
-# function JR2D.update_phase_ratios!(
-#         phase_ratios::JustPIC.PhaseRatios{CUDABackend, T}, phase_arrays, xci, xvi
-#     ) where {T <: CuArray}
-#     update_phase_ratios!(phase_ratios, phase_arrays, xci, xvi)
-#     return nothing
-# end
-
-function JR2D.update_phase_ratios!(
-        phase_ratios::JustPIC.PhaseRatios{CUDABackend, T}, phase_arrays, xci, xvi
-    ) where {T <: AbstractMatrix}
+function JR2D.update_phase_ratios_2D!(
+        phase_ratios::JustPIC.PhaseRatios{CUDABackend, T},
+        phase_arrays::NTuple{N, CuArray{U, 2}},
+        xci,
+        xvi
+    ) where {T <: AbstractMatrix, N, U}
     phase_ratios_center_from_arrays!(phase_ratios, phase_arrays)
     phase_ratios_vertex_from_arrays!(phase_ratios, phase_arrays, xvi, xci)
     # velocity nodes
@@ -491,42 +487,5 @@ function JR2D.update_phase_ratios!(
     phase_ratios_face_from_arrays!(phase_ratios.Vy, phase_arrays, xci, :y)
     return nothing
 end
-
-# function JR2D.phase_ratios_center_from_arrays!(phase_ratios::JustPIC.PhaseRatios{CUDABackend}, phase_arrays)
-#     ni = size(first(phase_arrays))
-
-#     @parallel (@idx ni) phase_ratios_center_from_arrays_kernel!(
-#         phase_ratios.center, phase_arrays
-#     )
-#     return nothing
-# end
-
-# function JR2D.phase_ratios_vertex_from_arrays!(
-#         phase_ratios::JustPIC.PhaseRatios{CUDABackend}, phase_arrays::NTuple{N, AbstractMatrix}, xvi::NTuple{ND}, xci::NTuple{ND}
-#     ) where {N, ND}
-
-#     ni = size(first(phase_arrays)) .+1
-#     di = compute_dx(xvi)
-
-#     @parallel (@idx ni) phase_ratios_vertex_from_arrays_kernel!(
-#         phase_ratios.vertex, phase_arrays, xci, xvi, di
-#     )
-#     return nothing
-# end
-
-# function JR2D.phase_ratios_face_from_arrays!(
-#         phase_face::JustPIC.PhaseRatios{CUDABackend}, phase_arrays::NTuple{N, AbstractMatrix}, xci::NTuple{ND}, dimension::Symbol
-#     ) where {N, ND}
-#     ni = size(first(phase_arrays))  # Cell grid size
-#     face_size = size(phase_face)  # Face grid size (including phase dimension)
-#     # ni_face = face_size[2:end]  # Face grid size excluding phase dimension
-#     di = compute_dx(xci)
-#     offsets = face_offset(Val(ND), dimension)
-
-#     @parallel (@idx ni) phase_ratios_face_from_arrays_kernel!(
-#         phase_face, phase_arrays, xci, di, offsets, ni
-#     )
-#     return nothing
-# end
 
 end

[…]
