### 2025-09-24 697e010
fix IO test
