+++ test/test_IO.jl
@@ -62,15 +62,18 @@ using WriteVTK, JLD2
         @test isfile(joinpath(dst, "Project.toml"))
 
         # Call the function
-        checkpointing_jld2(dst, stokes, thermal, time, dt, igg)
         checkpointing_jld2(dst, stokes, thermal, time, dt)
+        # Check that the file was created
+        fname = joinpath(dst, "checkpoint.jld2")
+        @test isfile(fname)
 
+        checkpointing_jld2(dst, stokes, thermal, time, dt, igg)
         # Check that the file was created
         fname = joinpath(dst, "checkpoint" * lpad("$(igg.me)", 4, "0") * ".jld2")
         @test isfile(fname)
 
         # Load the data from the file
-        stokes1, thermal1, t, dt1 = load_checkpoint_jld2(fname)
+        stokes1, thermal1, t, dt1 = load_checkpoint_jld2(dst, igg)
 
         @test stokes1.viscosity.η[1] == 1.0
         @test stokes1.V.Vy[1] == 10
@@ -194,15 +197,19 @@ using WriteVTK, JLD2
 
 
         # Call the function
-        checkpointing_jld2(dst, stokes, thermal, time, dt, igg)
         checkpointing_jld2(dst, stokes, thermal, time, dt)
+        # Check that the file was created
+        fname = joinpath(dst, "checkpoint.jld2")
+        @test isfile(fname)
+
+        checkpointing_jld2(dst, stokes, thermal, time, dt, igg)
 
         # Check that the file was created
         fname = joinpath(dst, "checkpoint" * lpad("$(igg.me)", 4, "0") * ".jld2")
         @test isfile(fname)
 
         # Load the data from the file
-        stokes, thermal, time, dt = load_checkpoint_jld2(fname)
+        stokes, thermal, time, dt = load_checkpoint_jld2(dst, igg)
 
         @test stokes.viscosity.η[1] == 1.0
         @test stokes.V.Vy[1] == 10
@@ -210,13 +217,18 @@ using WriteVTK, JLD2
         @test !isnothing(stokes.V.Vz)
 
         checkpointing_jld2(dst, stokes, time, dt)
+
+        # Check that the file was created
+        fname = joinpath(dst, "checkpoint.jld2")
+        @test isfile(fname)
+
         checkpointing_jld2(dst, stokes, time, dt, igg)
 
         # Check that the file was created
         fname = joinpath(dst, "checkpoint" * lpad("$(igg.me)", 4, "0") * ".jld2")
         @test isfile(fname)
         # Load the data from the file
-        stokes, _, time, dt = load_checkpoint_jld2(fname)
+        stokes, _, time, dt = load_checkpoint_jld2(dst, igg)
 
         @test stokes.viscosity.η[1] == 1.0
         @test stokes.V.Vy[1] == 10
