### 2025-09-24 ce14aa7
add GMG to dependency check
