+++ .github/workflows/Dependency.yml
@@ -9,10 +9,14 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v4
 
-      - name: Check for GLMakie and JustPIC dependencies
+      - name: Check for GLMakie and GMG dependencies
         run: |
           if grep -q "GLMakie" ./Project.toml; then
             echo "GLMakie dependency found, failing the test."
             exit 1
           fi
-          echo "Neither GLMakie dependencies found."
+          if grep -q "GeophysicalModelGenerator" ./Project.toml; then
+            echo "GeophysicalModelGenerator dependency found, failing the test."
+            exit 1
+          fi
+          echo "Neither GLMakie or GMG dependencies found."
