### 2025-09-24 c1b9765
fix gpu import of `compute_dx`
