+++ src/ext/CUDA/2D.jl
@@ -26,7 +26,7 @@ import JustRelax:
 
 import JustRelax: normal_stress, shear_stress, shear_vorticity, unwrap
 
-import JustPIC._2D: numphases, nphases#, PhaseRatios, update_phase_ratios!, compute_dx, face_offset
+import JustPIC._2D: numphases, nphases, PhaseRatios, update_phase_ratios!, compute_dx, face_offset
 
 __init__() = @init_parallel_stencil(CUDA, Float64, 2)
 
