### 2025-09-26 780a798
non dimensionalize 3D subduction miniapp
