+++ miniapps/subduction/3D/Subduction3D_MPI.jl
@@ -1,19 +1,37 @@
-# using CUDA
-using JustRelax, JustRelax.JustRelax3D, JustRelax.DataIO
-const backend_JR = CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
-# const backend_JR = CUDABackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+# Load script dependencies
+using GeoParams, GLMakie
 
-using ParallelStencil
-using ParallelStencil.FiniteDifferences3D
-# @init_parallel_stencil(CUDA, Float64, 3)
-@init_parallel_stencil(Threads, Float64, 3)
+const isCUDA = false
+
+@static if isCUDA
+    using CUDA
+end
+
+using JustRelax, JustRelax.JustRelax3D, JustRelax.DataIO
+
+const backend_JR = @static if isCUDA
+    CUDABackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+else
+    JustRelax.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+end
+
+using ParallelStencil, ParallelStencil.FiniteDifferences3D
+
+@static if isCUDA
+    @init_parallel_stencil(CUDA, Float64, 3)
+else
+    @init_parallel_stencil(Threads, Float64, 3)
+end
 
 using JustPIC, JustPIC._3D
-# const backend_JP = CUDABackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
-const backend_JP = JustPIC.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
-
-# Load script dependencies
-using Printf, LinearAlgebra, GeoParams, CairoMakie
+# Threads is the default backend,
+# to run on a CUDA GPU load CUDA.jl (i.e. "using CUDA") at the beginning of the script,
+# and to run on an AMD GPU load AMDGPU.jl (i.e. "using AMDGPU") at the beginning of the script.
+const backend_JP = @static if isCUDA
+    CUDABackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+else
+    JustPIC.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+end
 
 # Load file with all the rheology configurations
 include("Subduction3D_rheology.jl")
@@ -36,16 +54,24 @@ end
 
 ## BEGIN OF MAIN SCRIPT --------------------------------------------------------------
 function main3D(x_global, y_global, z_global, li, origin, phases_GMG, igg; nx = 16, ny = 16, nz = 16, figdir = "figs3D", do_vtk = false)
+    thickness = 660e3 * m
+    η0        = 1.0e20 * Pa * s
+    CharDim   = GEO_units(;
+        length = thickness, viscosity = η0, temperature = 1000 * K
+    )
+    li_nd     = nondimensionalize(li .* m, CharDim)
+    origin_nd = nondimensionalize(origin .* m, CharDim)
+   
     # LOCAL Physical domain ------------------------------------
     ni = nx, ny, nz           # number of cells
-    di = @. li / (nx_g(), ny_g(), nz_g())           # grid steps
-    grid = Geometry(ni, li; origin = origin)
-    (; xci, xvi) = grid # nodes at the center and vertices of the cells
+    di = @. li_nd / (nx_g(), ny_g(), nz_g())           # grid steps
+    grid         = Geometry(ni, li_nd; origin = origin_nd)
+    (; xci, xvi) = grid              # nodes at the center and vertices of the cells
     # ----------------------------------------------------
[…]
