+++ miniapps/subduction/3D/Subduction3D.jl
@@ -54,29 +54,29 @@ end
 ## BEGIN OF MAIN SCRIPT --------------------------------------------------------------
 function main3D(li, origin, phases_GMG, igg; nx = 16, ny = 16, nz = 16, figdir = "figs3D", do_vtk = false)
 
-    thickness = 660e3 * m
-    η0        = 1.0e20 * Pa * s
-    CharDim   = GEO_units(;
+    thickness = 660.0e3 * m
+    η0 = 1.0e20 * Pa * s
+    CharDim = GEO_units(;
         length = thickness, viscosity = η0, temperature = 1000 * K
     )
-    li_nd     = nondimensionalize(li .* m, CharDim)
+    li_nd = nondimensionalize(li .* m, CharDim)
     origin_nd = nondimensionalize(origin .* m, CharDim)
-   
+
     # Physical domain ------------------------------------
-    ni           = nx, ny, nz        # number of cells
-    di           = @. li_nd / ni        # grid steps
-    grid         = Geometry(ni, li_nd; origin = origin_nd)
+    ni = nx, ny, nz        # number of cells
+    di = @. li_nd / ni        # grid steps
+    grid = Geometry(ni, li_nd; origin = origin_nd)
     (; xci, xvi) = grid              # nodes at the center and vertices of the cells
     # ----------------------------------------------------
 
     # Physical properties using GeoParams ----------------
-    rheology  = init_rheologies(CharDim)
-    dt        = nondimensionalize(10.0e3 * yr, CharDim) # diffusive CFL timestep limiter
+    rheology = init_rheologies(CharDim)
+    dt = nondimensionalize(10.0e3 * yr, CharDim) # diffusive CFL timestep limiter
     # ----------------------------------------------------
 
     # Initialize particles -------------------------------
     nxcell, max_xcell, min_xcell = 150, 175, 125
-    particles      = init_particles(backend_JP, nxcell, max_xcell, min_xcell, xvi, di, ni)
+    particles = init_particles(backend_JP, nxcell, max_xcell, min_xcell, xvi, di, ni)
     subgrid_arrays = SubgridDiffusionCellArrays(particles)
     # velocity grids
     grid_vx, grid_vy, grid_vz = velocity_grids(xci, xvi, di)
@@ -85,36 +85,36 @@ function main3D(li, origin, phases_GMG, igg; nx = 16, ny = 16, nz = 16, figdir =
 
     # Assign particles phases anomaly
     phases_device = PTArray(backend_JR)(phases_GMG)
-    phase_ratios  = PhaseRatios(backend_JP, length(rheology), ni)
+    phase_ratios = PhaseRatios(backend_JP, length(rheology), ni)
     init_phases!(pPhases, phases_device, particles, xvi)
     update_phase_ratios!(phase_ratios, particles, xci, xvi, pPhases)
     # ----------------------------------------------------
 
     # STOKES ---------------------------------------------
     # Allocate arrays needed for every Stokes problem
-    stokes     = StokesArrays(backend_JR, ni)
-    pt_stokes  = PTStokesCoeffs(li_nd, di; ϵ_abs = 5.0e-3, ϵ_rel = 5.0e-3, CFL = 0.99 / √3.1)
+    stokes = StokesArrays(backend_JR, ni)
+    pt_stokes = PTStokesCoeffs(li_nd, di; ϵ_abs = 5.0e-3, ϵ_rel = 5.0e-3, CFL = 0.99 / √3.1)
     # ----------------------------------------------------
 
     # TEMPERATURE PROFILE --------------------------------
-    thermal    = ThermalArrays(backend_JR, ni)
+    thermal = ThermalArrays(backend_JR, ni)
[…]
