+++ miniapps/subduction/3D/Subduction3D_rheology.jl
@@ -3,39 +3,39 @@ function init_rheologies(CharDim)
     return rheology = (
         # Name              = "slab",
         SetMaterialParams(;
-            Phase            = 1,
-            Density          = ConstantDensity(; ρ = 3.28e3kg/m^3),
-            CompositeRheology= CompositeRheology((LinearViscous(η = 2.0e23*Pa*s),)),
+            Phase = 1,
+            Density = ConstantDensity(; ρ = 3.28e3kg / m^3),
+            CompositeRheology = CompositeRheology((LinearViscous(η = 2.0e23 * Pa * s),)),
             # Elasticity     = el_upper_crust,
-            Gravity          = ConstantGravity(; g = 9.81m/s^2),
-            CharDim          = CharDim,
+            Gravity = ConstantGravity(; g = 9.81m / s^2),
+            CharDim = CharDim,
         ),
         # Name              = "crust",
         SetMaterialParams(;
-            Phase            = 2,
-            Density          = ConstantDensity(; ρ = 3.28e3kg/m^3),
-            CompositeRheology= CompositeRheology((LinearViscous(η = 1.0e21*Pa*s),)),
+            Phase = 2,
+            Density = ConstantDensity(; ρ = 3.28e3kg / m^3),
+            CompositeRheology = CompositeRheology((LinearViscous(η = 1.0e21 * Pa * s),)),
             # Elasticity     = el_upper_crust,
-            Gravity          = ConstantGravity(; g = 9.81m/s^2),
-            CharDim          = CharDim,
+            Gravity = ConstantGravity(; g = 9.81m / s^2),
+            CharDim = CharDim,
         ),
         # Name              = "mantle",
         SetMaterialParams(;
-            Phase            = 3,
-            Density          = ConstantDensity(; ρ = 3.2e3kg/m^3),
-            CompositeRheology= CompositeRheology((LinearViscous(η = 1.0e21*Pa*s),)),
+            Phase = 3,
+            Density = ConstantDensity(; ρ = 3.2e3kg / m^3),
+            CompositeRheology = CompositeRheology((LinearViscous(η = 1.0e21 * Pa * s),)),
             # Elasticity     = el_upper_crust,
-            Gravity          = ConstantGravity(; g = 9.81m/s^2),
-            CharDim          = CharDim,
+            Gravity = ConstantGravity(; g = 9.81m / s^2),
+            CharDim = CharDim,
         ),
         # Name              = "StickyAir",
         SetMaterialParams(;
-            Phase            = 4,
-            Density          = ConstantDensity(; ρ = 100kg/m^3), # water density
-            HeatCapacity     = ConstantHeatCapacity(; Cp = 3.0e3J/kg/K),
-            Conductivity     = ConstantConductivity(; k = 1.0Watt/K/m),
-            CompositeRheology= CompositeRheology((LinearViscous(; η = 1.0e19*Pa*s),)),
-            CharDim          = CharDim,
+            Phase = 4,
+            Density = ConstantDensity(; ρ = 100kg / m^3), # water density
+            HeatCapacity = ConstantHeatCapacity(; Cp = 3.0e3J / kg / K),
+            Conductivity = ConstantConductivity(; k = 1.0Watt / K / m),
+            CompositeRheology = CompositeRheology((LinearViscous(; η = 1.0e19 * Pa * s),)),
+            CharDim = CharDim,
         ),
     )
 end

[…]
