+++ src/variational_stokes/Stokes2D.jl
@@ -44,8 +44,8 @@ function _solve_VS!(
     (; η, η_vep) = stokes.viscosity
     ni = size(stokes.P)
 
-    nRx = length(@views stokes.R.Rx[ϕ.Vx[2:end-1,:] .> 0])
-    nRy = length(@views stokes.R.Ry[ϕ.Vy[:, 2:end-1] .> 0])
+    nRx = length(@views stokes.R.Rx[ϕ.Vx[2:(end - 1), :] .> 0])
+    nRy = length(@views stokes.R.Ry[ϕ.Vy[:, 2:(end - 1)] .> 0])
     nRP = length(@views stokes.R.RP[ϕ.center .> 0])
 
     # ~preconditioner
@@ -230,9 +230,9 @@ function _solve_VS!(
             #     norm_mpi(@views stokes.R.RP[ϕ.center .> 0]) /
             #         √length(@views stokes.R.RP[ϕ.center .> 0]),
             # )
-             errs = (
-                norm_mpi(@views stokes.R.Rx[ϕ.Vx[2:end-1,:] .> 0]) / nRx,
-                norm_mpi(@views stokes.R.Ry[ϕ.Vy[:,2:end-1] .> 0]) / nRy,
+            errs = (
+                norm_mpi(@views stokes.R.Rx[ϕ.Vx[2:(end - 1), :] .> 0]) / nRx,
+                norm_mpi(@views stokes.R.Ry[ϕ.Vy[:, 2:(end - 1)] .> 0]) / nRy,
                 norm_mpi(@views stokes.R.RP[ϕ.center .> 0]) / nRP,
             )
 
