### 2025-10-02 81112fa
Update terminology from 'visco-elasto-viscoelastic' to 'visco-elasto-viscoplastic'
