### 2025-10-02 63f447e
Remove the obvious
