### 2025-10-06 874c797
Stokes Benchmarks - fix errors
