+++ miniapps/benchmarks/stokes2D/solvi/SolVi.jl
@@ -51,8 +51,8 @@ function solVi(;
     ni = nx, ny # number of nodes in x- and y-
     li = lx, ly  # domain length in x- and y-
     origin = 0, 0
-    igg = IGG(init_global_grid(nx, ny, 1; init_MPI = true)...) #init MPI
-    # igg          = IGG(init_global_grid(nx, ny, 1; init_MPI=init_MPI)...) #init MPI
+    # igg = IGG(init_global_grid(nx, ny, 1; init_MPI = true)...) #init MPI
+    igg          = IGG(init_global_grid(nx, ny, 1; init_MPI=init_MPI)...) #init MPI
     di = @. li / (nx_g(), ny_g()) # grid step in x- and -y
     grid = Geometry(ni, li; origin = origin)
     (; xci, xvi) = grid # nodes at the center and vertices of the cells
