+++ miniapps/benchmarks/stokes2D/free_surface_stabilization/Crameri2D.jl
@@ -143,7 +143,7 @@ function main(igg, nx, ny)
     # RockRatios
     air_phase = 1
     ϕ = RockRatio(backend, ni)
-    update_rock_ratio!(ϕ, phase_ratios, (phase_ratios.Vx, phase_ratios.Vy), air_phase)
+    update_rock_ratio!(ϕ, phase_ratios, air_phase)
     # ----------------------------------------------------
 
     # STOKES ---------------------------------------------
@@ -161,7 +161,7 @@ function main(igg, nx, ny)
     args = (; T = thermal.Tc, P = stokes.P, dt = Inf)
     compute_ρg!(ρg[2], phase_ratios, rheology, args)
     stokes.P .= PTArray(backend)(reverse(cumsum(reverse((ρg[2]) .* di[2], dims = 2), dims = 2), dims = 2))
-    compute_viscosity!(stokes, phase_ratios, args, rheology, air_phase, (1.0e18, 1.0e24))
+    compute_viscosity!(stokes, phase_ratios, args, rheology, (1.0e18, 1.0e24); air_phase = air_phase)
 
     # Boundary conditions
     flow_bcs = VelocityBoundaryConditions(;
@@ -174,11 +174,14 @@ function main(igg, nx, ny)
     Vy_v = @zeros(ni .+ 1...)
 
     figdir = "Crameri2012"
+    vtkdir = joinpath(figdir, "vtk")
     take(figdir)
+    take(vtkdir)
 
     # Time loop
     t, it = 0.0, 0
     dt = 10.0e3 * (3600 * 24 * 365.25)
+    dt_max = dt
 
     while it < 20
 
@@ -215,117 +218,70 @@ function main(igg, nx, ny)
         inject_particles_phase!(particles, pPhases, (), (), xvi)
         # update phase ratios
         update_phase_ratios!(phase_ratios, particles, xci, xvi, pPhases)
-        update_rock_ratio!(ϕ, phase_ratios, (phase_ratios.Vx, phase_ratios.Vy), air_phase)
+        update_rock_ratio!(ϕ, phase_ratios, air_phase)
 
         @show it += 1
         t += dt
 
 
         (; η_vep, η) = stokes.viscosity
-        # if do_vtk
+        if do_vtk && (it == 1 || rem(it, 1) == 0)
         velocity2vertex!(Vx_v, Vy_v, @velocity(stokes)...)
         velocity_v = @. √(Vx_v^2 .+ Vy_v^2)
         data_v = (;
-            τII = Array(stokes.τ.II),
-            εII = Array(stokes.ε.II),
             Vx = Array(Vx_v),
             Vy = Array(Vy_v),
             Vel = Array(velocity_v),
         )
         data_c = (;
+            stress_II = Array(stokes.τ.II),
+            strain_rate_II = Array(stokes.ε.II),
             P = Array(stokes.P),
-            η = Array(η_vep),
+            visc_eff = Array(η_vep),
         )
         velocity_v = (
             Array(Vx_v),
             Array(Vy_v),
         )
         JustRelax.DataIO.save_vtk(
-            joinpath(figdir, "vtk_" * lpad("$it", 6, "0")),
+            joinpath(vtkdir, "vtk_" * lpad("$it", 6, "0")),
             xvi,
             xci,
             data_v,
             data_c,
             velocity_v;
-            t = t
+            t = t,
+            pvd = joinpath(vtkdir, "Crameri2012")
         )
-        # end
+        end
 
[…]
