### 2025-10-06 39f1f2b
update some scripts
