+++ miniapps/benchmarks/stokes2D/free_surface_stabilization/PlumeFreeSurface_VariationalStokes.jl
@@ -249,7 +249,7 @@ function main(igg, nx, ny)
             chain_y = chain.coords[2].data[:] ./ 1.0e3
 
             scatter!(ax, Array(pxv[idxv]), Array(pyv[idxv]), color = Array(clr[idxv]), markersize = 5)
-            arrows!(
+            arrows2d!(
                 ax,
                 xvi[1][1:nt:(end - 1)] ./ 1.0e3, xvi[2][1:nt:(end - 1)] ./ 1.0e3, Array.((Vx_v[1:nt:(end - 1), 1:nt:(end - 1)], Vy_v[1:nt:(end - 1), 1:nt:(end - 1)]))...,
                 lengthscale = 25 / max(maximum(Vx_v), maximum(Vy_v)),
