+++ miniapps/benchmarks/stokes2D/VanKeken/VanKeken.jl
@@ -79,7 +79,7 @@ function main2D(igg; ny = 64, nx = 64, figdir = "model_figs")
     # Initialize particles -------------------------------
     nxcell, max_p, min_p = 20, 30, 10
     particles = init_particles(
-        backend, nxcell, max_p, min_p, xvi, di, ni
+        backend, nxcell, max_p, min_p, xvi...
     )
     # velocity grids
     grid_vx, grid_vy = velocity_grids(xci, xvi, di)
@@ -157,7 +157,7 @@ function main2D(igg; ny = 64, nx = 64, figdir = "model_figs")
 
         # Compute U rms ---------------
         Urms_it = let
-            velocity2vertex!(Vx_v, Vy_v, stokes.V.Vx, stokes.V.Vy; ghost_nodes = true)
+            velocity2vertex!(Vx_v, Vy_v, stokes.V.Vx, stokes.V.Vy)
             @. Vx_v .= hypot.(Vx_v, Vy_v) # we reuse Vx_v to store the velocity magnitude
             sum(Vx_v .^ 2) * prod(di) |> sqrt
         end
@@ -179,9 +179,25 @@ function main2D(igg; ny = 64, nx = 64, figdir = "model_figs")
 
         # Plotting ---------------------
         if it == 1 || rem(it, 25) == 0 || t >= tmax
-            fig = Figure(size = (1000, 1000), title = "t = $t")
-            ax1 = Axis(fig[1, 1], aspect = 1 / λ, title = "t=$t")
-            heatmap!(ax1, xvi[1], xvi[2], Array(ρg[2]), colormap = :oleron)
+            fig = Figure(size = (1000, 1000), font = "TeX Gyre Heros Makie")
+            ax1 = Axis(fig[1:2, 1], aspect = 1 / λ, title = "VanKeken ",
+                titlesize = 20,
+                yticklabelsize = 12,
+                xticklabelsize = 12,
+                xlabelsize = 12,
+                ylabelsize = 12
+            )
+            h=heatmap!(ax1, xvi[1], xvi[2], Array(ρg[2]), colormap = :lapaz)
+            Colorbar(fig[1:2, 2], h; height = Relative(1.0), label = "Density", labelsize = 20, ticklabelsize = 12)
+            ax2 = Axis(fig[3, 1], aspect = 2.25, title = L"V_{RMS}", xlabel = L"Time", ylabel = L"Velocity",
+                titlesize = 20,
+                yticklabelsize = 12,
+                xticklabelsize = 12,
+                xlabelsize = 12,
+                ylabelsize = 12
+            )
+            lines!(ax2, trms, Urms, color = :black)
+
             save(joinpath(figdir, "$(it).png"), fig)
             fig
         end
