### 2025-10-07 03880eb
save datafram fro comparison
