### 2025-10-08 dc3f086
eliminate particles crossing marker chain
