### 2025-10-10 0450d6e
switch CI to `lts` and `1` testing rather than `1.10`, `1.11`, `1.12`
