### 2025-10-10 cf8ebe2
Fix function type signature in extensions
