+++ previews/PR357/assets/man_subduction2D_setup.md.C67yTyz_.js
@@ -1,0 +1,50 @@
+import{_ as k,c as t,o as h,j as s,aA as n,a}from"./chunks/framework.BoMdGI0i.js";const l="/JustRelax.jl/previews/PR357/assets/Subduction2D_setup_4.B4FIC6Zt.png",p="/JustRelax.jl/previews/PR357/assets/Subduction2D_setup_2.CvGRnUfs.png",e="/JustRelax.jl/previews/PR357/assets/Subduction2D_setup_3.D6dkTcls.png",d="/JustRelax.jl/previews/PR357/assets/Subduction2D_setup_5.DlnIM8Zs.png",m=JSON.parse('{"title":"Model setup","description":"","frontmatter":{},"headers":[],"relativePath":"man/subduction2D/setup.md","filePath":"man/subduction2D/setup.md","lastUpdated":null}'),r={name:"man/subduction2D/setup.md"},E={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},g={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.566ex"},xmlns:"http://www.w3.org/2000/svg",width:"23.634ex",height:"2.262ex",role:"img",focusable:"false",viewBox:"0 -750 10446.3 1000","aria-hidden":"true"};function y(o,i,F,T,c,Q){return h(),t("div",null,[i[6]||(i[6]=s("h1",{id:"Model-setup",tabindex:"-1"},[a("Model setup "),s("a",{class:"header-anchor",href:"#Model-setup","aria-label":'Permalink to "Model setup {#Model-setup}"'},"​")],-1)),s("p",null,[i[2]||(i[2]=a("As described in the original ",-1)),i[3]||(i[3]=s("a",{href:"https://doi.org/10.5194/se-15-567-2024",target:"_blank",rel:"noreferrer"},"paper",-1)),i[4]||(i[4]=a(", the domain consists of a Cartesian box of ",-1)),s("mjx-container",E,[(h(),t("svg",g,[...i[0]||(i[0]=[n('<g stroke="currentColor" fill="currentColor" stroke-width="0" transform="scale(1,-1)"><g data-mml-node="math"><g data-mml-node="mi"><path data-c="3A9" d="M55 454Q55 503 75 546T127 617T197 665T272 695T337 704H352Q396 704 404 703Q527 687 596 615T666 454Q666 392 635 330T559 200T499 83V80H543Q589 81 600 83T617 93Q622 102 629 135T636 172L637 177H677V175L660 89Q645 3 644 2V0H552H488Q461 0 456 3T451 20Q451 89 499 235T548 455Q548 512 530 555T483 622T424 656T361 668Q332 668 303 658T243 626T193 560T174 456Q174 380 222 233T270 20Q270 7 263 0H77V2Q76 3 61 89L44 175V177H84L85 172Q85 171 88 155T96 119T104 93Q109 86 120 84T178 80H222V83Q206 132 162 199T87 329T55 454Z" style="stroke-width:3;"></path></g><g data-mml-node="mo" transform="translate(999.8,0)"><path data-c="2208" d="M84 250Q84 372 166 450T360 539Q361 539 377 539T419 540T469 540H568Q583 532 583 520Q583 511 570 501L466 500Q355 499 329 494Q280 482 242 458T183 409T147 354T129 306T124 272V270H568Q583 262 583 250T568 230H124V228Q124 207 134 177T167 112T231 48T328 7Q355 1 466 0H570Q583 -10 583 -20Q583 -32 568 -40H471Q464 -40 446 -40T417 -41Q262 -41 172 45Q84 127 84 250Z" style="stroke-width:3;"></path></g><g dat
[…]
