+++ dev/assets/man_backend.md.DW4AZowj.lean.js
@@ -1,0 +1,1 @@
+import{_ as e,c as s,o as i,aA as t}from"./chunks/framework.CSEylmgj.js";const u=JSON.parse('{"title":"Selecting the backend","description":"","frontmatter":{},"headers":[],"relativePath":"man/backend.md","filePath":"man/backend.md","lastUpdated":null}'),l={name:"man/backend.md"};function n(d,a,p,o,h,c){return i(),s("div",null,[...a[0]||(a[0]=[t("",7)])])}const k=e(l,[["render",n]]);export{u as __pageData,k as default};
