### 2025-10-13 d6e44f3
shortened paper
