### 2025-10-13 65247b4
add missing DOI
