+++ docs/paper/paper.bib
@@ -1,24 +1,27 @@
 @article{bezanson2017julia,
-  title={Julia: A fresh approach to numerical computing},
-  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-  journal={SIAM review},
-  volume={59},
-  number={1},
-  pages={65--98},
-  year={2017},
-  publisher={SIAM},
-  url={10.1137/141000671}
+  author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+  title = {Julia: A Fresh Approach to Numerical Computing},
+  journal = {SIAM Review},
+  volume = {59},
+  number = {1},
+  pages = {65-98},
+  year = {2017},
+  doi = {10.1137/141000671},
+  URL = {https://doi.org/10.1137/141000671}
 }
 
 @article{larionov2017variational,
-  title={Variational stokes: A unified pressure-viscosity solver for accurate viscous liquids},
-  author={Larionov, Egor and Batty, Christopher and Bridson, Robert},
-  journal={ACM Transactions on Graphics (TOG)},
-  volume={36},
-  number={4},
-  pages={1--11},
-  year={2017},
-  publisher={ACM New York, NY, USA}
+  author = {Larionov, Egor and Batty, Christopher and Bridson, Robert}, 
+  title = {Variational stokes: a unified pressure-viscosity solver for accurate viscous liquids}, 
+  year = {2017}, 
+  issue_date = {August 2017}, 
+  publisher = {Association for Computing Machinery}, 
+  address = {New York, NY, USA}, 
+  volume = {36}, 
+  number = {4},  
+  issn = {0730-0301}, 
+  url = {https://doi.org/10.1145/3072959.3073628}, 
+  doi = {10.1145/3072959.3073628},
 }
 
 @article{Harlow1965,
@@ -55,8 +58,13 @@
   number={1-4},
   pages={83--105},
   year={2007},
+  note = {Computational Challenges in the Earth Sciences},
+  issn = {0031-9201},
+  doi = {https://doi.org/10.1016/j.pepi.2007.04.015},
+  url = {https://www.sciencedirect.com/science/article/pii/S0031920107000969},
   publisher={Elsevier}
 }
+
 @article{gerya2015plate,
   title={Plate tectonics on the Earth triggered by plume-induced subduction initiation},
   author={Gerya, Taras V and Stern, Robert J and Baes, Marzieh and Sobolev, Stephan V and Whattam, Scott A},
@@ -65,6 +73,8 @@
   number={7577},
   pages={221--225},
   year={2015},
+  doi = {10.1038/nature15752},
+  url = {https://doi.org/10.1038/nature15752},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -76,6 +86,8 @@
   number={14},
   pages={5757--5786},
   year={2022},
+  url = {https://gmd.copernicus.org/articles/15/5757/2022/},
+  doi = {10.5194/gmd-15-5757-2022},
   publisher={Copernicus Publications G{\"o}ttingen, Germany}
 }
 
@@ -87,16 +99,19 @@
   number={1-4},
   pages={7--18},
   year={2008},
+  doi = {10.1016/j.pepi.2008.08.005},
   publisher={Elsevier}
 }
 
 
 @inproceedings{lamem,
-  title={3D modelling of non-linear visco-elasto-plastic crustal and lithospheric processes using LaMEM},
-  author={Popov, Anton and Kaus, Boris},
-  booktitle={EGU General Assembly Conference Abstracts},
-  pages={EPSC2016--10128},
-  year={2016}
[…]
