+++ docs/paper/paper_long.md
@@ -1,0 +1,247 @@
+---
+title: 'JustRelax.jl: A Julia package for geodynamic modeling with matrix-free solvers'
+tags:
+  - Julia
+  - Geosciences
+  - Geodynamics
+  - Tectonics
+  - Geophysics
+  - Computational geosciences
+authors:
+  - name: Albert de Montserrat
+    orcid: 0000-0003-1694-3735
+    affiliation: 1
+  - name: Pascal S. Aellig
+    orcid: 0009-0008-9039-5646
+    affiliation: 2
+  - name: Christian Schuler
+    orcid: 0009-0004-9873-9774
+    affiliation: 2
+  - name: Ivan Navarrete
+    orcid: 0009-0004-9272-511X
+    affiliation: 3
+  - name: Ludovic Räss
+    orcid: 0000-0002-1136-899X
+    affiliation: 4
+  - name: Lukas Fuchs
+    orcid: 0000-0002-9165-6384
+    affiliation: 5
+  - name: Boris J.P. Kaus
+    orcid: 0000-0002-0247-8660
+    affiliation: 2 # (Multiple affiliations must be quoted)
+  - name: Hugo Dominguez
+    orcid: 0009-0001-8425-0503
+    affiliation: 2
+
+affiliations:
+ - name: ETH Zürich, Switzerland
+   index: 1
+ - name: Johannes Gutenberg-University Mainz, Germany
+   index: 2
+ - name: École Normale Supérieure - PSL University, France
+   index: 3
+ - name: University of Lausanne, Switzerland
+   index: 4
+ - name: Frankfurt University, Germany
+   index: 5
+
+date: 21 August 2025
+bibliography: paper.bib
+---
+
+# Summary
+
+JustRelax.jl is an open source, highly portable, and high-performance Julia package designed for geodynamic modeling. It employs the Accelerated Pseudo-Transient (APT) method to solve the Stokes and diffusion equations to simulate complex geodynamic processes. The APT is a matrix-free and embarrassingly parallel iterative method, making it especially well-suited to exploit Graphics Processing Unit (GPU) hardware accelerators available in distributed systems of most HPC centers.
+
+JustRelax.jl incorporates a wide range of features critical to computational geodynamics. These include complex and highly non-linear rheologies, free surface and its good spatial locality (compared to e.g. matrix-matrix or matrix-vector multiplication) particle-in-cell method to track and advect material phases and properties. Several of the features JustRelax.jl is built upon are outsourced to specialized packages. This modular design considerably reduces the size of the core code base, making the entire package infrastructure easier to maintain. This also improves code reusability, as the external packages can be re-used either as stand-alone packages or as building blocks for some other package.
+
+Most of JustRelax.jl's core dependencies are fully written in the [Julia programming language](https://julialang.org/) [@bezanson2017julia], which facilitates contributions from users without extensive experience in programming, compared to other compiled languages common in HPC-focused software (C/C++/Fortran). The interactive environment of Julia allows users to prototype and implement new code in a single high-level language, improving the overall development experience in similar fashion to Python and MATLAB.
+
+# Statement of Need
+
[…]
