### 2025-10-14 e345ac6
fix _update_phases_given_markerchain_kernel!
