### 2025-10-15 98a0040
rework viscosity
