### 2025-10-16 6ef1c0a
a couple of fix ups
