+++ src/variational_stokes/Stokes3D.jl
@@ -69,7 +69,7 @@ function _solve_VS!(
 
     # compute buoyancy forces and viscosity
     compute_ρg!(ρg, phase_ratios, rheology, args)
-    compute_viscosity_τII!(stokes, phase_ratios, args, rheology, air_phase, viscosity_cutoff)
+    compute_viscosity!(stokes, phase_ratios, args, rheology, air_phase, viscosity_cutoff)
 
     # convert displacement to velocity
     displacement2velocity!(stokes, dt, flow_bcs)
