+++ src/stokes/Stokes2D.jl
@@ -346,7 +346,7 @@ function _solve!(
 
     # compute buoyancy forces and viscosity
     compute_ρg!(ρg[end], rheology, args)
-    compute_viscosity_τII!(stokes, args, rheology, viscosity_cutoff)
+    compute_viscosity!(stokes, args, rheology, viscosity_cutoff)
 
     # convert displacement to velocity
     displacement2velocity!(stokes, dt, flow_bcs)
@@ -548,7 +548,7 @@ function _solve!(
 
     # compute buoyancy forces and viscosity
     compute_ρg!(ρg, phase_ratios, rheology, args)
-    compute_viscosity_τII!(stokes, phase_ratios, args, rheology, viscosity_cutoff)
+    compute_viscosity!(stokes, phase_ratios, args, rheology, viscosity_cutoff)
     displacement2velocity!(stokes, dt, flow_bcs)
 
     while iter ≤ iterMax
