### 2025-10-16 a28bfad
fix test warnings
