+++ test/test_shearband2D_MPI.jl
@@ -227,30 +227,32 @@ function main(igg; nx = 64, ny = 64)
 
 end
 
-if CSCS_CI != true
-    @suppress begin
-        if backend_JR == CPUBackend
-            N = 30
-            nx = N * 2  # if only 2 CPU/GPU are used nx = 67 - 2 with N =128
-            ny = N * 2
-            igg = if !(JustRelax.MPI.Initialized())
-                IGG(init_global_grid(nx, ny, 1; init_MPI = true)...)
+let
+    if CSCS_CI != true
+        @suppress begin
+            if backend_JR == CPUBackend
+                N = 30
+                nx = N * 2  # if only 2 CPU/GPU are used nx = 67 - 2 with N =128
+                ny = N * 2
+                igg = if !(JustRelax.MPI.Initialized())
+                    IGG(init_global_grid(nx, ny, 1; init_MPI = true)...)
+                else
+                    igg
+                end
+                main(igg; nx = nx, ny = ny)
             else
-                igg
+                println("This test is only for CPU CI yet")
             end
-            main(igg; nx = nx, ny = ny)
-        else
-            println("This test is only for CPU CI yet")
         end
-    end
-else
-    N = 30
-    nx = N * 2  # if only 2 CPU/GPU are used nx = 67 - 2 with N =128
-    ny = N * 2
-    igg = if !(JustRelax.MPI.Initialized())
-        IGG(init_global_grid(nx, ny, 1; init_MPI = true, select_device = false)...)
     else
-        igg
+        N = 30
+        nx = N * 2  # if only 2 CPU/GPU are used nx = 67 - 2 with N =128
+        ny = N * 2
+        igg = if !(JustRelax.MPI.Initialized())
+            IGG(init_global_grid(nx, ny, 1; init_MPI = true, select_device = false)...)
+        else
+            igg
+        end
+        main(igg; nx = nx, ny = ny)
     end
-    main(igg; nx = nx, ny = ny)
 end
