### 2025-10-16 e91f4ca
fix stopping criteria of PT solvers when using MPI
