+++ src/variational_stokes/Stokes3D.jl
@@ -171,7 +171,7 @@ function _solve_VS!(
             for (norm_Ri, Ri) in zip((norm_Rx, norm_Ry, norm_Rz), @residuals(stokes.R))
                 push!(
                     norm_Ri,
-                    norm_mpi(Ri[2:(end - 1), 2:(end - 1), 2:(end - 1)]) / length(Ri),
+                    norm_mpi(Ri[2:(end - 1), 2:(end - 1), 2:(end - 1)]) / ((nx_g() - 1) * (ny_g() - 1) * (nz_g() - 1)),
                 )
             end
             push!(norm_∇V, norm_mpi(stokes.R.RP) / length(stokes.R.RP))
