+++ src/thermal_diffusion/DiffusionPT_solver.jl
@@ -30,7 +30,7 @@ function _heatdiffusion_PT!(
     _dt = inv(dt)
     _di = inv.(di)
 
-    _sq_len_RT = inv(sqrt((nx_g() + 1) * (ny_g() + 1)  * (nz_g() + (nz_g() > 1) )))
+    _sq_len_RT = inv(sqrt((nx_g() + 1) * (ny_g() + 1) * (nz_g() + (nz_g() > 1))))
     ϵ = pt_thermal.ϵ
     ni = size(thermal.Tc)
     @copy thermal.Told thermal.T
@@ -138,7 +138,7 @@ function _heatdiffusion_PT!(
     # Compute some constant stuff
     _dt = inv(dt)
     _di = inv.(di)
-    _sq_len_RT = inv(sqrt((nx_g() + 1) * (ny_g() + 1)  * (nz_g() + (nz_g() > 1) )))
+    _sq_len_RT = inv(sqrt((nx_g() + 1) * (ny_g() + 1) * (nz_g() + (nz_g() > 1))))
     ϵ = pt_thermal.ϵ
     ni = size(thermal.Tc)
     @copy thermal.Told thermal.T
