+++ previews/PR343/assets/chunks/VPLocalSearchBox.Bv_MuAt9.js
@@ -1,0 +1,8 @@
+var Nt=Object.defineProperty;var Ft=(a,e,t)=>e in a?Nt(a,e,{enumerable:!0,configurable:!0,writable:!0,value:t}):a[e]=t;var Re=(a,e,t)=>Ft(a,typeof e!="symbol"?e+"":e,t);import{V as Ot,D as le,h as ge,aB as et,aC as Ct,aD as Rt,aE as Mt,q as $e,aF as At,d as Lt,aG as tt,p as he,aH as Dt,ac as Pt,s as zt,aI as Vt,v as Me,P as fe,O as _e,aJ as $t,ak as jt,W as Bt,R as Wt,$ as Kt,b as Jt,o as H,j as x,a0 as qt,a2 as Ut,k as L,a5 as Gt,ad as Ht,c as Z,e as Se,n as st,B as nt,F as it,a as pe,t as ve,aa as Qt,aK as rt,aL as Yt,aq as Zt,av as Xt,aM as es,_ as ts}from"./framework.B94pNcpO.js";import{u as ss,c as ns}from"./theme.ChkRgJH1.js";const is={root:()=>Ot(()=>import("./@localSearchIndexroot.o309KOQY.js"),[])};/*!
+* tabbable 6.2.0
+* @license MIT, https://github.com/focus-trap/tabbable/blob/master/LICENSE
+*/var vt=["input:not([inert])","select:not([inert])","textarea:not([inert])","a[href]:not([inert])","button:not([inert])","[tabindex]:not(slot):not([inert])","audio[controls]:not([inert])","video[controls]:not([inert])",'[contenteditable]:not([contenteditable="false"]):not([inert])',"details>summary:first-of-type:not([inert])","details:not([inert])"],ke=vt.join(","),mt=typeof Element>"u",re=mt?function(){}:Element.prototype.matches||Element.prototype.msMatchesSelector||Element.prototype.webkitMatchesSelector,Ne=!mt&&Element.prototype.getRootNode?function(a){var e;return a==null||(e=a.getRootNode)===null||e===void 0?void 0:e.call(a)}:function(a){return a==null?void 0:a.ownerDocument},Fe=function a(e,t){var s;t===void 0&&(t=!0);var n=e==null||(s=e.getAttribute)===null||s===void 0?void 0:s.call(e,"inert"),r=n===""||n==="true",i=r||t&&e&&a(e.parentNode);return i},rs=function(e){var t,s=e==null||(t=e.getAttribute)===null||t===void 0?void 0:t.call(e,"contenteditable");return s===""||s==="true"},gt=function(e,t,s){if(Fe(e))return[];var n=Array.prototype.slice.apply(e.querySelectorAll(ke));return t&&re.call(e,ke)&&n.unshift(e),n=n.filter(s),n},bt=function a(e,t,s){for(var n=[],r=Array.from(e);r.length;){var i=r.shift();if(!Fe(i,!1))if(i.tagName==="SLOT"){var o=i.assignedElements(),l=o.length?o:i.children,c=a(l,!0,s);s.flatten?n.push.apply(n,c):n.push({scopeParent:i,candidates:c})}else{var h=re.call(i,ke);h&&s.filter(i)&&(t||!e.includes(i))&&n.push(i);var m=i.shadowRoot||typeof s.getShadowRoot=="function"&&s.getShadowRoot(i),f=!Fe(m,!1)&&(!s.shadowRootFilter||s.shadowRootFilter(i));if(m&&f){var g=a(m===!0?i.children:m.children,!0,s);s.flatten?n.push.apply(n,g):n.push({scopeParent:i,candidates:g})}else r.unshift.apply(r,i.children)}}return n},yt=function(
[…]
