+++ previews/PR369/assets/chunks/stokes_stag2D.BSDaukLS.js
@@ -1,0 +1,1 @@
+const s="/JustRelax.jl/previews/PR369/assets/temp_stag2D.CPrcX1ke.png",t="/JustRelax.jl/previews/PR369/assets/stokes_stag2D.5cFNPKDo.png";export{s as _,t as a};
