+++ previews/PR357/assets/man_security.md.XHxGW9xR.js
@@ -1,0 +1,1 @@
+import{_ as t,c as i,o as r,aA as a}from"./chunks/framework.BoMdGI0i.js";const y=JSON.parse('{"title":"Security Policy","description":"","frontmatter":{},"headers":[],"relativePath":"man/security.md","filePath":"man/security.md","lastUpdated":null}'),o={name:"man/security.md"};function s(n,e,l,u,p,c){return r(),i("div",null,[...e[0]||(e[0]=[a('<h1 id="Security-Policy" tabindex="-1">Security Policy <a class="header-anchor" href="#Security-Policy" aria-label="Permalink to &quot;Security Policy {#Security-Policy}&quot;">​</a></h1><p>We take security issues seriously. We appreciate all efforts to responsibly disclose any security issues and will make every effort to acknowledge contributions.</p><h2 id="Supported-Versions" tabindex="-1">Supported Versions <a class="header-anchor" href="#Supported-Versions" aria-label="Permalink to &quot;Supported Versions {#Supported-Versions}&quot;">​</a></h2><p>The current stable release following the interpretation of <a href="https://julialang.github.io/Pkg.jl/dev/compatibility/#Version-specifier-format-1" target="_blank" rel="noreferrer">semantic versioning (SemVer)</a> used in the Julia ecosystem is supported with security updates.</p><h2 id="Reporting-a-Vulnerability" tabindex="-1">Reporting a Vulnerability <a class="header-anchor" href="#Reporting-a-Vulnerability" aria-label="Permalink to &quot;Reporting a Vulnerability {#Reporting-a-Vulnerability}&quot;">​</a></h2><p>To report a security issue, please use the GitHub Security Advisory <a href="https://github.com/PTsolvers/JustRelax.jl/security/advisories/new" target="_blank" rel="noreferrer">&quot;Report a Vulnerability&quot;</a> tab.</p><p>We will send a response indicating the next steps in handling your report. After the initial reply to your report, we will keep you informed of the progress towards a fix and full announcement, and may ask for additional information or guidance.</p><p>Please report security bugs in third-party modules directly to the person or team maintaining the module.</p><p>Public notifications of vulnerabilities will be shared in community channels such as Slack.</p>',9)])])}const h=t(o,[["render",s]]);export{y as __pageData,h as default};

[…]
