### 2025-10-17 48c152b
add GMD scripts (convergence tests), fix some bugs. Add ExactFieldSolutions as its required by benchs
