+++ Project.toml
@@ -7,6 +7,7 @@ version = "0.4.2"
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
 CellArrays = "d35fcfd7-7af4-4c67-b1aa-d78070614af4"
 Crayons = "a8cc5b0e-0ffa-5ad4-8c14-923d3ee1735f"
+ExactFieldSolutions = "2a6b1ac7-2de5-4354-bffb-bf009f6a4bef"
 GeoParams = "e018b62d-d9de-4a26-8697-af89c310ae38"
 HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
 ImplicitGlobalGrid = "4d7a3746-15be-11ea-1130-334b0c4f5fa0"
@@ -37,6 +38,7 @@ Adapt = "4"
 CUDA = "5"
 CellArrays = "0.3.2"
 Crayons = "4.1.1"
+ExactFieldSolutions = "0.1.6"
 GeoParams = "0.7"
 HDF5 = "0.17.1"
 ImplicitGlobalGrid = "0.16"
