+++ previews/PR357/assets/chunks/@localSearchIndexroot.DasgyE8A.js
@@ -1,1 +1,0 @@
-const e='{"documentCount":151,"nextId":151,"documentIds":{"0":"/JustRelax.jl/previews/PR357/#What-is-JustRelax.jl?","1":"/JustRelax.jl/previews/PR357/#How-to-Install-JustRelax.jl?","2":"/JustRelax.jl/previews/PR357/#funding","3":"/JustRelax.jl/previews/PR357/man/Blankenbach#Blankenbach-benchmark","4":"/JustRelax.jl/previews/PR357/man/Blankenbach#Initialize-packages","5":"/JustRelax.jl/previews/PR357/man/Blankenbach#script","6":"/JustRelax.jl/previews/PR357/man/Blankenbach#Model-domain","7":"/JustRelax.jl/previews/PR357/man/Blankenbach#rheology","8":"/JustRelax.jl/previews/PR357/man/Blankenbach#Initialize-particles","9":"/JustRelax.jl/previews/PR357/man/Blankenbach#Temperature-anomaly","10":"/JustRelax.jl/previews/PR357/man/Blankenbach#Stokes-and-heat-diffusion-arrays","11":"/JustRelax.jl/previews/PR357/man/Blankenbach#Initialize-thermal-profile-and-viscosity-fields","12":"/JustRelax.jl/previews/PR357/man/Blankenbach#Boundary-conditions","13":"/JustRelax.jl/previews/PR357/man/Blankenbach#Pseuo-transient-coefficients","14":"/JustRelax.jl/previews/PR357/man/Blankenbach#Just-before-solving-the-problem...","15":"/JustRelax.jl/previews/PR357/man/Blankenbach#Advancing-one-time-step","16":"/JustRelax.jl/previews/PR357/man/Blankenbach#visualization","17":"/JustRelax.jl/previews/PR357/man/Blankenbach#fields","18":"/JustRelax.jl/previews/PR357/man/Blankenbach#Final-model","19":"/JustRelax.jl/previews/PR357/man/GeoParams#List-of-GeoParams-functions","20":"/JustRelax.jl/previews/PR357/man/ShearBands#ShearBand-benchmark","21":"/JustRelax.jl/previews/PR357/man/ShearBands#Initialize-packages","22":"/JustRelax.jl/previews/PR357/man/ShearBands#script","23":"/JustRelax.jl/previews/PR357/man/ShearBands#Model-domain","24":"/JustRelax.jl/previews/PR357/man/ShearBands#Physical-properties-using-GeoParams","25":"/JustRelax.jl/previews/PR357/man/ShearBands#rheology","26":"/JustRelax.jl/previews/PR357/man/ShearBands#Phase-anomaly","27":"/JustRelax.jl/previews/PR357/man/ShearBands#Stokes-arrays","28":"/JustRelax.jl/previews/PR357/man/ShearBands#Initialize-viscosity-fields","29":"/JustRelax.jl/previews/PR357/man/ShearBands#Boundary-conditions","30":"/JustRelax.jl/previews/PR357/man/ShearBands#Pseuo-transient-coefficients","31":"/JustRelax.jl/previews/PR357/man/ShearBands#Just-before-solving-the-problem...","32":"/JustRelax.jl/previews/PR357/man/ShearBands#Advancing-one-time-step","33":"/JustRelax.jl/previews/PR357/man/ShearBands#visualization","34":"/JustRelax.jl/previews/PR357/man/ShearBands#fields","35":"/JustRelax.jl/previews/PR357/man/ShearBands#Final-model","36":"/JustRelax.jl/previews/PR357/man/JustPIC#List-of-JustPIC-functions","37":"/JustRelax.jl/previews/PR357/man/advection#Field-advection","38"
[…]
