### 2025-10-17 6190a65
fix viscosity update in absence of phase ratios
