### 2025-10-17 93ba968
make `relλ` a kwarg with default of `relλ=1`
