+++ src/stokes/Stokes2D.jl
@@ -113,6 +113,7 @@ function _solve!(
             push!(
                 norm_∇V, norm_mpi(stokes.∇V) / (Vmax - Vmin) * lx / sqrt(length(stokes.∇V))
             )
+
             err = maximum_mpi(norm_Rx[end], norm_Ry[end], norm_∇V[end])
             push!(err_evo1, err)
             push!(err_evo2, iter)
@@ -427,10 +428,10 @@ function _solve!(
 
             errs = (
                 norm_mpi(@views stokes.R.Rx[2:(end - 1), 2:(end - 1)]) /
-                    ((nx_g() - 2) * (ny_g() - 1)),
+                    √((nx_g() - 2) * (ny_g() - 1)),
                 norm_mpi(@views stokes.R.Ry[2:(end - 1), 2:(end - 1)]) /
-                    ((nx_g() - 1) * (ny_g() - 2)),
-                norm_mpi(stokes.R.RP) / (nx_g() * ny_g()),
+                    √((nx_g() - 1) * (ny_g() - 2)),
+                norm_mpi(stokes.R.RP) / √(nx_g() * ny_g()),
             )
             push!(norm_Rx, errs[1])
             push!(norm_Ry, errs[2])
@@ -499,6 +500,7 @@ function _solve!(
         strain_increment = false,
         viscosity_cutoff = (-Inf, Inf),
         viscosity_relaxation = 1.0e-2,
+        λ_relaxation = 1e0,
         iterMax = 50.0e3,
         iterMin = 1.0e2,
         free_surface = false,
@@ -541,7 +543,7 @@ function _solve!(
     # solver loop
     @copy stokes.P0 stokes.P
     wtime0 = 0.0
-    relλ = 0.2
+    relλ = λ_relaxation
     θ = deepcopy(stokes.P)
     λ = @zeros(ni...)
     λv = @zeros(ni .+ 1...)
@@ -551,7 +553,6 @@ function _solve!(
     for Aij in @tensor_center(stokes.ε_pl)
         Aij .= 0.0
     end
-
 
     # compute buoyancy forces and viscosity
     compute_ρg!(ρg, phase_ratios, rheology, args)
@@ -703,11 +704,12 @@ function _solve!(
 
             errs = (
                 norm_mpi(@views stokes.R.Rx[2:(end - 1), 2:(end - 1)]) /
-                    ((nx_g() - 2) * (ny_g() - 1)),
+                    √((nx_g() - 2) * (ny_g() - 1)),
                 norm_mpi(@views stokes.R.Ry[2:(end - 1), 2:(end - 1)]) /
-                    ((nx_g() - 1) * (ny_g() - 2)),
-                norm_mpi(stokes.R.RP) / (nx_g() * ny_g()),
+                    √((nx_g() - 1) * (ny_g() - 2)),
+                norm_mpi(stokes.R.RP) / √(nx_g() * ny_g()),
             )
+
             push!(norm_Rx, errs[1])
             push!(norm_Ry, errs[2])
             push!(norm_∇V, errs[3])
