### 2025-10-17 eacc465
update residual calculations
