### 2025-10-17 eba6a32
fix tests
