+++ previews/PR369/assets/man_backend.md.Dt-KS2Gn.lean.js
@@ -1,1 +1,0 @@
-import{_ as e,c as s,o as i,aA as t}from"./chunks/framework.D_OaAAct.js";const u=JSON.parse('{"title":"Selecting the backend","description":"","frontmatter":{},"headers":[],"relativePath":"man/backend.md","filePath":"man/backend.md","lastUpdated":null}'),l={name:"man/backend.md"};function n(d,a,p,h,o,c){return i(),s("div",null,[...a[0]||(a[0]=[t("",7)])])}const k=e(l,[["render",n]]);export{u as __pageData,k as default};

[…]
