+++ src/stokes/Stokes2D.jl
@@ -242,7 +242,15 @@ function _solve!(
             @parallel (@idx ni) compute_Res!(
                 stokes.R.Rx, stokes.R.Ry, stokes.P, @stress(stokes)..., ρg..., _di...
             )
-            errs = maximum_mpi.((abs.(stokes.R.Rx), abs.(stokes.R.Ry), abs.(stokes.R.RP)))
+            
+            Acell = prod(di)
+            errs = (
+                norm_mpi(@views stokes.R.Rx[2:(end - 1), 2:(end - 1)]) * √(Acell),
+                norm_mpi(@views stokes.R.Ry[2:(end - 1), 2:(end - 1)]) * √(Acell),
+                norm_mpi(stokes.R.RP) * √(Acell),
+            )
+
+            # errs = maximum_mpi.((abs.(stokes.R.Rx), abs.(stokes.R.Ry), abs.(stokes.R.RP)))
             push!(norm_Rx, errs[1])
             push!(norm_Ry, errs[2])
             push!(norm_∇V, errs[3])
@@ -426,13 +434,21 @@ function _solve!(
                 dt * free_surface,
             )
 
+            # errs = (
+            #     norm_mpi(@views stokes.R.Rx[2:(end - 1), 2:(end - 1)]) /
+            #         √((nx_g() - 2) * (ny_g() - 1)),
+            #     norm_mpi(@views stokes.R.Ry[2:(end - 1), 2:(end - 1)]) /
+            #         √((nx_g() - 1) * (ny_g() - 2)),
+            #     norm_mpi(stokes.R.RP) / √(nx_g() * ny_g()),
+            # )
+
+            Acell = prod(di)
             errs = (
-                norm_mpi(@views stokes.R.Rx[2:(end - 1), 2:(end - 1)]) /
-                    √((nx_g() - 2) * (ny_g() - 1)),
-                norm_mpi(@views stokes.R.Ry[2:(end - 1), 2:(end - 1)]) /
-                    √((nx_g() - 1) * (ny_g() - 2)),
-                norm_mpi(stokes.R.RP) / √(nx_g() * ny_g()),
+                norm_mpi(@views stokes.R.Rx[2:(end - 1), 2:(end - 1)]) * √(Acell),
+                norm_mpi(@views stokes.R.Ry[2:(end - 1), 2:(end - 1)]) * √(Acell),
+                norm_mpi(stokes.R.RP) * √(Acell),
             )
+
             push!(norm_Rx, errs[1])
             push!(norm_Ry, errs[2])
             push!(norm_∇V, errs[3])
@@ -702,13 +718,19 @@ function _solve!(
                 dt * free_surface,
             )
 
+            Acell = prod(di)
             errs = (
-                norm_mpi(@views stokes.R.Rx[2:(end - 1), 2:(end - 1)]) /
-                    √((nx_g() - 2) * (ny_g() - 1)),
-                norm_mpi(@views stokes.R.Ry[2:(end - 1), 2:(end - 1)]) /
-                    √((nx_g() - 1) * (ny_g() - 2)),
-                norm_mpi(stokes.R.RP) / √(nx_g() * ny_g()),
+                norm_mpi(@views stokes.R.Rx[2:(end - 1), 2:(end - 1)]) * √(Acell),
+                norm_mpi(@views stokes.R.Ry[2:(end - 1), 2:(end - 1)]) * √(Acell),
+                norm_mpi(stokes.R.RP) * √(Acell),
             )
+            # errs = (
+            #     norm_mpi(@views stokes.R.Rx[2:(end - 1), 2:(end - 1)]) /
+            #         √((nx_g() - 2) * (ny_g() - 1)),
[…]
