### 2025-10-17 519ecc2
make lambda relaxation a kwarg
