### 2025-10-17 fd98958
set  lambda relaxation default to 0.2
