### 2025-10-20 8daaafc
revert residual calculation changes
