### 2025-10-20 b1e83e1
add `precision = Float32` keyword argument to save hdf5 and vtk functions
