### 2025-10-20 d20d5b3
fix `namevar` macro
