+++ src/IO/VTK.jl
@@ -274,9 +274,9 @@ end
 function save_particles2D(particles, pPhases, precision; conversion = 1.0e3, fname::String = "./particles", pvd::Union{Nothing, String} = nothing, t::Number = 0.0)
     p = particles.coords
     ppx, ppy = p
-    pxv  = precision.(Array(ppx.data)[:] ./ conversion)
-    pyv  = precision.(Array(ppy.data)[:] ./ conversion)
-    clr  = precision.(Array(pPhases.data)[:])
+    pxv = precision.(Array(ppx.data)[:] ./ conversion)
+    pyv = precision.(Array(ppy.data)[:] ./ conversion)
+    clr = precision.(Array(pPhases.data)[:])
     idxv = Array(particles.index.data[:])
 
     x = pxv[idxv]
@@ -301,10 +301,10 @@ end
 function save_particles3D(particles, pPhases, precision; conversion = 1.0e3, fname::String = "./particles", pvd::Union{Nothing, String} = nothing, t::Number = 0.0)
     p = particles.coords
     ppx, ppy = p
-    pxv  = precision.(Array(ppx.data)[:] ./ conversion)
-    pyv  = precision.(Array(ppy.data)[:] ./ conversion)
-    pzv  = precision.(Array(ppz.data)[:] ./ conversion)
-    clr  = precision.(Array(pPhases.data)[:])
+    pxv = precision.(Array(ppx.data)[:] ./ conversion)
+    pyv = precision.(Array(ppy.data)[:] ./ conversion)
+    pzv = precision.(Array(ppz.data)[:] ./ conversion)
+    clr = precision.(Array(pPhases.data)[:])
     idxv = Array(particles.index.data[:])
 
     x = pxv[idxv]
