+++ src/rheology/Viscosity_old.jl
@@ -1,392 +1,0 @@
-# # Traits
-
-# without phase ratios
-@inline function update_viscosity!(
-        stokes::JustRelax.StokesArrays, args, rheology, cutoff; relaxation = 1.0e0
-    )
-    update_viscosity!(
-        islinear(rheology), stokes, args, rheology, cutoff; relaxation = relaxation
-    )
-    return nothing
-end
-
-@inline function update_viscosity!(
-        ::LinearRheologyTrait,
-        stokes::JustRelax.StokesArrays,
-        args,
-        rheology,
-        cutoff;
-        relaxation = 1.0e0,
-    )
-    return nothing
-end
-
-@inline function update_viscosity!(
-        ::NonLinearRheologyTrait,
-        stokes::JustRelax.StokesArrays,
-        args,
-        rheology,
-        cutoff;
-        relaxation = 1.0e0,
-    )
-    compute_viscosity!(stokes, args, rheology, cutoff; relaxation = relaxation)
-    return nothing
-end
-
-# with phase ratios
-@inline function update_viscosity!(
-        stokes::JustRelax.StokesArrays,
-        phase_ratios,
-        args,
-        rheology,
-        cutoff;
-        air_phase::Integer = 0,
-        relaxation = 1.0e0,
-    )
-    update_viscosity!(
-        islinear(rheology),
-        stokes,
-        phase_ratios,
-        args,
-        rheology,
-        air_phase,
-        cutoff;
-        relaxation = relaxation,
-    )
-    return nothing
-end
-
-@inline function update_viscosity!(
-        ::LinearRheologyTrait,
-        stokes::JustRelax.StokesArrays,
-        phase_ratios,
-        args,
-        rheology,
-        air_phase,
-        cutoff;
-        relaxation = 1.0e0,
-    )
-    return nothing
-end
-
-@inline function update_viscosity!(
-        ::NonLinearRheologyTrait,
-        stokes::JustRelax.StokesArrays,
-        phase_ratios,
-        args,
-        rheology,
-        air_phase,
-        cutoff;
-        relaxation = 1.0e0,
-    )
-    compute_viscosity!(
-        stokes, phase_ratios, args, rheology, cutoff; relaxation = relaxation, air_phase = air_phase
-    )
-    return nothing
-end
-
-## 2D KERNELS
-function compute_viscosity!(
-        stokes::JustRelax.StokesArrays, args, rheology, cutoff; relaxation = 1.0e0
-    )
-    return compute_viscosity!(backend(stokes), stokes, relaxation, args, rheology, cutoff)
-end
-
-function compute_viscosity!(::CPUBackendTrait, stokes, ν, args, rheology, cutoff)
-    return _compute_viscosity!(stokes, ν, args, rheology, cutoff)
-end
-
-function _compute_viscosity!(stokes::JustRelax.StokesArrays, ν, args, rheology, cutoff)
-    ni = size(stokes.viscosity.η)
-    @parallel (@idx ni) compute_viscosity_kernel!(
-        stokes.viscosity.η, ν, @strain(stokes)..., args, rheology, cutoff
-    )
-    return nothing
-end
-
-@parallel_indices (I...) function compute_viscosity_kernel!(
-        η, ν, εxx, εyy, εxyv, args, rheology, cutoff
-    )
-
-    # convenience closure
-    Base.@propagate_inbounds @inline gather(A) = _gather(A, I...)
-
-    @inbounds begin
-        # cache
-        ε = εxx[I...], εyy[I...]
-
-        # we need strain rate not to be zero, otherwise we get NaNs
-        εII_0 = allzero(ε...) * eps()
-
-        # argument fields at local index
[…]
