### 2025-10-20 92bd3db
fix `@namevar` macro
