+++ previews/PR357/assets/man_plume3D_rheology.md.C3mAiliG.lean.js
@@ -1,0 +1,1 @@
+import{_ as i,c as a,o as h,aA as k}from"./chunks/framework.BoMdGI0i.js";const g=JSON.parse('{"title":"Using GeoParams.jl to define the rheology of the material phases","description":"","frontmatter":{},"headers":[],"relativePath":"man/plume3D/rheology.md","filePath":"man/plume3D/rheology.md","lastUpdated":null}'),n={name:"man/plume3D/rheology.md"};function l(t,s,p,e,E,r){return h(),a("div",null,[...s[0]||(s[0]=[k("",12)])])}const y=i(n,[["render",l]]);export{g as __pageData,y as default};
