+++ previews/PR357/assets/man_contributing.md.gowVPpRC.lean.js
@@ -1,0 +1,1 @@
+import{_ as s,c as a,o as e,aA as t}from"./chunks/framework.BoMdGI0i.js";const d=JSON.parse('{"title":"Contributing","description":"","frontmatter":{},"headers":[],"relativePath":"man/contributing.md","filePath":"man/contributing.md","lastUpdated":null}'),i={name:"man/contributing.md"};function p(r,n,o,l,c,h){return e(),a("div",null,[...n[0]||(n[0]=[t("",6)])])}const f=s(i,[["render",p]]);export{d as __pageData,f as default};

[…]
