### 2025-10-21 fd16b77
even shorter paper for JOSS
