+++ docs/paper/paper.md
@@ -10,9 +10,11 @@ tags:
 authors:
   - name: Albert de Montserrat
     orcid: 0000-0003-1694-3735
+    equal-contrib: true
     affiliation: 1
   - name: Pascal S. Aellig
     orcid: 0009-0008-9039-5646
+    equal-contrib: true
     affiliation: 2
   - name: Christian Schuler
     orcid: 0009-0004-9873-9774
@@ -45,83 +47,55 @@ affiliations:
  - name: Frankfurt University, Germany
    index: 5
 
-date: 21 August 2025
+date: 21 October 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-JustRelax.jl is an open source, highly portable, and high-performance Julia package designed for geodynamic modeling. It employs the Accelerated Pseudo-Transient (APT) method to solve the Stokes and diffusion equations to simulate complex geodynamic processes. The APT is a matrix-free and embarrassingly parallel iterative method, making it especially well-suited to exploit Graphics Processing Unit (GPU) hardware accelerators available in distributed systems of most HPC centers.
+JustRelax.jl is an open source, highly portable, and high-performance Julia package for geodynamic modeling. It employs the Accelerated Pseudo-Transient (APT) method—a matrix-free, embarrassingly parallel iterative method—to solve the Stokes and diffusion equations, making it well-suited to exploit GPU hardware in modern distributed HPC systems.
 
-JustRelax.jl incorporates a wide range of features critical to computational geodynamics. These include complex and highly non-linear rheologies, free surface and its good spatial locality (compared to e.g. matrix-matrix or matrix-vector multiplication) particle-in-cell method to track and advect material phases and properties. Several of the features JustRelax.jl is built upon are outsourced to specialized packages. This modular design considerably reduces the size of the core code base, making the entire package infrastructure easier to maintain. This also improves code reusability, as the external packages can be re-used either as stand-alone packages or as building blocks for some other package.
+The package incorporates features critical to computational geodynamics, including complex non-linear rheologies, free surface tracking, and a particle-in-cell method to advect material phases and properties. Its modular design outsources specialized functionality to external packages, reducing the core code base and improving maintainability and reusability.
 
-Most of JustRelax.jl's core dependencies are fully written in the [Julia programming language](https://julialang.org/) [@bezanson2017julia], which facilitates contributions from users without extensive experience in programming, compared to other compiled languages common in HPC-focused software (C/C++/Fortran). The interactive environment of Julia allows users to prototype and implement new code in a single high-level language, improving the overall development experience in similar fashion to Python and MATLAB.
+JustRelax.jl and its dependencies are written in [Julia](https://julialang.org/) [@bezanson2017julia], which lowers the barrier to contribution compared to traditional HPC languages (C/C++/Fortran). Julia's interactive environment enables rapid prototyping in a single high-level language, similar to Python and MATLAB, while maintaining high performance.
 
 # Statement of Need
 
 Simulating Earth's thermo-mechanical evolution requires solving coupled, nonlinear Stokes and heat‑diffusion problems across large domains with sharp material contrasts, complex rheologies and long time scales. These calculations are computationally demanding and traditionally rely on monolithic, CPU‑oriented codes built around matrix‑based linear solvers. Such designs are hard to port efficiently to modern heterogeneous HPC systems (e.g. multi‑GPU nodes).
 
[…]
