+++ previews/PR343/assets/chunks/theme.CaUsv2-r.js
@@ -1,0 +1,2 @@
+const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/chunks/VPLocalSearchBox.Ba_GU0tg.js","assets/chunks/framework.B94pNcpO.js"])))=>i.map(i=>d[i]);
+import{d as $,c as h,r as v,n as T,o as i,a as Q,t as M,b as L,w as m,T as re,e as _,_ as V,u as Se,i as Et,f as Wt,g as Fe,h as P,j as c,k as s,l as le,m as Y,p as w,q as W,s as Le,v as D,x as xe,y as Ue,z as Ht,A as Bt,F as O,B as F,C as ae,D as it,E as we,G as y,H as q,I as lt,J as ve,K as te,L as z,M as Ot,N as ce,O as Oe,P as Pe,Q as rt,R as Ve,S as It,U as Ct,V as Rt,W as ut,X as De,Y as jt,Z as Ft,$ as Ut,a0 as Dt,a1 as ct,a2 as de,a3 as zt,a4 as dt,a5 as Kt,a6 as Me,a7 as ht,a8 as ne,a9 as Ie,aa as pt,ab as ze,ac as Ke,ad as Ge,ae as he,af as Gt,ag as qt,ah as Jt,ai as vt,aj as Yt,ak as Xt,al as Qt,am as We}from"./framework.B94pNcpO.js";const Zt=$({__name:"VPBadge",props:{text:{},type:{default:"tip"}},setup(e){return(t,a)=>(i(),h("span",{class:T(["VPBadge",e.type])},[v(t.$slots,"default",{},()=>[Q(M(e.text),1)])],2))}}),ea={key:0,class:"VPBackdrop"},ta=$({__name:"VPBackdrop",props:{show:{type:Boolean}},setup(e){return(t,a)=>(i(),L(re,{name:"fade"},{default:m(()=>[e.show?(i(),h("div",ea)):_("",!0)]),_:1}))}}),aa=V(ta,[["__scopeId","data-v-b06cdb19"]]),N=Se;function na(e,t){let a,o=!1;return()=>{a&&clearTimeout(a),o?a=setTimeout(e,t):(e(),(o=!0)&&setTimeout(()=>o=!1,t))}}function Ce(e){return e.startsWith("/")?e:`/${e}`}function qe(e){const{pathname:t,search:a,hash:o,protocol:n}=new URL(e,"http://a.com");if(Et(e)||e.startsWith("#")||!n.startsWith("http")||!Wt(t))return e;const{site:u}=N(),r=t.endsWith("/")||t.endsWith(".html")?e:e.replace(/(?:(^\.+)\/)?.*$/,`$1${t.replace(/(\.md)?$/,u.value.cleanUrls?"":".html")}${a}${o}`);return Fe(r)}function be({correspondingLink:e=!1}={}){const{site:t,localeIndex:a,page:o,theme:n,hash:u}=N(),r=P(()=>{var l,p;return{label:(l=t.value.locales[a.value])==null?void 0:l.label,link:((p=t.value.locales[a.value])==null?void 0:p.link)||(a.value==="root"?"/":`/${a.value}/`)}});return{localeLinks:P(()=>Object.entries(t.value.locales).flatMap(([l,p])=>r.value.label===p.label?[]:{text:p.label,link:oa(p.link||(l==="root"?"/":`/${l}/`),n.value.i18nRouting!==!1&&e,o.value.relativePath.slice(r.value.link.length-1),!t.value.cleanUrls)+u.value})),currentLang:r}}function oa(e,t,a,o){return t?e.replace(/\/$/,"")+Ce(a.replace(/(^|\/)index\.md$/,"$1").replace(
[…]
