### 2025-10-21 050322d
revised short paper
