+++ docs/paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'JustRelax.jl: A Julia package for geodynamic modeling with matrix-free solvers'
+title: '`JustRelax.jl`: A Julia package for geodynamic modeling with matrix-free solvers'
 tags:
   - Julia
   - Geosciences
@@ -45,56 +45,53 @@ affiliations:
  - name: Frankfurt University, Germany
    index: 5
 
-date: 21 October 2025
+date: 21 August 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-JustRelax.jl is an open source, highly portable, and high-performance Julia package for geodynamic modeling. It employs the Accelerated Pseudo-Transient (APT) method—a matrix-free, embarrassingly parallel iterative method—to solve the Stokes and diffusion equations, making it well-suited to exploit GPU hardware in modern distributed HPC systems.
+`JustRelax.jl` is an open source, highly portable, and high-performance [Julia](https://julialang.org/) [@bezanson2017julia] package designed for geodynamic modeling. It employs the Accelerated Pseudo-Transient (APT) method —a matrix-free, embarrassingly parallel iterative method— to solve the Stokes and diffusion equations, making it well-suited to exploit Graphics Processing Unit (GPU) hardware accelerators.
 
-The package incorporates features critical to computational geodynamics, including complex non-linear rheologies, free surface tracking, and a particle-in-cell method to advect material phases and properties. Its modular design outsources specialized functionality to external packages, reducing the core code base and improving maintainability and reusability.
+`JustRelax.jl` incorporates a wide range of features critical to computational geodynamics. These include complex and highly non-linear rheologies, free surface, and a particle-in-cell method to track and advect material information. Several of the features available in  `JustRelax.jl` are are outsourced to specialized external packages, reducing the core code base, and improving maintainability and reusability.
 
-JustRelax.jl and its dependencies are written in [Julia](https://julialang.org/) [@bezanson2017julia], which lowers the barrier to contribution compared to traditional HPC languages (C/C++/Fortran). Julia's interactive environment enables rapid prototyping in a single high-level language, similar to Python and MATLAB, while maintaining high performance.
+<!-- Most of `JustRelax.jl`'s core dependencies are fully written in the [Julia programming language](https://julialang.org/) [@bezanson2017julia], which facilitates contributions from users without extensive experience in programming, compared to other compiled languages common in HPC-focused software (C/C++/Fortran). The interactive environment of Julia allows users to prototype and implement new code in a single high-level language, improving the overall development experience in similar fashion to Python and MATLAB. -->
 
 # Statement of Need
 
 Simulating Earth's thermo-mechanical evolution requires solving coupled, nonlinear Stokes and heat‑diffusion problems across large domains with sharp material contrasts, complex rheologies and long time scales. These calculations are computationally demanding and traditionally rely on monolithic, CPU‑oriented codes built around matrix‑based linear solvers. Such designs are hard to port efficiently to modern heterogeneous HPC systems (e.g. multi‑GPU nodes).
 
-JustRelax.jl addresses this gap by providing a compact, modular, GPU‑ready toolkit implemented in Julia. It implements a matrix‑free Accelerated Pseudo‑Transient (APT) solver that reduces global linear‑algebra bottlenecks and ports efficiently to GPUs, while outsourcing local material physics and advection (particle‑in‑cell) to specialized external packages. This keeps the core solver as contained as possible, and makes the software more flexible. Julia’s high‑level syntax and package manager simplify installation, scripting and extension, lowering the barrier for students and researchers to prototype and test new physics.
[…]
