### 2025-10-21 f5de93c
Update docs/paper/paper.md
Co-authored-by: Ludovic Räss <61313342+luraess@users.noreply.github.com>
