+++ previews/PR343/assets/chunks/@localSearchIndexroot.CtFANGL-.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":145,"nextId":145,"documentIds":{"0":"/JustRelax.jl/previews/PR343/#What-is-JustRelax.jl?","1":"/JustRelax.jl/previews/PR343/#How-to-Install-JustRelax.jl?","2":"/JustRelax.jl/previews/PR343/#funding","3":"/JustRelax.jl/previews/PR343/man/Blankenbach#Blankenbach-benchmark","4":"/JustRelax.jl/previews/PR343/man/Blankenbach#Initialize-packages","5":"/JustRelax.jl/previews/PR343/man/Blankenbach#script","6":"/JustRelax.jl/previews/PR343/man/Blankenbach#Model-domain","7":"/JustRelax.jl/previews/PR343/man/Blankenbach#rheology","8":"/JustRelax.jl/previews/PR343/man/Blankenbach#Initialize-particles","9":"/JustRelax.jl/previews/PR343/man/Blankenbach#Temperature-anomaly","10":"/JustRelax.jl/previews/PR343/man/Blankenbach#Stokes-and-heat-diffusion-arrays","11":"/JustRelax.jl/previews/PR343/man/Blankenbach#Initialize-thermal-profile-and-viscosity-fields","12":"/JustRelax.jl/previews/PR343/man/Blankenbach#Boundary-conditions","13":"/JustRelax.jl/previews/PR343/man/Blankenbach#Pseuo-transient-coefficients","14":"/JustRelax.jl/previews/PR343/man/Blankenbach#Just-before-solving-the-problem...","15":"/JustRelax.jl/previews/PR343/man/Blankenbach#Advancing-one-time-step","16":"/JustRelax.jl/previews/PR343/man/Blankenbach#visualization","17":"/JustRelax.jl/previews/PR343/man/Blankenbach#fields","18":"/JustRelax.jl/previews/PR343/man/Blankenbach#Final-model","19":"/JustRelax.jl/previews/PR343/man/GeoParams#List-of-GeoParams-functions","20":"/JustRelax.jl/previews/PR343/man/JustPIC#List-of-JustPIC-functions","21":"/JustRelax.jl/previews/PR343/man/ShearBands#ShearBand-benchmark","22":"/JustRelax.jl/previews/PR343/man/ShearBands#Initialize-packages","23":"/JustRelax.jl/previews/PR343/man/ShearBands#script","24":"/JustRelax.jl/previews/PR343/man/ShearBands#Model-domain","25":"/JustRelax.jl/previews/PR343/man/ShearBands#Physical-properties-using-GeoParams","26":"/JustRelax.jl/previews/PR343/man/ShearBands#rheology","27":"/JustRelax.jl/previews/PR343/man/ShearBands#Phase-anomaly","28":"/JustRelax.jl/previews/PR343/man/ShearBands#Stokes-arrays","29":"/JustRelax.jl/previews/PR343/man/ShearBands#Initialize-viscosity-fields","30":"/JustRelax.jl/previews/PR343/man/ShearBands#Boundary-conditions","31":"/JustRelax.jl/previews/PR343/man/ShearBands#Pseuo-transient-coefficients","32":"/JustRelax.jl/previews/PR343/man/ShearBands#Just-before-solving-the-problem...","33":"/JustRelax.jl/previews/PR343/man/ShearBands#Advancing-one-time-step","34":"/JustRelax.jl/previews/PR343/man/ShearBands#visualization","35":"/JustRelax.jl/previews/PR343/man/ShearBands#fields","36":"/JustRelax.jl/previews/PR343/man/ShearBands#Final-model","37":"/JustRelax.jl/previews/PR343/man/advection#Field-advection","38":"
[…]
