+++ previews/PR343/assets/chunks/VPLocalSearchBox.BXfimL6H.js
@@ -1,4 +1,4 @@
-var Nt=Object.defineProperty;var Ft=(a,e,t)=>e in a?Nt(a,e,{enumerable:!0,configurable:!0,writable:!0,value:t}):a[e]=t;var Re=(a,e,t)=>Ft(a,typeof e!="symbol"?e+"":e,t);import{V as Ot,D as le,h as ge,aB as et,aC as Ct,aD as Rt,aE as Mt,q as $e,aF as At,d as Lt,aG as tt,p as he,aH as Dt,ac as Pt,s as zt,aI as Vt,v as Me,P as fe,O as _e,aJ as $t,ak as jt,W as Bt,R as Wt,$ as Kt,b as Jt,o as H,j as x,a0 as qt,a2 as Ut,k as L,a5 as Gt,ad as Ht,c as Z,e as Se,n as st,B as nt,F as it,a as pe,t as ve,aa as Qt,aK as rt,aL as Yt,aq as Zt,av as Xt,aM as es,_ as ts}from"./framework.B94pNcpO.js";import{u as ss,c as ns}from"./theme.YGEUpIk4.js";const is={root:()=>Ot(()=>import("./@localSearchIndexroot.BTfzu0Vm.js"),[])};/*!
+var Nt=Object.defineProperty;var Ft=(a,e,t)=>e in a?Nt(a,e,{enumerable:!0,configurable:!0,writable:!0,value:t}):a[e]=t;var Re=(a,e,t)=>Ft(a,typeof e!="symbol"?e+"":e,t);import{V as Ot,D as le,h as ge,aB as et,aC as Ct,aD as Rt,aE as Mt,q as $e,aF as At,d as Lt,aG as tt,p as he,aH as Dt,ac as Pt,s as zt,aI as Vt,v as Me,P as fe,O as _e,aJ as $t,ak as jt,W as Bt,R as Wt,$ as Kt,b as Jt,o as H,j as x,a0 as qt,a2 as Ut,k as L,a5 as Gt,ad as Ht,c as Z,e as Se,n as st,B as nt,F as it,a as pe,t as ve,aa as Qt,aK as rt,aL as Yt,aq as Zt,av as Xt,aM as es,_ as ts}from"./framework.B94pNcpO.js";import{u as ss,c as ns}from"./theme.CQgmfMIs.js";const is={root:()=>Ot(()=>import("./@localSearchIndexroot.CtFANGL-.js"),[])};/*!
 * tabbable 6.2.0
 * @license MIT, https://github.com/focus-trap/tabbable/blob/master/LICENSE
[…]
