### 2025-10-21 #9022 @albert-de-montserrat
@kthyng Thanks for your comment. We have addressed the following:

- Word count is now below 1000 words.
- Added missing DOIs.

As for potential reviewers, here is a list with some suggestions:

| Name             | GH handle    | Affiliation        |
|------------------|------------- |--------------------|
| Jordi Bolibar    | JordiBolibar | CNRS, IGE          |
| Francesc Verdugo | fverdugo     | VU Amsterdam       |
| Nobuaki Fuji     | seismobassoon| IPGP               |
| Rene Gassmoeller | gassmoeller  | Geomar             |
| Cedric Thielout  | cedrict      | Utrecht University |
