### 2025-10-21 44283d3
working GPU phase ratios w/ arrays (`@MVector`)
