+++ src/phases/PhaseRatios.jl
@@ -85,33 +85,29 @@ function phase_ratios_center_from_arrays!(phase_ratios::JustPIC.PhaseRatios, pha
 end
 
 @parallel_indices (I...) function phase_ratios_center_from_arrays_kernel!(
-        ratio_centers, phase_arrays::NTuple{N, AbstractArray}
-    ) where {N}
+    ratio_centers, phase_arrays::NTuple{N, AbstractArray}
+) where {N}
 
     values = ntuple(i -> phase_arrays[i][I...], Val(N))
-
-    total = zero(eltype(values[1]))
-    for i in 1:N
-        total += values[i]
-    end
+    # Sum
+    total = sum(values)
 
     # Normalize
-    normalized_values = ntuple(i -> values[i] / total, Val(N))
-    # Manual clamping
-    normalized_values = ntuple(i -> min(max(normalized_values[i], 0.0), 1.0), Val(N))
-    # Clean up very small values (round to zero if < 1e-5)
-    cleaned_values = ntuple(i -> normalized_values[i] < 1.0e-5 ? zero(eltype(normalized_values)) : normalized_values[i], Val(N))
-    # Renormalize to ensure sum = 1
-    final_total = zero(eltype(values[1]))  # Fixed: was 'value[1]'
-    for i in 1:N
-        final_total += cleaned_values[i]
-    end
+    normalized = values ./ total
 
-    normalized_values = ntuple(i -> cleaned_values[i] / final_total, Val(N))
+    # Clamp
+    clamped = map(x -> min(max(x, 0.0), 1.0), normalized)
 
-    # Update phase ratios array
+    # Threshold small values
+    cleaned = map(x -> x < 1e-5 ? zero(eltype(values)) : x, clamped)
+
+    # Renormalize
+    final_total = sum(cleaned)
+    final = cleaned ./ final_total
+
+    # Write back
     for k in 1:N
-        @index ratio_centers[k, I...] = normalized_values[k]
+        @index ratio_centers[k, I...] = final[k]
     end
 
     return nothing
@@ -137,84 +133,75 @@ end
 @parallel_indices (I...) function phase_ratios_vertex_from_arrays_kernel!(
         ratio_vertices, phase_arrays::NTuple{N, AbstractArray}, xci::NTuple{ND}, xvi::NTuple{ND}, di::NTuple{ND, T}
     ) where {N, ND, T}
-    w = ntuple(_ -> zero(T), Val(N))
+    w_vals = @MVector zeros(T, N)
+    total_weight = zero(T)
+
     # Vertex position
     cell_vertex = ntuple(d -> xvi[d][I[d]], Val(ND))
-    if ND == 2
-        ni = size(first(phase_arrays))
-        total_weight = zero(T)
-        w_vals = ntuple(_ -> zero(T), Val(N))
-        x_v, y_v = cell_vertex
+    ni = size(first(phase_arrays))
 
+    if ND == 2
         for offset₁ in -1:0, offset₂ in -1:0
             i_cell = I[1] + offset₁
             j_cell = I[2] + offset₂
             if 1 <= i_cell <= ni[1] && 1 <= j_cell <= ni[2]
                 x_c = xci[1][i_cell]
                 y_c = xci[2][j_cell]
-                wx = 1.0 - abs(x_v - x_c) / di[1]
-                wy = 1.0 - abs(y_v - y_c) / di[2]
+                wx = muladd(-abs(cell_vertex[1] - x_c), inv(di[1]), 1.0)
+                wy = muladd(-abs(cell_vertex[2] - y_c), inv(di[2]), 1.0)
                 weight = wx * wy
                 total_weight += weight
 
                 for k in 1:N
-                    phase_val = phase_arrays[k][i_cell, j_cell]
[…]
