+++ src/ext/AMDGPU/3D.jl
@@ -491,7 +491,7 @@ function JR3D.rotate_stress!(
 end
 
 # Phase ratios with arrays
-function JR3D.update_phase_ratios!(
+function JR3D.update_phase_ratios_3D!(
         phase_ratios::JustPIC.PhaseRatios{AMDGPUBackend, T},
         phase_arrays::NTuple{N, ROCArray{U, 3}},
         xci,
