+++ previews/PR357/assets/man_backend.md.B4EvVDT8.lean.js
@@ -1,0 +1,1 @@
+import{_ as e,c as s,o as i,aA as t}from"./chunks/framework.BoMdGI0i.js";const k=JSON.parse('{"title":"Selecting the backend","description":"","frontmatter":{},"headers":[],"relativePath":"man/backend.md","filePath":"man/backend.md","lastUpdated":null}'),l={name:"man/backend.md"};function n(d,a,p,h,o,c){return i(),s("div",null,[...a[0]||(a[0]=[t("",7)])])}const u=e(l,[["render",n]]);export{k as __pageData,u as default};

[…]
