### 2025-10-22 44691fc
add Makie extension
