+++ src/JustRelax.jl
@@ -90,4 +90,7 @@ include("IO/DataIO.jl")
 include("types/type_conversions.jl")
 export Array, copy
 
+function plot_particles end
+function plot_field end
+
 end # module
