+++ miniapps/benchmarks/stokes3D/StickyAirSubduction/Subduction3D.jl
@@ -66,7 +66,7 @@ function main3D(li, origin, phases_GMG, igg; nx = 16, ny = 16, nz = 16, figdir =
 
     # Initialize particles -------------------------------
     nxcell, max_xcell, min_xcell = 125, 150, 75
-       particles = init_particles(
+    particles = init_particles(
         backend_JP, nxcell, max_xcell, min_xcell, xvi...
     )
     # velocity grids

[…]
