### 2025-10-27 0b093da
export functions
