+++ src/JustRelax.jl
@@ -93,4 +93,6 @@ export Array, copy
 function plot_particles end
 function plot_field end
 
+export plot_particles, plot_field
+
 end # module
