+++ miniapps/benchmarks/stokes2D/sinking_block/SinkingBlock2D_WENO5.jl
@@ -103,22 +103,7 @@ end
     return nothing
 end
 
-function plot_particles(particles, pPhases)
-    p = particles.coords
-    # pp = [argmax(p) for p in phase_ratios.center] #if you want to plot it in a heatmap rather than scatter
-    ppx, ppy = p
-    # pxv = ustrip.(dimensionalize(ppx.data[:], km, CharDim))
-    # pyv = ustrip.(dimensionalize(ppy.data[:], km, CharDim))
-    pxv = ppx.data[:]
-    pyv = ppy.data[:]
-    clr = pPhases.data[:]
-    # clr = pϕ.data[:]
-    idxv = particles.index.data[:]
-    f, ax, h = scatter(Array(pxv[idxv]), Array(pyv[idxv]), color = Array(clr[idxv]), colormap = :roma, markersize = 1)
-    Colorbar(f[1, 2], h)
-    return f
-end
-# --------------------------------------------------------------------------------
+# # --------------------------------------------------------------------------------
 # BEGIN MAIN SCRIPT
 # --------------------------------------------------------------------------------
 function sinking_block2D(igg; ar = 8, ny = 16, nx = ny * 8, figdir = "figs2D", thermal_perturbation = :circular)
