### 2025-10-28 ab9a914
fix wrong interpolation at face and midpoints
