### 2025-10-28 67d6e5b
revert 3D `shear2center_kernel` as locations are staggered
