+++ simple-rheology/Rift2D_displacement.jl
@@ -1,574 +1,0 @@
-const isCUDA = false
-#const isCUDA = true
-
-@static if isCUDA
-    using CUDA
-end
-
-using JustRelax, JustRelax.JustRelax2D, JustRelax.DataIO
-import JustRelax.@cell
-
-const backend_JR = @static if isCUDA
-    CUDABackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
-else
-    JustRelax.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
-end
-
-using ParallelStencil, ParallelStencil.FiniteDifferences2D
-
-@static if isCUDA
-    @init_parallel_stencil(CUDA, Float64, 2)
-else
-    @init_parallel_stencil(Threads, Float64, 2)
-end
-
-using JustPIC, JustPIC._2D
-# Threads is the default backend_JR,
-# to run on a CUDA GPU load CUDA.jl (i.e. "using CUDA") at the beginning of the script,
-# and to run on an AMD GPU load AMDGPU.jl (i.e. "using AMDGPU") at the beginning of the script.
-const backend_JP = @static if isCUDA
-    CUDABackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
-else
-    JustPIC.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
-end
-
-# Load script dependencies
-using GeoParams, CairoMakie
-#using  CairoMakie
-using JLD2
-
-
-# Load file with all the rheology configurations
-include("Rift2DSetupFaultInclusion.jl")
-include("RiftRheology.jl")
-
-
-## SET OF HELPER FUNCTIONS PARTICULAR FOR THIS SCRIPT --------------------------------
-
-import ParallelStencil.INDICES
-const idx_k = INDICES[2]
-macro all_k(A)
-    esc(:($A[$idx_k]))
-end
-
-function copyinn_x!(A, B)
-    @parallel function f_x(A, B)
-        @all(A) = @inn_x(B)
-        return nothing
-    end
-
-    @parallel f_x(A, B)
-end
-
-@parallel_indices (i, j) function update_Dirichlet_mask!(mask, phase_ratio_vertex, air_phase)
-    @inbounds mask[i + 1, j] = @index(phase_ratio_vertex[air_phase, i, j]) == 1
-    nothing
-end
-
-# Initial pressure profile - not accurate
-@parallel function init_P!(P, ρg, z)
-    @all(P) = abs(@all(ρg) * @all_k(z)) #* <(@all_k(z), 0.0)
-    return nothing
-end
-
-function apply_pure_shear(Vx, Vy, εbg, xvi, lx, ly)
-    xv, yv = xvi
-
-    @parallel_indices (i, j) function pure_shear_x!(Vx, εbg, lx)
-        xi = xv[i]
-        Vx[i, j + 1] = εbg * (xi - lx * 0.5)
-        return nothing
-    end
-
-    @parallel_indices (i, j) function pure_shear_y!(Vy, εbg, ly)
-        yi = yv[j]
-        Vy[i + 1, j] = abs(yi) * εbg
-        return nothing
-    end
-
-    nx, ny = size(Vx)
-    @parallel (1:nx, 1:(ny - 2)) pure_shear_x!(Vx, εbg, lx)
-    nx, ny = size(Vy)
-    @parallel (1:(nx - 2), 1:ny) pure_shear_y!(Vy, εbg, ly)
-
-    return nothing
-end
-
-function update_bc!(Ux, Uy, εbg, xvi, lx, ly, dt)
-    xv, yv = xvi
-
[…]
