+++ src/stokes/PressureKernels.jl
@@ -195,15 +195,15 @@ function _compute_P!(P, P0, ∇V, Q, ΔTc, α, η, K, dt, r, θ_dτ)
 end
 
 
-## Displacement based 
+## Displacement based
 
 # With GeoParams
 
 @parallel_indices (I...) function compute_P_displacement!(
-        P, P0, RP, ∇U, Q, η, rheology::NTuple{N, MaterialParams}, phase, r, lτ, Vpdτ, θ_dτ,dt, args
+        P, P0, RP, ∇U, Q, η, rheology::NTuple{N, MaterialParams}, phase, r, lτ, Vpdτ, θ_dτ, dt, args
     ) where {N}
     K = get_bulk_modulus(rheology, phase[I...])
-    @inbounds RP[I...], P[I...] = _compute_P_displacement!(P[I...], P0[I...], ∇U[I...], Q[I...], η[I...], K,  r, lτ, Vpdτ, θ_dτ,dt)
+    @inbounds RP[I...], P[I...] = _compute_P_displacement!(P[I...], P0[I...], ∇U[I...], Q[I...], η[I...], K, r, lτ, Vpdτ, θ_dτ, dt)
     return nothing
 end
 
@@ -216,8 +216,8 @@ function compute_P_displacement!(
         η,
         rheology::NTuple{N, MaterialParams},
         phase_ratio,
-        r, 
-        lτ, 
+        r,
+        lτ,
         Vpdτ,
         θ_dτ,
         dt,
@@ -235,8 +235,8 @@ function compute_P_displacement!(
         η,
         rheology::NTuple{N, MaterialParams},
         phase_ratio,
-        r, 
-        lτ, 
+        r,
+        lτ,
         Vpdτ,
         θ_dτ,
         dt;
@@ -246,7 +246,7 @@ function compute_P_displacement!(
     ) where {N}
     ni = size(P)
     @parallel (@idx ni) compute_P_displacement_kernel!(
-        P, P0, RP, ∇U, Q, η, rheology, phase_ratio.center, r, lτ, Vpdτ, θ_dτ,dt, ΔTc, melt_fraction
+        P, P0, RP, ∇U, Q, η, rheology, phase_ratio.center, r, lτ, Vpdτ, θ_dτ, dt, ΔTc, melt_fraction
     )
     return nothing
 end
@@ -260,8 +260,8 @@ end
         η,
         rheology::NTuple{N, MaterialParams},
         phase_ratio::C,
-        r, 
-        lτ, 
+        r,
+        lτ,
         Vpdτ,
         θ_dτ,
         dt,
@@ -269,7 +269,7 @@ end
         ::Nothing,
     ) where {N, C <: JustRelax.CellArray}
     K = fn_ratio(get_bulk_modulus, rheology, @cell(phase_ratio[I...]))
-    @inbounds RP[I...], P[I...] = _compute_P_displacement!(P[I...], P0[I...], ∇U[I...], Q[I...], η[I...], K,  r, lτ, Vpdτ, θ_dτ,dt)
+    @inbounds RP[I...], P[I...] = _compute_P_displacement!(P[I...], P0[I...], ∇U[I...], Q[I...], η[I...], K, r, lτ, Vpdτ, θ_dτ, dt)
     return nothing
 end
 
@@ -282,8 +282,8 @@ end
         η,
         rheology::NTuple{N, MaterialParams},
         phase_ratio::C,
-        r, 
-        lτ, 
+        r,
[…]
