### 2025-10-29 839f8fd
Fix Project.toml file
