+++ Project.toml
@@ -18,7 +18,6 @@ MuladdMacro = "46d2c3a1-f734-5fdb-9937-b9b9aeba4221"
 ParallelStencil = "94395366-693c-11ea-3b26-d9b7aac5d958"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
-Runic = "62bfec6d-59d7-401d-8490-b29ee721c001"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
@@ -47,7 +46,6 @@ MPI = "0.20"
 MuladdMacro = "0.2"
 ParallelStencil = "0.14.2"
 Reexport = "1.2.2"
-Runic = "1.5.1"
 StaticArrays = "1"
 Statistics = "1"
 TOML = "1.0.3"
