+++ miniapps/benchmarks/stokes2D/Blankenbach2D/Benchmark2D_sgd_GMD.jl
@@ -1,543 +1,0 @@
-const isCUDA = false
-# const isCUDA = true
-
-@static if isCUDA
-    using CUDA
-end
-
-using JustRelax, JustRelax.JustRelax2D, JustRelax.DataIO
-
-const backend_JR = @static if isCUDA
-    CUDABackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
-else
-    JustRelax.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
-end
-
-using ParallelStencil, ParallelStencil.FiniteDifferences2D
-
-@static if isCUDA
-    @init_parallel_stencil(CUDA, Float64, 2)
-else
-    @init_parallel_stencil(Threads, Float64, 2)
-end
-
-using JustPIC, JustPIC._2D
-# Threads is the default backend,
-# to run on a CUDA GPU load CUDA.jl (i.e. "using CUDA") at the beginning of the script,
-# and to run on an AMD GPU load AMDGPU.jl (i.e. "using AMDGPU") at the beginning of the script.
-const backend = @static if isCUDA
-    CUDABackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
-else
-    JustPIC.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
-end
-# Load script dependencies
-using Printf, LinearAlgebra, GeoParams, CairoMakie, CellArrays
-
-# Load file with all the rheology configurations
-include("Blankenbach_Rheology.jl")
-
-## SET OF HELPER FUNCTIONS PARTICULAR FOR THIS SCRIPT --------------------------------
-
-function copyinn_x!(A, B)
-
-    @parallel function f_x(A, B)
-        @all(A) = @inn_x(B)
-        return nothing
-    end
-
-    return @parallel f_x(A, B)
-end
-
-# Initial thermal profile
-@parallel_indices (i, j) function init_T!(T, y)
-    depth = -y[j]
-
-    dTdZ = (1273 - 273) / 1000.0e3
-    offset = 273.0e0
-    T[i, j] = (depth) * dTdZ + offset
-    return nothing
-end
-
-# Thermal rectangular perturbation
-function rectangular_perturbation!(T, xc, yc, r, xvi)
-    @parallel_indices (i, j) function _rectangular_perturbation!(T, xc, yc, r, x, y)
-        if ((x[i] - xc)^2 ≤ r^2) && ((y[j] - yc)^2 ≤ r^2)
-            T[i + 1, j] += 20.0
-        end
-        return nothing
-    end
-    nx, ny = size(T)
-    @parallel (1:(nx - 2), 1:ny) _rectangular_perturbation!(T, xc, yc, r, xvi...)
-    return nothing
-end
-## END OF HELPER FUNCTION ------------------------------------------------------------
-
-## BEGIN OF MAIN SCRIPT --------------------------------------------------------------
-function main2D(igg; ar = 1, nx = 32, ny = 32, nit = 1.0e1, figdir = "figs2D", do_vtk = false, finalize_MPI = true)
-
-    # Physical domain ------------------------------------
-    ly = 1000.0e3               # domain length in y
-    lx = ly                   # domain length in x
-    ni = nx, ny               # number of cells
-    li = lx, ly               # domain length in x- and y-
-    di = @. li / ni           # grid step in x- and -y
-    origin = 0.0, -ly             # origin coordinates
-    grid = Geometry(ni, li; origin = origin)
-    (; xci, xvi) = grid # nodes at the center and vertices of the cells
-    # ----------------------------------------------------
-
-    # Physical properties using GeoParams ----------------
[…]
