### 2025-11-03 3dc1182
add tensile cap (VM) to plastic stress correction
