+++ src/variational_stokes/StressKernels.jl
@@ -59,7 +59,13 @@
         τIIv_ij = second_invariant(dτijv .+ τijv)
 
         # yield function @ center
-        Fv = τIIv_ij - Cv * cosϕv - Pv_ij * sinϕv
+        Fv = if Pv_ij ≥ 0
+            # DP in extension
+            τIIv_ij - Cv * cosϕv - Pv_ij * sinϕv
+        else
+            # VM in extension
+            τIIv_ij - Cv
+        end
 
         if is_pl && !iszero(τIIv_ij) && Fv > 0
             # stress correction @ vertex
@@ -99,7 +105,14 @@
             dτij = compute_stress_increment(τij, τij_o, ηij, εij, _Gdt, dτ_r)
             τII_ij = second_invariant(dτij .+ τij)
             # yield function @ center
-            F = τII_ij - C * cosϕ - Pr[I...] * sinϕ
+            P_ij = Pr[I...]
+            F = if P_ij ≥ 0
+                # DP in extension
+                τII_ij - C * cosϕ - P_ij * sinϕ
+            else
+                # VM in extension
+                τII_ij - C
+            end
 
             τII_ij = if is_pl && !iszero(τII_ij) && F > 0
                 # stress correction @ center
@@ -223,7 +236,13 @@ end
         τIIv_ij = second_invariant(τijv .+ dτijv)
 
         # yield function @ vertex
-        Fv = τIIv_ij - Cv * cosϕv - Pv_ij * sinϕv
+                Fv = if Pv_ij ≥ 0
+            # DP in extension
+            τIIv_ij - Cv * cosϕv - Pv_ij * sinϕv
+        else
+            # VM in extension
+            τIIv_ij - Cv
+        end
         if is_pl && !iszero(τIIv_ij) && Fv > 0
             # stress correction @ vertex
             λv[1][I...] =
@@ -288,7 +307,13 @@ end
         τIIv_ij = second_invariant(τijv .+ dτijv)
 
         # yield function @ vertex
-        Fv = τIIv_ij - Cv * cosϕv - Pv_ij * sinϕv
+                Fv = if Pv_ij ≥ 0
+            # DP in extension
+            τIIv_ij - Cv * cosϕv - Pv_ij * sinϕv
+        else
+            # VM in extension
+            τIIv_ij - Cv
+        end
         if is_pl && !iszero(τIIv_ij) && Fv > 0
             # stress correction @ vertex
             λv[2][I...] =
@@ -354,7 +379,13 @@ end
         τIIv_ij = second_invariant(τijv .+ dτijv)
 
         # yield function @ vertex
-        Fv = τIIv_ij - Cv * cosϕv - Pv_ij * sinϕv
+                Fv = if Pv_ij ≥ 0
+            # DP in extension
+            τIIv_ij - Cv * cosϕv - Pv_ij * sinϕv
+        else
+            # VM in extension
+            τIIv_ij - Cv
+        end
         if is_pl && !iszero(τIIv_ij) && Fv > 0
             # stress correction @ vertex
             λv[3][I...] =
@@ -392,7 +423,14 @@ end
         dτij = compute_stress_increment(τij, τij_o, ηij, εij, _Gdt, dτ_r)
         τII_ij = second_invariant(dτij .+ τij)
         # yield function @ center
-        F = τII_ij - C * cosϕ - Pr[I...] * sinϕ
+        P_ij = Pr[I...]
+        F = if P_ij ≥ 0
+            # DP in extension
+            τII_ij - C * cosϕ - P_ij * sinϕ
+        else
+            # VM in extension
+            τII_ij - C
+        end
 
         if is_pl && !iszero(τII_ij) && F > 0
             # stress correction @ center
@@ -496,7 +534,13 @@ end
[…]
