### 2025-11-05 cbc0f69
update installation guide for cloned repos
