+++ docs/src/man/installation.md
@@ -18,6 +18,13 @@ julia> ]
     add JustRelax#main
     ```
 
+If you downloaded or cloned the repository manually, you need to instantiate the package to install all dependencies. Navigate to the directory where you have JustRelax.jl and run:
+```julia
+julia> ]
+(@v1.10) pkg> instatiate
+```
+
+
 After installation, you can test the package by running the following commands:
 ```julia-repl
 using JustRelax
