### 2025-11-06 6bb4ac3
WIP: DYREL kernels
