+++ DYREL/ShearBand2D_AutoTune.jl
@@ -1,0 +1,427 @@
+# using CUDA
+using GeoParams, GLMakie, CellArrays
+using JustRelax, JustRelax.JustRelax2D
+using ParallelStencil
+# @init_parallel_stencil(CUDA, Float64, 2)
+@init_parallel_stencil(Threads, Float64, 2)
+
+const backend = CPUBackend
+# const backend = CUDABackend
+
+using JustPIC, JustPIC._2D
+import JustPIC._2D.GridGeometryUtils as GGU
+
+const backend_JP = JustPIC.CPUBackend
+# const backend_JP = CUDABackend
+
+# HELPER FUNCTIONS ----------------------------------- ----------------------------
+solution(ε, t, G, η) = 2 * ε * η * (1 - exp(-G * t / η))
+
+# Initialize phases on the particles
+function init_phases!(phase_ratios, xci, xvi, circle)
+    ni = size(phase_ratios.center)
+
+    @parallel_indices (i, j) function init_phases!(phases, xc, yc, circle)
+        x, y = xc[i], yc[j]
+        p    = GGU.Point(x, y)
+
+        if GGU.inside(p, circle)
+            @index phases[1, i, j] = 0.0
+            @index phases[2, i, j] = 1.0
+
+        else
+            @index phases[1, i, j] = 1.0
+            @index phases[2, i, j] = 0.0
+
+        end
+        return nothing
+    end
+
+    @parallel (@idx ni) init_phases!(phase_ratios.center, xci..., circle)
+    @parallel (@idx ni .+ 1) init_phases!(phase_ratios.vertex, xvi..., circle)
+
+    return nothing
+end
+
+
+# MAIN SCRIPT --------------------------------------------------------------------
+function main(igg; nx = 64, ny = 64, figdir = "model_figs")
+
+    # Physical domain ------------------------------------
+    ly = 1e0          # domain length in y
+    lx = 2e0           # domain length in x
+    ni = nx, ny       # number of cells
+    li = lx, ly       # domain length in x- and y-
+    di = @. li / ni   # grid step in x- and -y
+    origin = -lx/2, -ly/2     # origin coordinates
+    grid = Geometry(ni, li; origin = origin)
+    (; xci, xvi) = grid # nodes at the center and vertices of the cells
+
+    ### DYREL stuff
+
+    # Bulk viscosity
+    Kb   = 1e-10
+    ηb   = @fill(Kb * dt, ni...)
+    # Select γ
+    γfact = 20           # penalty: multiplier to the arithmetic mean of η
+    γi    = γfact * mean(stokes.viscosity.η)
+    # (Pseudo-)compressibility
+    γ_eff = @zeros(ni...) 
+    γ_num = @fill(γi, ni...)
+    γ_phy = ηb
+    γ_eff = ((γ_phy.*γ_num)./(γ_phy.+γ_num))
+
+    # Diagonal preconditioner arrays
+    Dx     = @zeros(ni[1]-1, ni[2])
+    Dy     = @zeros(ni[1], ni[2]-1)
[…]
