+++ DYREL/stress_kernel.jl
@@ -1,102 +1,105 @@
+function compute_stress_DRYEL!(stokes, rheology, phase_ratios, λ_relaxation, dt)
+    ni = size(phase_ratios.vertex)
+    @parallel (@idx ni) compute_stress_DRYEL!(stokes, rheology, phase_ratios.center, phase_ratios.vertex, λ_relaxation, dt)
+    return nothing
+end
+
+@parallel_indices (I...) function compute_stress_DRYEL!(stokes::JustRelax.StokesArrays, rheology, phase_ratios_center, phase_ratios_vertex, λ_relaxation, dt)
+
+    Base.@propagate_inbounds @inline av(A) = sum(JustRelax2D._gather(A, I...)) / 4
+
+    ni = size(phase_ratios_center)
+
+    ## VERTEX CALCULATION
+    @inbounds begin
+        Ic    = clamped_indices(ni, I...)
+        τij_o = av_clamped(stokes.τ_o.xx, Ic...), av_clamped(stokes.τ_o.yy, Ic...), stokes.τ_o.xy[I...]
+        εij   = av_clamped(stokes.ε.xx, Ic...),   av_clamped(stokes.ε.yy, Ic...),   stokes.ε.xy[I...]
+        λvij  = stokes.λv[I...]
+        ηij   = av_clamped(stokes.viscosity.η, Ic...)
+        Pij   = av_clamped(stokes.P, Ic...)
+        ratio = phase_ratios_vertex[I...]
+        # # compute local stress
+        _, _, τxy_I, _, _, λ_I, _ = compute_local_stress(εij, τij_o, ηij, Pij, λvij, λ_relaxation, rheology, ratio, dt)
+        # update arrays
+        stokes.τ.xy[I...] = τxy_I
+        stokes.λv[I...]   = λ_I
+
+        ## CENTER CALCULATION
+        !all(I .≤ ni) && continue
+        τij_o = stokes.τ_o.xx[I...], stokes.τ_o.yy[I...], stokes.τ_o.xy_c[I...]
+        εij   = stokes.ε.xx[I...], stokes.ε.yy[I...], av(stokes.ε.xy)
+        λij   = stokes.λ[I...]
+        ηij   = stokes.viscosity.η[I...]
+        Pij   = stokes.P[I...]
+        ratio = phase_ratios_center[I...]
+
+        # compute local stress
+        τxx_I, τyy_I, τxy_I, τII_I, ηvep_I, λ_I, ΔPψ_I = compute_local_stress(εij, τij_o, ηij, Pij, λij, λ_relaxation, rheology, ratio, dt)
+        # update arrays
+        stokes.τ.xx[I...], stokes.τ.yy[I...], stokes.τ.xy_c[I...] = τxx_I, τyy_I, τxy_I
+        stokes.τ.II[I...]            = τII_I
+        stokes.viscosity.η_vep[I...] = ηvep_I
+        stokes.λ[I...]               = λ_I
+        stokes.ΔPψ[I...]             = ΔPψ_I
+    end
+
+    return nothing
+end
+
 @generated function compute_local_stress(εij, τij_o, η, P, λ, λ_relaxation, rheology, phase_ratio::SVector{N}, dt) where N
     quote
         @inline
+        # iterate over phases
         v_phases = Base.@ntuple $N phase -> begin
+            # get phase ratio
             ratio_I = phase_ratio[phase]
-            v =  if iszero(ratio_I)  
[…]
