+++ DYREL/ShearBand2D_AutoTune.jl
@@ -43,6 +43,7 @@ function init_phases!(phase_ratios, xci, xvi, circle)
     return nothing
 end
 
+using Statistics
 
 # MAIN SCRIPT --------------------------------------------------------------------
 function main(igg; nx = 64, ny = 64, figdir = "model_figs")
@@ -56,34 +57,6 @@ function main(igg; nx = 64, ny = 64, figdir = "model_figs")
     origin = -lx/2, -ly/2     # origin coordinates
     grid = Geometry(ni, li; origin = origin)
     (; xci, xvi) = grid # nodes at the center and vertices of the cells
-
-    ### DYREL stuff
-
-    # Bulk viscosity
-    Kb   = 1e-10
-    ηb   = @fill(Kb * dt, ni...)
-    # Select γ
-    γfact = 20           # penalty: multiplier to the arithmetic mean of η
-    γi    = γfact * mean(stokes.viscosity.η)
-    # (Pseudo-)compressibility
-    γ_eff = @zeros(ni...) 
-    γ_num = @fill(γi, ni...)
-    γ_phy = ηb
-    γ_eff = ((γ_phy.*γ_num)./(γ_phy.+γ_num))
-
-    # Diagonal preconditioner arrays
-    Dx     = @zeros(ni[1]-1, ni[2])
-    Dy     = @zeros(ni[1], ni[2]-1)
-    # maximum eigenvalue estimates
-    λmaxVx = @zeros(ni[1]-1, ni[2])
-    λmaxVy = @zeros(ni[1], ni[2]-1)
-
-    βVx = @zeros(ni[1]-1, ni[2])
-    βVy = @zeros(ni[1], ni[2]-1)
-    cVx = @zeros(ni[1]-1, ni[2])
-    cVy = @zeros(ni[1], ni[2]-1)
-    αVx = @zeros(ni[1]-1, ni[2])
-    αVy = @zeros(ni[1], ni[2]-1)
 
     # Physical properties using GeoParams ----------------
     τ_y      = 1.6           # yield stress. If do_DP=true, τ_y stand for the cohesion: c*cos(ϕ)
@@ -145,6 +118,34 @@ function main(igg; nx = 64, ny = 64, figdir = "model_figs")
     ρg = @zeros(ni...), @zeros(ni...)
     args = (; T = @zeros(ni...), P = stokes.P, dt = dt, perturbation_C = perturbation_C)
 
+     ### DYREL stuff
+
+    # Bulk viscosity
+    Kb   = 1e-10
+    ηb   = @fill(Kb * dt, ni...)
+    # Select γ
+    γfact = 20           # penalty: multiplier to the arithmetic mean of η
+    γi    = γfact * mean(stokes.viscosity.η)
+    # (Pseudo-)compressibility
+    γ_eff = @zeros(ni...) 
+    γ_num = @fill(γi, ni...)
+    γ_phy = ηb
+    γ_eff = ((γ_phy.*γ_num)./(γ_phy.+γ_num))
+
+    # Diagonal preconditioner arrays
+    Dx     = @zeros(ni[1]-1, ni[2])
+    Dy     = @zeros(ni[1], ni[2]-1)
[…]
