+++ DYREL/ShearBand2D_AutoTune.jl
@@ -44,6 +44,7 @@ function init_phases!(phase_ratios, xci, xvi, circle)
 end
 
 using Statistics
+include("DYREL_solver.jl")
 
 # MAIN SCRIPT --------------------------------------------------------------------
 function main(igg; nx = 64, ny = 64, figdir = "model_figs")
@@ -103,7 +104,7 @@ function main(igg; nx = 64, ny = 64, figdir = "model_figs")
 
     # Initialize phase ratios -------------------------------
     phase_ratios = PhaseRatios(backend_JP, length(rheology), ni)
-    radius = 0.1
+    radius = 0.09
     origin = 0e0, 0e0
     circle = GGU.Circle(origin, radius)
     init_phases!(phase_ratios, xci, xvi, circle)
@@ -118,7 +119,24 @@ function main(igg; nx = 64, ny = 64, figdir = "model_figs")
     ρg = @zeros(ni...), @zeros(ni...)
     args = (; T = @zeros(ni...), P = stokes.P, dt = dt, perturbation_C = perturbation_C)
 
-     ### DYREL stuff
+    # Rheology
+    compute_viscosity!(stokes, phase_ratios, args, rheology, (-Inf, Inf))
+    center2vertex!(stokes.viscosity.ηv, stokes.viscosity.η)
+
+    # Boundary conditions
+    flow_bcs = VelocityBoundaryConditions(;
+        free_slip = (left = true, right = true, top = true, bot = true),
+        no_slip   = (left = false, right = false, top = false, bot = false),
+    )
+    stokes.V.Vx .= PTArray(backend)([ x * εbg for x in xvi[1], _ in 1:(ny + 2)])
+    stokes.V.Vy .= PTArray(backend)([-y * εbg for _ in 1:(nx + 2), y in xvi[2]])
+    stokes.V.Vx[2:end-1,:] .= 0 # ensure non zero initial pressure residual
+    stokes.V.Vy[:,2:end-1] .= 0 # ensure non zero initial pressure residual
+    
+    flow_bcs!(stokes, flow_bcs) # apply boundary conditions
+    update_halo!(@velocity(stokes)...)
+
+    ### DYREL stuff
 
     # Bulk viscosity
     Kb   = 1e-10
@@ -131,6 +149,7 @@ function main(igg; nx = 64, ny = 64, figdir = "model_figs")
     γ_num = @fill(γi, ni...)
     γ_phy = ηb
     γ_eff = ((γ_phy.*γ_num)./(γ_phy.+γ_num))
+    c_fact   = 0.5          # damping factor
 
     # Diagonal preconditioner arrays
     Dx     = @zeros(ni[1]-1, ni[2])
@@ -138,38 +157,46 @@ function main(igg; nx = 64, ny = 64, figdir = "model_figs")
     # maximum eigenvalue estimates
     λmaxVx = @zeros(ni[1]-1, ni[2])
     λmaxVy = @zeros(ni[1], ni[2]-1)
+    dVxdτ  = @zeros(ni[1]-1, ni[2])
+    dVydτ  = @zeros(ni[1], ni[2]-1)
+    dτVx   = @zeros(ni[1]-1, ni[2])
+    dτVy   = @zeros(ni[1], ni[2]-1)
+    dVx    = @zeros(ni[1]-1, ni[2])
+    dVy    = @zeros(ni[1], ni[2]-1)
+    βVx    = @zeros(ni[1]-1, ni[2])
+    βVy    = @zeros(ni[1], ni[2]-1)
+    cVx    = @zeros(ni[1]-1, ni[2])
+    cVy    = @zeros(ni[1], ni[2]-1)
+    αVx    = @zeros(ni[1]-1, ni[2])
[…]
