+++ DYREL/stress_kernels.jl
@@ -11,7 +11,7 @@ end
     ni = size(phase_ratios_center)
 
     ## VERTEX CALCULATION
-    @inbounds begin
+    # @inbounds begin
         Ic    = clamped_indices(ni, I...)
         τij_o = av_clamped(stokes.τ_o.xx, Ic...), av_clamped(stokes.τ_o.yy, Ic...), stokes.τ_o.xy[I...]
         εij   = av_clamped(stokes.ε.xx, Ic...),   av_clamped(stokes.ε.yy, Ic...),   stokes.ε.xy[I...]
@@ -26,23 +26,24 @@ end
         stokes.λv[I...]   = λ_I
 
         ## CENTER CALCULATION
-        !all(I .≤ ni) && continue
-        τij_o = stokes.τ_o.xx[I...], stokes.τ_o.yy[I...], stokes.τ_o.xy_c[I...]
-        εij   = stokes.ε.xx[I...], stokes.ε.yy[I...], av(stokes.ε.xy)
-        λij   = stokes.λ[I...]
-        ηij   = stokes.viscosity.η[I...]
-        Pij   = stokes.P[I...]
-        ratio = phase_ratios_center[I...]
-
-        # compute local stress
-        τxx_I, τyy_I, τxy_I, τII_I, ηvep_I, λ_I, ΔPψ_I = compute_local_stress(εij, τij_o, ηij, Pij, λij, λ_relaxation, rheology, ratio, dt)
-        # update arrays
-        stokes.τ.xx[I...], stokes.τ.yy[I...], stokes.τ.xy_c[I...] = τxx_I, τyy_I, τxy_I
-        stokes.τ.II[I...]            = τII_I
-        stokes.viscosity.η_vep[I...] = ηvep_I
-        stokes.λ[I...]               = λ_I
-        stokes.ΔPψ[I...]             = ΔPψ_I
-    end
+        if all(I .≤ ni)
+            τij_o = stokes.τ_o.xx[I...], stokes.τ_o.yy[I...], stokes.τ_o.xy_c[I...]
+            εij   = stokes.ε.xx[I...], stokes.ε.yy[I...], av(stokes.ε.xy)
+            λij   = stokes.λ[I...]
+            ηij   = stokes.viscosity.η[I...]
+            Pij   = stokes.P[I...]
+            ratio = phase_ratios_center[I...]
+            
+            # compute local stress
+            τxx_I, τyy_I, τxy_I, τII_I, ηvep_I, λ_I, ΔPψ_I = compute_local_stress(εij, τij_o, ηij, Pij, λij, λ_relaxation, rheology, ratio, dt)
+            # update arrays
+            stokes.τ.xx[I...], stokes.τ.yy[I...], stokes.τ.xy_c[I...] = τxx_I, τyy_I, τxy_I
+            stokes.τ.II[I...]            = τII_I
+            stokes.viscosity.η_vep[I...] = ηvep_I
+            stokes.λ[I...]               = λ_I
+            stokes.ΔPψ[I...]             = ΔPψ_I
+        end
+    # end
 
     return nothing
 end
@@ -59,8 +60,8 @@ end
 
             else
                 # get rheological properties for this phase
-                G  = fn_ratio(get_shear_modulus, rheology, phase)
-                Kb = fn_ratio(get_bulk_modulus, rheology, phase)
+                G  = get_shear_modulus(rheology, phase)
+                Kb = get_bulk_modulus(rheology, phase)
[…]

[…]
