### 2025-11-10 00ceaf5
replace airthm. mean by harmonic when average viscosity at vertices
