### 2025-11-10 4096820
DYREL now matching Thibault's implementation
