+++ dev/assets/man_equations_basic.md.C4rN1uyG.lean.js
@@ -1,0 +1,1 @@
+import{_ as s,c as Q,o as l,j as t,a,aA as o}from"./chunks/framework.UXI2lzaD.js";const e2=JSON.parse('{"title":"Governing equations","description":"","frontmatter":{},"headers":[],"relativePath":"man/equations_basic.md","filePath":"man/equations_basic.md","lastUpdated":null}'),n={name:"man/equations_basic.md"},i={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},T={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.566ex"},xmlns:"http://www.w3.org/2000/svg",width:"15.116ex",height:"2.262ex",role:"img",focusable:"false",viewBox:"0 -750 6681.4 1000","aria-hidden":"true"},r={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},m={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-1.809ex"},xmlns:"http://www.w3.org/2000/svg",width:"20.596ex",height:"4.749ex",role:"img",focusable:"false",viewBox:"0 -1299.5 9103.4 2099","aria-hidden":"true"},d={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},p={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.029ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.38ex",height:"1.034ex",role:"img",focusable:"false",viewBox:"0 -444 610 457","aria-hidden":"true"},h={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},g={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.439ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.138ex",height:"1.439ex",role:"img",focusable:"false",viewBox:"0 -442 503 636","aria-hidden":"true"},x={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},w={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.464ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.244ex",height:"2.059ex",role:"img",focusable:"false",viewBox:"0 -705 550 910","aria-hidden":"true"},u={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},c={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.018ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.283ex",height:"1.043ex",role:"img",focusable:"false",viewBox:"0 -453 567 461","aria-hidden":"true"},f={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},H={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.439ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.281ex",height:"2.034ex",role:"img",focusable:"false",viewBox:"0 -705 566 899","aria-hidden":"true"},v={class:"MathJax",jax:"SVG",style:{direction:"ltr",position
[…]
