### 2025-11-10 34f763b
add DYREL type
