### 2025-11-10 4e3076c
add plastic strain rate calculation
