+++ DYREL/stress_kernels.jl
@@ -21,7 +21,8 @@ end
         Pij   = av_clamped(stokes.P, Ic...)
         ratio = phase_ratios_vertex[I...]
         # # compute local stress
-        τxx_I, τyy_I, τxy_I, _, _, λ_I, _ = compute_local_stress(εij, τij_o, ηij, Pij, λvij, λ_relaxation, rheology, ratio, dt)
+        τxx_I, τyy_I, τxy_I, _, _, _, _, λ_I, = compute_local_stress(εij, τij_o, ηij, Pij, λvij, λ_relaxation, rheology, ratio, dt)
+
         # update arrays
         stokes.τ.xx_v[I...], stokes.τ.yy_v[I...], stokes.τ.xy[I...] = τxx_I, τyy_I, τxy_I
         stokes.λv[I...]   = λ_I
@@ -36,9 +37,10 @@ end
             ratio = phase_ratios_center[I...]
             
             # compute local stress
-            τxx_I, τyy_I, τxy_I, τII_I, ηvep_I, λ_I, ΔPψ_I = compute_local_stress(εij, τij_o, ηij, Pij, λij, λ_relaxation, rheology, ratio, dt)
+            τxx_I, τyy_I, τxy_I, εxx_pl, εyy_pl, εxy_pl, τII_I, λ_I, ΔPψ_I, ηvep_I = compute_local_stress(εij, τij_o, ηij, Pij, λij, λ_relaxation, rheology, ratio, dt)
             # update arrays
-            stokes.τ.xx[I...], stokes.τ.yy[I...], stokes.τ.xy_c[I...] = τxx_I, τyy_I, τxy_I
+            stokes.τ.xx[I...],    stokes.τ.yy[I...],   stokes.τ.xy_c[I...]     = τxx_I, τyy_I, τxy_I
+            stokes.ε_pl.xx[I...], stokes.ε_pl.yy[I...], stokes.ε_pl.xy_c[I...] = εxx_pl, εyy_pl, εxy_pl
             stokes.τ.II[I...]            = τII_I
             stokes.viscosity.η_vep[I...] = ηvep_I
             stokes.λ[I...]               = λ_I
@@ -49,7 +51,7 @@ end
     return nothing
 end
 
-@generated function compute_local_stress(εij, τij_o, η, P, λ, λ_relaxation, rheology, phase_ratio::SVector{N}, dt) where N
+@generated function compute_local_stress(εij, τij_o, η, P, λ, λ_relaxation, rheology, phase_ratio::SVector{N}, dt) where {N}
     quote
         @inline
         # iterate over phases
@@ -57,9 +59,10 @@ end
             # get phase ratio
             ratio_I = phase_ratio[phase]
             v = if iszero(ratio_I) # this phase does not contribute
-                0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0
+                empty_stress_solution(εij)
 
             else
+
                 # get rheological properties for this phase
                 G  = get_shear_modulus(rheology, phase)
                 Kb = get_bulk_modulus(rheology, phase)
@@ -70,7 +73,7 @@ end
         end
         # sum contributions from all phases
         v = reduce(.+, v_phases)
-        return v # this returns τxx, τyy, τxy, τII, ηvep, λ, ΔPψ
+        return v # this returns (τ_ij...), (εij_pl...), τII, λ, ΔPψ, ηvep
     end
 end
 
@@ -78,14 +81,15 @@ function _compute_local_stress(εij, τij_o, η, P, G, Kb, λ, λ_relaxation, is
 
[…]
