### 2025-11-10 d54d9c8
incorporate DYREL into JR machinery
