+++ src/DYREL/Gershgorin.jl
@@ -1,0 +1,142 @@
+
+function Gershgorin_Stokes2D_SchurComplement!(Dx, Dy, λmaxVx, λmaxVy, η, ηv, γ_eff, phase_ratios, rheology, di, dt)
+    ni = size(η)
+    @parallel (@idx ni) _Gershgorin_Stokes2D_SchurComplement!(Dx, Dy, λmaxVx, λmaxVy, η, ηv, γ_eff, di..., phase_ratios.vertex, phase_ratios.center, rheology, dt)
+    return nothing
+end
+
+@parallel_indices (i, j) function _Gershgorin_Stokes2D_SchurComplement!(Dx, Dy, λmaxVx, λmaxVy, η, ηv, γ_eff, dx, dy,
+    phase_vertex, phase_center, rheology, dt
+)
+    # @inbounds begin
+        phase = phase_vertex[i+1, j+1]
+        GN = fn_ratio(get_shear_modulus, rheology, phase)
+        phase = phase_vertex[i+1, j]
+        GS = fn_ratio(get_shear_modulus, rheology, phase)
+        phase = phase_center[i, j]
+        GW = fn_ratio(get_shear_modulus, rheology, phase)
+
+        # viscosity coefficients at surrounding points
+        ηN = ηv[i+1, j+1]
+        ηS = ηv[i+1, j]
+        ηW = η[i, j]
+        # # bulk viscosity coefficients at surrounding points
+        γW = γ_eff[i, j]
+        
+        if all( (i,j) .≤ size(Dx) )
+            phase = phase_center[i+1, j]
+            GE = fn_ratio(get_shear_modulus, rheology, phase)
+            ηE = η[i+1, j]
+            γE = γ_eff[i+1, j]
+            # effective viscoelastic viscosity 
+            ηN = 1 / (1 / ηN + 1 / (GN * dt))
+            ηS = 1 / (1 / ηS + 1 / (GS * dt))
+            ηW = 1 / (1 / ηW + 1 / (GW * dt))
+            ηE = 1 / (1 / ηE + 1 / (GE * dt))
+            # compute Gershgorin entries
+            Cxx = abs(ηN / dy ^ 2) + abs(ηS / dy ^ 2) + abs(γE / dx ^ 2 + (4 / 3) * ηE / dx ^ 2) + abs(γW / dx ^ 2 + (4 / 3) * ηW / dx ^ 2) + abs(-(-ηN / dy - ηS / dy) / dy + (γE / dx + γW / dx) / dx + ((4 / 3) * ηE / dx + (4 / 3) * ηW / dx) / dx)
+            Cxy = abs(γE / (dx * dy) - 2 / 3 * ηE / (dx * dy) + ηN / (dx * dy)) + abs(γE / (dx * dy) - 2 / 3 * ηE / (dx * dy) + ηS / (dx * dy)) + abs(γW / (dx * dy) + ηN / (dx * dy) - 2 / 3 * ηW / (dx * dy)) + abs(γW / (dx * dy) + ηS / (dx * dy) - 2 / 3 * ηW / (dx * dy))
+            # this is the preconditioner diagonal entry
+            Dx_ij = Dx[i, j] = -(-ηN / dy - ηS / dy) / dy + (γE / dx + γW / dx) / dx + ((4 / 3) * ηE / dx + (4 / 3) * ηW / dx) / dx
+            # maximum eigenvalue estimate
+            λmaxVx[i, j] = inv(Dx_ij) * (Cxx + Cxy)
+        end
+
+        
[…]

[…]
