### 2025-11-10 bd683fc
update DYREL miniapp
