### 2025-11-10 f01a01e
working CUDA version
