### 2025-11-10 a88b284
CUDA miniapp
