+++ previews/PR380/assets/man_equations_discretization.md.D-Q9w3LB.js
@@ -1,0 +1,1 @@
+import{_ as e,a as d}from"./chunks/stokes_stag2D.fEnPgM0u.js";import{_ as n,c as a,o as T,aA as m,j as Q,a as l}from"./chunks/framework.CTMZkvum.js";const v=JSON.parse('{"title":"Spatial discretization of the governing equations","description":"","frontmatter":{},"headers":[],"relativePath":"man/equations_discretization.md","filePath":"man/equations_discretization.md","lastUpdated":null}'),s={name:"man/equations_discretization.md"},o={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},r={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-8.57ex"},xmlns:"http://www.w3.org/2000/svg",width:"50.308ex",height:"18.272ex",role:"img",focusable:"false",viewBox:"0 -4288 22236 8076","aria-hidden":"true"},i={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},h={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-4.247ex"},xmlns:"http://www.w3.org/2000/svg",width:"13.568ex",height:"9.626ex",role:"img",focusable:"false",viewBox:"0 -2377.3 5997.2 4254.6","aria-hidden":"true"},g={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},H={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-2.194ex"},xmlns:"http://www.w3.org/2000/svg",width:"36.815ex",height:"5.52ex",role:"img",focusable:"false",viewBox:"0 -1469.9 16272.2 2439.7","aria-hidden":"true"},p={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},c={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-2.349ex"},xmlns:"http://www.w3.org/2000/svg",width:"38.904ex",height:"5.83ex",role:"img",focusable:"false",viewBox:"0 -1538.4 17195.7 2576.9","aria-hidden":"true"},u={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},L={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-2.827ex"},xmlns:"http://www.w3.org/2000/svg",width:"82.925ex",height:"6.785ex",role:"img",focusable:"false",viewBox:"0 -1749.5 36652.6 2999","aria-hidden":"true"},w={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},f={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-2.293ex"},xmlns:"http://www.w3.org/2000/svg",width:"45.565ex",height:"5.717ex",role:
[…]
