+++ previews/PR380/assets/man_authors.md.BI1qDbp0.lean.js
@@ -1,0 +1,1 @@
+import{_ as r,c as a,o as t,aA as o}from"./chunks/framework.CTMZkvum.js";const d=JSON.parse('{"title":"Authors","description":"","frontmatter":{},"headers":[],"relativePath":"man/authors.md","filePath":"man/authors.md","lastUpdated":null}'),i={name:"man/authors.md"};function l(s,e,n,u,p,h){return t(),a("div",null,[...e[0]||(e[0]=[o("",7)])])}const m=r(i,[["render",l]]);export{d as __pageData,m as default};

[…]
