### 2025-11-13 338a13f
add harmonic av to visc in VS
