+++ previews/PR380/assets/man_restart.md.BgEvuUKk.js
@@ -1,4 +1,4 @@
-import{_ as s,c as a,o as t,aA as e}from"./chunks/framework.CTMZkvum.js";const c=JSON.parse('{"title":"Restarting from a Checkpoint File","description":"","frontmatter":{},"headers":[],"relativePath":"man/restart.md","filePath":"man/restart.md","lastUpdated":null}'),n={name:"man/restart.md"};function l(h,i,p,k,d,r){return t(),a("div",null,[...i[0]||(i[0]=[e(`<h1 id="Restarting-from-a-Checkpoint-File" tabindex="-1">Restarting from a Checkpoint File <a class="header-anchor" href="#Restarting-from-a-Checkpoint-File" aria-label="Permalink to &quot;Restarting from a Checkpoint File {#Restarting-from-a-Checkpoint-File}&quot;">​</a></h1><p>To restart a simulation from a previously saved checkpoint file, you can make use of the checkpointing functions described in the <a href="./checkpointing">Checkpointing documentation</a>. Depending on the file format you used to save your checkpoint (HDF5 or JLD2), you can load the saved state of your simulation using the corresponding loading function.</p><p>In this example, we will demonstrate how to set up a script to restart a simulation from a JLD2 checkpoint file as we can save the entire structures of the <code>StokesArrays</code> and <code>ThermalArrays</code> which makes it easier to restart the simulation. We will assume that you have already saved a checkpoint file using the <code>checkpointing_jld2</code> function for the example of a 2D subduction model. Ideally, one does not need to change much in the initial script used to start the simulation from scratch. The main difference is that instead of initializing the <code>StokesArrays</code> and <code>ThermalArrays</code> from scratch, we will load them from the checkpoint file. For a detailed description of the 2D subduction model setup, please refer to the <a href="./subduction2D/subduction2D">2D subduction documentation</a>. The following example can be found <a href="https://github.com/PTsolvers/JustRelax.jl/blob/d63ca8f08860859700913b575c9befc33d5c4f2a/miniapps/subduction/2D/Subduction2D_restart" target="_blank" rel="noreferrer">here</a>.</p><p>Load JustRelax necessary modules and define backend.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> CUDA </span><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># comment this out if you are not using CUDA; or load AMDGPU.jl if you are using an AMD GPU</span></span>
[…]
