+++ dev/assets/man_advection.md.A5jd5YHK.js
@@ -1,0 +1,1 @@
+import{_ as a,c as t,o as i,aA as l}from"./chunks/framework.UXI2lzaD.js";const p=JSON.parse('{"title":"Field advection","description":"","frontmatter":{},"headers":[],"relativePath":"man/advection.md","filePath":"man/advection.md","lastUpdated":null}'),r={name:"man/advection.md"};function n(o,e,c,s,d,_){return i(),t("div",null,[...e[0]||(e[0]=[l('<h1 id="Field-advection" tabindex="-1">Field advection <a class="header-anchor" href="#Field-advection" aria-label="Permalink to &quot;Field advection {#Field-advection}&quot;">​</a></h1><h2 id="particles-in-cell" tabindex="-1">Particles-in-Cell <a class="header-anchor" href="#particles-in-cell" aria-label="Permalink to &quot;Particles-in-Cell&quot;">​</a></h2><p><a href="https://github.com/PTsolvers/JustRelax.jl" target="_blank" rel="noreferrer">JustRelax.jl</a> relies on <a href="https://github.com/JuliaGeodynamics/JustPIC.jl" target="_blank" rel="noreferrer">JustPIC.jl</a> for advections of particles containing material information.</p>',3)])])}const h=a(r,[["render",n]]);export{p as __pageData,h as default};
