+++ dev/assets/man_plume3D_rheology.md.HLk1omkE.js
@@ -1,0 +1,68 @@
+import{_ as i,c as a,o as h,aA as k}from"./chunks/framework.UXI2lzaD.js";const g=JSON.parse('{"title":"Using GeoParams.jl to define the rheology of the material phases","description":"","frontmatter":{},"headers":[],"relativePath":"man/plume3D/rheology.md","filePath":"man/plume3D/rheology.md","lastUpdated":null}'),n={name:"man/plume3D/rheology.md"};function l(t,s,p,e,E,r){return h(),a("div",null,[...s[0]||(s[0]=[k(`<h1 id="Using-GeoParams.jl-to-define-the-rheology-of-the-material-phases" tabindex="-1">Using GeoParams.jl to define the rheology of the material phases <a class="header-anchor" href="#Using-GeoParams.jl-to-define-the-rheology-of-the-material-phases" aria-label="Permalink to &quot;Using GeoParams.jl to define the rheology of the material phases {#Using-GeoParams.jl-to-define-the-rheology-of-the-material-phases}&quot;">​</a></h1><p>In this example, we will define a rheology for a 3D plume model using the <code>GeoParams.jl</code> package. The model will consist of for four different rheology layers: upper crust, lower crust, lithospheric mantle, and sublithospheric mantle. We will use a non-Newtoninan visco-elasto-plastic rheology.</p><p>We will start by defining the physical parameters of the diffusion and dislocation creep laws of every phase:</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> GeoParams</span></span>
[…]

[…]
