+++ dev/assets/chunks/@localSearchIndexroot.DafLMf_L.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":151,"nextId":151,"documentIds":{"0":"/JustRelax.jl/dev/#What-is-JustRelax.jl?","1":"/JustRelax.jl/dev/#How-to-Install-JustRelax.jl?","2":"/JustRelax.jl/dev/#funding","3":"/JustRelax.jl/dev/man/Blankenbach#Blankenbach-benchmark","4":"/JustRelax.jl/dev/man/Blankenbach#Initialize-packages","5":"/JustRelax.jl/dev/man/Blankenbach#script","6":"/JustRelax.jl/dev/man/Blankenbach#Model-domain","7":"/JustRelax.jl/dev/man/Blankenbach#rheology","8":"/JustRelax.jl/dev/man/Blankenbach#Initialize-particles","9":"/JustRelax.jl/dev/man/Blankenbach#Temperature-anomaly","10":"/JustRelax.jl/dev/man/Blankenbach#Stokes-and-heat-diffusion-arrays","11":"/JustRelax.jl/dev/man/Blankenbach#Initialize-thermal-profile-and-viscosity-fields","12":"/JustRelax.jl/dev/man/Blankenbach#Boundary-conditions","13":"/JustRelax.jl/dev/man/Blankenbach#Pseuo-transient-coefficients","14":"/JustRelax.jl/dev/man/Blankenbach#Just-before-solving-the-problem...","15":"/JustRelax.jl/dev/man/Blankenbach#Advancing-one-time-step","16":"/JustRelax.jl/dev/man/Blankenbach#visualization","17":"/JustRelax.jl/dev/man/Blankenbach#fields","18":"/JustRelax.jl/dev/man/Blankenbach#Final-model","19":"/JustRelax.jl/dev/man/GeoParams#List-of-GeoParams-functions","20":"/JustRelax.jl/dev/man/JustPIC#List-of-JustPIC-functions","21":"/JustRelax.jl/dev/man/ShearBands#ShearBand-benchmark","22":"/JustRelax.jl/dev/man/ShearBands#Initialize-packages","23":"/JustRelax.jl/dev/man/ShearBands#script","24":"/JustRelax.jl/dev/man/ShearBands#Model-domain","25":"/JustRelax.jl/dev/man/ShearBands#Physical-properties-using-GeoParams","26":"/JustRelax.jl/dev/man/ShearBands#rheology","27":"/JustRelax.jl/dev/man/ShearBands#Phase-anomaly","28":"/JustRelax.jl/dev/man/ShearBands#Stokes-arrays","29":"/JustRelax.jl/dev/man/ShearBands#Initialize-viscosity-fields","30":"/JustRelax.jl/dev/man/ShearBands#Boundary-conditions","31":"/JustRelax.jl/dev/man/ShearBands#Pseuo-transient-coefficients","32":"/JustRelax.jl/dev/man/ShearBands#Just-before-solving-the-problem...","33":"/JustRelax.jl/dev/man/ShearBands#Advancing-one-time-step","34":"/JustRelax.jl/dev/man/ShearBands#visualization","35":"/JustRelax.jl/dev/man/ShearBands#fields","36":"/JustRelax.jl/dev/man/ShearBands#Final-model","37":"/JustRelax.jl/dev/man/advection#Field-advection","38":"/JustRelax.jl/dev/man/advection#particles-in-cell","39":"/JustRelax.jl/dev/man/authors#authors","40":"/JustRelax.jl/dev/man/authors#Principal-Developers","41":"/JustRelax.jl/dev/man/authors#contributors","42":"/JustRelax.jl/dev/man/backend#Selecting-the-backend","43":"/JustRelax.jl/dev/man/boundary_cond
[…]
