### 2025-11-24 d41ec51
add some exports
