### 2025-11-24 9d23389
update DYREL examples
