### 2025-11-25 PTsolvers/JustRelax.jl#381 @shipengcheng1230
JOSS Review Feedback/Suggestions
Hi team,

I'd like to use this issue as a communication portal as part of the https://github.com/openjournals/joss-reviews/issues/9365. First of all, I thoroughly enjoy your package of its versatile functionality and rich examples.

I have so far had the following suggestions in the submitted paper:
- line 29 as well as Figure 1, many packages mentioned here have their own references. Please check their GitHub and the DOI.
- line 54: in the definition of parallel efficiency, based on the numbers in Figure 2, would it be more precise to say *the wall-time of the single-process run normalized by the product of the number of processes and the wall-time of the parallel run*?
- line 61: [GeoParams.jl](https://github.com/JuliaGeodynamics/GeoParams.jl) need a reference.
- line 66, [ParallelStencil.jl](https://github.com/omlins/ParallelStencil.jl), following others in the paper a link would be preferred.
- line 76-77, the links to the examples and document do not work. It appears they point to a deprecated location.

Also for the package:
- When experimenting some of the examples, for instance, the [Shear Bands](https://ptsolvers.github.io/JustRelax.jl/dev/man/ShearBands), it has some discrepancies compared to the one [here](https://github.com/PTsolvers/JustRelax.jl/blob/main/miniapps/benchmarks/stokes2D/shear_band/ShearBand2D.jl). Following the one in the documentation does not work, for me I encountered issue in [this section](https://ptsolvers.github.io/JustRelax.jl/dev/man/ShearBands#Phase-anomaly) where `PhaseRatios` multidispatch found no matching method, and `backend` was not defined. I understand many APIs are under imminent change because of the rapid development, building these docs from real Julia script could be CI/CD time consuming, it'd be beneficial at least to include your working examples in the [miniapps](https://github.com/PTsolvers/JustRelax.jl/tree/main/miniapps).
