+++ docs/paper/paper.md
@@ -61,7 +61,7 @@ Simulating Earth's thermo-mechanical evolution requires solving coupled, nonline
 
 `JustRelax.jl` addresses this gap by providing a compact, modular, GPU‑ready toolkit implemented in Julia. It implements a matrix‑free Accelerated Pseudo‑Transient (APT) solver that reduces global linear‑algebra bottlenecks and ports efficiently to accelerators, while outsourcing some physical calculations, parallalelization, and IO to other packages, shown in \ref{fig:dependencies}. This keeps the core solver as contained as possible and makes the software more flexible. Julia’s high‑level syntax and package manager simplify installation, scripting and extension, lowering the barrier for students and researchers to prototype and test new physics.
 
-![Core dependencies of `JustRelax.jl`. \label{fig:dependencies}](figs/dependencies.png){width=75%}
+![Sketch of the main dependencies of `JustRelax.jl`: a) physics packages: material properties and rheology ([GeoParams.jl](https://github.com/JuliaGeodynamics/GeoParams.jl) [@GeoParams]), Particles-in-Cell based advection ([JustPIC.jl](https://github.com/JuliaGeodynamics/JustPIC.jl) [@JustPIC], where data structures are handled by [CellArrays.jl](https://github.com/omlins/CellArrays.jl)); b) parallelization: domain decomposition and distributed parallelism is handled by [ImplicitGlobalGrid.jl](https://github.com/eth-cscs/ImplicitGlobalGrid.jl) [@Omlin2024] and [MPI.jl](https://github.com/JuliaParallel/MPI.jl) [Byrne2021], and backend abstraction is implemented with [ParallelStencil.jl](https://github.com/omlins/ParallelStencil.jl) [@Omlin2024], which supports Julia's native multi-threading ([Threads.jl](https://docs.julialang.org/en/v1/manual/multi-threading/)), and NVIDIA ([CUDA.jl](https://github.com/JuliaGPU/CUDA.jl) [besard2018juliagpu]) and [AMDGPU.jl](https://github.com/JuliaGPU/AMDGPU.jl) GPU cards; and c) I/O and data visualization: HDF5 binary files ([HDF5.jl](https://github.com/JuliaIO/HDF5.jl) and [JLD2.jl](https://github.com/JuliaIO/JLD2.jl)), and VTK files for visualization ([WriteVTK.jl](https://github.com/JuliaVTK/WriteVTK.jl) [@WriteVTK]). \label{fig:dependencies}](figs/dependencies.png){width=75%}
 
 In short, `JustRelax.jl` delivers a high‑performance, portable alternative to legacy geodynamics codes: it (i) exploits modern GPUs without extensive rewrites; (ii) avoids costly matrix assembly via matrix‑free algorithms; and (iii) promotes modularity, reproducibility, and rapid development. Together with CI, scalable I/O and checkpointing, these features make `JustRelax.jl` practical for both exploratory studies and production‑scale simulations on modern multi-XPU platforms.
 
@@ -73,13 +73,13 @@ In short, `JustRelax.jl` delivers a high‑performance, portable alternative to 
 
 `JustRelax.jl` features:
 
[…]
