+++ docs/paper/paper.bib
@@ -10,6 +10,51 @@
   URL = {https://doi.org/10.1137/141000671}
 }
 
+@article{GeoParams,
+  author = {Kaus, Boris and de Montserrat, Albert and Medinger, Nils and Aellig, Pascal and Dominguez, Hugo and Riel, Nicolas and Cosarinsky, Matias and Spang, Arne and Berlie, Nicolas and Kiss, Daniel and Ranocha, Hendrik and Fuchs, Lukas and Frasunkiewicz, Jacob and Räss, Ludovic and Lueder, Mona and Seiler, Alexander and Duretz, Thibault},
+  title = {GeoParams.jl: v0.7.4},
+  year = {2025},
+  URL = {https://doi.org/10.5281/zenodo.15719680}
+}
+
+@article{JustPIC,
+  author = {de Montserrat, Albert and Aellig, Pascal and Räss, Ludovic and Duretz, Thibault and Utkin, Ivan and Kaus, Boris and Ranocha, Hendrik},
+  title = {JustPIC.jl: v0.5.10},
+  year = {2025},
+  URL = {https://doi.org/10.5281/zenodo.15704602}
+}
+
+@article{WriteVTK,
+  author = {Polanco, J. I.},
+  title = {WriteVTK.jl: a Julia package for writing VTK XML files (v1.18.0)},
+  year = {2023},
+  URL = {https://doi.org/10.5281/zenodo.7804590}
+}
+@article{Byrne2021, 
+  doi = {10.21105/jcon.00068}, 
+  url = {https://doi.org/10.21105/jcon.00068}, 
+  year = {2021}, 
+  publisher = {The Open Journal}, 
+  volume = {1}, 
+  number = {1}, 
+  pages = {68}, 
+  author = {Simon Byrne and Lucas C. Wilcox and Valentin Churavy}, 
+  title = {MPI.jl: Julia bindings for the Message Passing Interface}, 
+  journal = {Proceedings of the JuliaCon Conferences} 
+} 
+
+@article{besard2018juliagpu,
+  author        = {Besard, Tim and Foket, Christophe and De Sutter, Bjorn},
+  title         = {Effective Extensible Programming: Unleashing {Julia} on {GPUs}},
+  journal       = {IEEE Transactions on Parallel and Distributed Systems},
+  year          = {2018},
+  doi           = {10.1109/TPDS.2018.2872064},
+  ISSN          = {1045-9219},
+  archivePrefix = {arXiv},
+  eprint        = {1712.03112},
+  primaryClass  = {cs.PL},
+}
+
 @article{larionov2017variational,
   author = {Larionov, Egor and Batty, Christopher and Bridson, Robert}, 
   title = {Variational stokes: a unified pressure-viscosity solver for accurate viscous liquids}, 
