### 2025-11-25 303b1f1
add Hugo's suggestion
