### 2025-11-25 e79e643
update shear band docs
