+++ docs/src/man/ShearBands.md
@@ -10,6 +10,12 @@ using JustRelax, JustRelax.JustRelax2D, JustRelax.DataIO
 const backend_JR = CPUBackend
 ```
 
+```julia
+using JustPIC, JustPIC._2D
+import JustPIC._2D.GridGeometryUtils as GGU
+const backend_JP = JustPIC.CPUBackend
+```
+
 We will also use [ParallelStencil.jl](https://github.com/omlins/ParallelStencil.jl) to write some device-agnostic helper functions:
 ```julia
 using ParallelStencil
@@ -23,6 +29,7 @@ using GeoParams
 ## Script
 
 ### Model domain
+
 ```julia
 nx = ny      = 64                       # number of cells per dimension
 igg          = IGG(
@@ -36,7 +43,6 @@ di           = @. li / ni               # grid step in x- and -y
 origin       = 0.0, 0.0                 # origin coordinates
 grid         = Geometry(ni, li; origin = origin)
 (; xci, xvi) = grid                     # nodes at the center and vertices of the cells
-dt           = Inf
 ```
 
 ### Physical properties using GeoParams
@@ -46,71 +52,79 @@ dt           = Inf
 C       = τ_y           # Cohesion
 η0      = 1.0           # viscosity
 G0      = 1.0           # elastic shear modulus
-Gi      = G0/(6.0-4.0)  # elastic shear modulus perturbation
+Gi      = G0 / (6 - 4)  # elastic shear modulus perturbation
 εbg     = 1.0           # background strain-rate
 η_reg   = 8e-3          # regularisation "viscosity"
-dt      = η0/G0/4.0     # assumes Maxwell time of 4
-el_bg   = ConstantElasticity(; G=G0, Kb=4)
-el_inc  = ConstantElasticity(; G=Gi, Kb=4)
-visc    = LinearViscous(; η=η0)
+dt      = η0 / G0 / 4.0 # assumes Maxwell time of 4
+el_bg   = ConstantElasticity(; G = G0, Kb = 4.0)
+el_inc  = ConstantElasticity(; G = Gi, Kb = 4.0)
+visc    = LinearViscous(; η = η0)
 pl      = DruckerPrager_regularised(;  # non-regularized plasticity
-    C    = C,
+    C    = C / cosd(ϕ),
     ϕ    = ϕ,
     η_vp = η_reg,
     Ψ    = 0
 )
 ```
-### Rheology
-```julia
-    rheology = (
-        # Low density phase
-        SetMaterialParams(;
-            Phase             = 1,
-            Density           = ConstantDensity(; ρ = 0.0),
-            Gravity           = ConstantGravity(; g = 0.0),
-            CompositeRheology = CompositeRheology((visc, el_bg, pl)),
-            Elasticity        = el_bg,
 
-        ),
-        # High density phase
-        SetMaterialParams(;
-            Density           = ConstantDensity(; ρ = 0.0),
-            Gravity           = ConstantGravity(; g = 0.0),
-            CompositeRheology = CompositeRheology((visc, el_inc, pl)),
-            Elasticity        = el_inc,
-        ),
-    )
+### Rheology
+
+```julia
+rheology = (
+    # Low density phase
+    SetMaterialParams(;
+        Phase             = 1,
+        Density           = ConstantDensity(; ρ = 0.0),
+        Gravity           = ConstantGravity(; g = 0.0),
+        CompositeRheology = CompositeRheology((visc, el_bg, pl)),
[…]
