### 2025-11-25 2dafa71
add viscosity update
