+++ DYREL/ShearBand2D_AutoTune_Thibault.jl
@@ -1,0 +1,240 @@
+# using CUDA
+using GeoParams, CellArrays
+using GLMakie
+# using Plots
+
+using JustRelax, JustRelax.JustRelax2D
+using ParallelStencil
+# @init_parallel_stencil(CUDA, Float64, 2)
+@init_parallel_stencil(Threads, Float64, 2)
+
+const backend = CPUBackend
+# const backend = CUDABackend
+
+using JustPIC, JustPIC._2D
+import JustPIC._2D.GridGeometryUtils as GGU
+
+const backend_JP = JustPIC.CPUBackend
+# const backend_JP = CUDABackend
+
+# Initialize phases on the particles
+function init_phases!(phase_ratios, xci, xvi, circle)
+    ni = size(phase_ratios.center)
+
+    @parallel_indices (i, j) function init_phases!(phases, xc, yc, circle)
+        x, y = xc[i], yc[j]
+        p    = GGU.Point(x, y)
+
+        if GGU.inside(p, circle) # inclusion
+            @index phases[1, i, j] = 0.0
+            @index phases[2, i, j] = 1.0
+
+        else
+            @index phases[1, i, j] = 1.0
+            @index phases[2, i, j] = 0.0
+
+        end
+        return nothing
+    end
+
+    @parallel (@idx ni) init_phases!(phase_ratios.center, xci..., circle)
+    @parallel (@idx ni .+ 1) init_phases!(phase_ratios.vertex, xvi..., circle)
+
+    return nothing
+end
+
+
+using Statistics
+# include("DYREL_solver.jl")
+
+# MAIN SCRIPT --------------------------------------------------------------------
+function main(igg; nx = 64, ny = 64, figdir = "model_figs")
+
+    # Physical domain ------------------------------------
+    ly = 1e0          # domain length in y
+    lx = 2e0           # domain length in x
+    ni = nx, ny       # number of cells
+    li = lx, ly       # domain length in x- and y-
+    di = @. li / ni   # grid step in x- and -y
+    origin = -lx/2, -ly/2     # origin coordinates
+    grid = Geometry(ni, li; origin = origin)
+    (; xci, xvi) = grid # nodes at the center and vertices of the cells
+
+    # Physical properties using GeoParams ----------------
+    τ_y      = 1.6           # yield stress. If do_DP=true, τ_y stand for the cohesion: c*cos(ϕ)
+    ϕ        = 35            # friction angle
+    C        = 5             # Cohesion
+    η0       = 4e3           # viscosity
+    εbg      = -1.0          # background strain-rate
+    η_reg    = 5e-3          # regularisation "viscosity"
+    dt       = 1.5e9         # assumes Maxwell time of 4
+    el_bg    = ConstantElasticity(; G = 1e-10, Kb = 1e-10)
+    el_inc   = ConstantElasticity(; G = 1e-10, Kb = 1e-10)
+    visc     = LinearViscous(; η = η0)
+    visc_inc = LinearViscous(; η = 4e-2)
+    pl = DruckerPrager_regularised(;
+        # non-regularized plasticity
+        C = C * 1,
+        ϕ = ϕ,
+        η_vp = η_reg,
+        Ψ = 5e0,
+    )
+
+    rheology = (
+        # Host
+        SetMaterialParams(;
+            Phase = 1,
+            Density = ConstantDensity(; ρ = 0.0),
+            Gravity = ConstantGravity(; g = 0.0),
[…]
