+++ miniapps/Project.toml
@@ -1,0 +1,6 @@
+[deps]
+CellArrays = "d35fcfd7-7af4-4c67-b1aa-d78070614af4"
+GeoParams = "e018b62d-d9de-4a26-8697-af89c310ae38"
+JustPIC = "10dc771f-8528-4cd9-9d3b-b21b2e693339"
+JustRelax = "34418575-392d-4e26-8c6d-96b0910afa06"
+ParallelStencil = "94395366-693c-11ea-3b26-d9b7aac5d958"
