+++ miniapps/Project.toml
@@ -1,7 +1,9 @@
 [deps]
 CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 CellArrays = "d35fcfd7-7af4-4c67-b1aa-d78070614af4"
+ExactFieldSolutions = "2a6b1ac7-2de5-4354-bffb-bf009f6a4bef"
 GeoParams = "e018b62d-d9de-4a26-8697-af89c310ae38"
+GeophysicalModelGenerator = "3700c31b-fa53-48a6-808a-ef22d5a84742"
 JustPIC = "10dc771f-8528-4cd9-9d3b-b21b2e693339"
 JustRelax = "34418575-392d-4e26-8c6d-96b0910afa06"
 MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
