### 2025-11-27 799cdb1
Add compat
