+++ miniapps/Project.toml
@@ -8,3 +8,10 @@ JustPIC = "10dc771f-8528-4cd9-9d3b-b21b2e693339"
 JustRelax = "34418575-392d-4e26-8c6d-96b0910afa06"
 MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
 ParallelStencil = "94395366-693c-11ea-3b26-d9b7aac5d958"
+
+[compat]
+CellArrays = "0.3.2"
+ExactFieldSolutions = "0.1.6"
+GeoParams = "0.7.8"
+MPI = "0.20"
+ParallelStencil = "0.14.2"
