### 2025-11-27 a5dd183
Ludovic editing suggestion
