+++ src/variational_stokes/StressKernels.jl
@@ -22,6 +22,7 @@
         phase_vertex,
         ϕ::JustRelax.RockRatio,
     ) where {T}
+    εijv_pl = ε_pl[3]
     τxyv = τshear_v[1]
     τxyv_old = τshear_ov[1]
     ni = size(Pr)
@@ -75,9 +76,11 @@
             dQdτxy = 0.5 * (τxyv[I...] + dτxyv) / τIIv_ij
             εij_plv = λv[I...] * dQdτxy
             τxyv[I...] += @muladd dτxyv - 2.0 * ηv_ij * εij_plv * dτ_rv
+            ε_pl[3][I...] = εij_plv
         else
             # stress correction @ vertex
             τxyv[I...] += dτxyv
+            ε_pl[3][I...] = 0.0
         end
     else
         τxyv[I...] = zero(eltype(T))
@@ -125,14 +128,18 @@
                 τij = dτij .+ τij
                 Base.@nexprs 3 i -> begin
                     @inbounds τ[i][I...] = τij[i]
+                    if i < 3
                     @inbounds ε_pl[i][I...] = εij_pl[i]
+                    end
                 end
                 τII_ij = second_invariant(τij)
             else
                 # stress correction @ center
                 Base.@nexprs 3 i -> begin
-                    @inbounds τ[i][I...] = dτij[i] + τij[i]
+                    τ[i][I...] = dτij[i] + τij[i]
+                    if i < 3
                     @inbounds ε_pl[i][I...] = 0.0
+                    end
                 end
                 τII_ij
             end
@@ -145,6 +152,7 @@
             η_vep[I...] = zero(eltype(T))
             Base.@nexprs 3 i -> begin
                 τ[i][I...] = zero(eltype(T))
+                ε_pl[i][I...] = zero(eltype(T))
             end
 
         end
@@ -199,6 +207,8 @@ end
         εxzv_ij = av_clamped_yz_y(ε[5], Ic...)
         εxyv_ij = av_clamped_yz_z(ε[6], Ic...)
 
+        ε_plyzv_ij = ε_pl[4][I...]
+
         τxxv_ij = av_clamped_yz(τ[1], Ic...)
         τyyv_ij = av_clamped_yz(τ[2], Ic...)
         τzzv_ij = av_clamped_yz(τ[3], Ic...)
@@ -250,10 +260,13 @@ end
                 relλ * (max(Fv, 0.0) / (ηv_ij * dτ_rv + η_regv + volumev))
 
             dQdτyz = 0.5 * (τyzv_ij + dτyzv) / τIIv_ij
-            τyzv[I...] += dτyzv - 2.0 * ηv_ij * λv[1][I...] * dQdτyz * dτ_rv
+            ε_plxyv_ij = λv[1][I...] * dQdτyz
+            τyzv[I...] += dτyzv - 2.0 * ηv_ij * ε_plxyv_ij * dτ_rv
+            ε_pl[4][I...] = ε_plxyv_ij
         else
             # stress correction @ vertex
             τyzv[I...] += dτyzv
+            ε_pl[4][I...] = 0.0
         end
     else
         τyzv[I...] = zero(eltype(T))
@@ -283,6 +296,7 @@ end
         τyzv_old_ij = av_clamped_xz_x(τyzv_old, Ic...)
         τxzv_old_ij = τxzv_old[I...]
         τxyv_old_ij = av_clamped_xz_z(τxyv_old, Ic...)
+        ε_plxzv_ij = ε_pl[5][I...]
 
         # vertex parameters
         phase = @inbounds phase_xz[I...]
@@ -321,10 +335,13 @@ end
                 relλ * (max(Fv, 0.0) / (ηv_ij * dτ_rv + η_regv + volumev))
 
[…]
