### 2025-12-01 c0f6fa5
fix v1.10 precompilation error?
