### 2025-12-01 e0b59b1
update DYREL solver & miniapps
