+++ miniapps/subduction/2D/Subduction2D_DYREL_ND.jl
@@ -1,0 +1,390 @@
+# Load script dependencies
+using GeoParams, GLMakie
+
+const isCUDA = false
+
+@static if isCUDA
+    using CUDA
+end
+
+using JustRelax, JustRelax.JustRelax2D, JustRelax.DataIO
+
+const backend = @static if isCUDA
+    CUDABackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+else
+    JustRelax.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+end
+
+using ParallelStencil, ParallelStencil.FiniteDifferences2D
+
+@static if isCUDA
+    @init_parallel_stencil(CUDA, Float64, 2)
+else
+    @init_parallel_stencil(Threads, Float64, 2)
+end
+
+using JustPIC, JustPIC._2D
+# Threads is the default backend,
+# to run on a CUDA GPU load CUDA.jl (i.e. "using CUDA") at the beginning of the script,
+# and to run on an AMD GPU load AMDGPU.jl (i.e. "using AMDGPU") at the beginning of the script.
+const backend_JP = @static if isCUDA
+    CUDABackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+else
+    JustPIC.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+end
+
+# Load file with all the rheology configurations
+include("Subduction2D_setup.jl")
+include("Subduction2D_rheology_ND.jl")
+
+## SET OF HELPER FUNCTIONS PARTICULAR FOR THIS SCRIPT --------------------------------
+
+import ParallelStencil.INDICES
+const idx_k = INDICES[2]
+macro all_k(A)
+    return esc(:($A[$idx_k]))
+end
+
+function copyinn_x!(A, B)
+    @parallel function f_x(A, B)
+        @all(A) = @inn_x(B)
+        return nothing
+    end
+
+    return @parallel f_x(A, B)
+end
+
+# Initial pressure profile - not accurate
+@parallel function init_P!(P, ρg, z)
+    @all(P) = abs(@all(ρg) * @all_k(z)) * <(@all_k(z), 0.0)
+    return nothing
+end
+## END OF HELPER FUNCTION ------------------------------------------------------------
+
+## BEGIN OF MAIN SCRIPT --------------------------------------------------------------
+function main(li, origin, T_GMG, phases_GMG, igg; nx = 16, ny = 16, figdir = "figs2D", do_vtk = false)
+
+    thickness = 120e3 * m
+    η0        = 1.0e22 * Pa * s
+    CharDim   = GEO_units(;
+        length = thickness, viscosity = η0, temperature = 1.0e3K
+    )
+    li_nd     = nondimensionalize(li .* m, CharDim)
+    origin_nd = nondimensionalize(origin .* m, CharDim)
+
+    # Physical domain ------------------------------------
+    ni    = nx, ny           # number of cells
+    di    = @. li_nd / ni       # grid steps
+    grid  = Geometry(ni, li_nd; origin = origin_nd)
+    (; xci, xvi) = grid # nodes at the center and vertices of the cells
+    # ----------------------------------------------------
+
+    # Physical properties using GeoParams ----------------
+    # rheology = init_rheology_nonNewtonian_plastic(CharDim)
+    rheology = init_rheology_linear(CharDim)
+    dt       = nondimensionalize(10e3 * yr, CharDim)
+    # ----------------------------------------------------
+
+    # Initialize particles -------------------------------
+    nxcell    = 40
+    max_xcell = 60
+    min_xcell = 20
+    particles = init_particles(
+        backend_JP, nxcell, max_xcell, min_xcell, xvi...
+    )
+    subgrid_arrays = SubgridDiffusionCellArrays(particles)
[…]
