### 2025-12-01 2259462
update models figure & caption
