### 2025-12-01 61eadb2
update parallel efficiency definition
