### 2025-12-01 3ae386e
update DYREL solver
