### 2025-12-01 2820ebc
update subduction miniapp
