+++ miniapps/subduction/2D/Subduction2D_DYREL.jl
@@ -278,7 +278,7 @@ function main(li, origin, phases_GMG, igg; nx = 16, ny = 16, figdir = "figs2D", 
         t += dt
 
         # Data I/O and plotting ---------------------
-        if it == 1 || rem(it, 10) == 0
+        if it == 1 || rem(it, 5) == 0
             checkpointing_jld2(checkpoint, stokes, thermal, t, dt; it = it)
             checkpointing_particles(checkpoint, particles; phases = pPhases, phase_ratios = phase_ratios, particle_args = particle_args, particle_args_reduced = particle_args_reduced, t = t, dt = dt, it = it)
             (; η_vep, η) = stokes.viscosity
@@ -357,7 +357,7 @@ end
 ## END OF MAIN SCRIPT ----------------------------------------------------------------
 do_vtk = true # set to true to generate VTK files for ParaView
 figdir = "Subduction2D"
-n = 64
+n = 128 ÷ 2
 # n = 32 # *4
 nx, ny = n * 2, n
 
