### 2025-12-01 22c959d
update miniapp & solver
