### 2025-12-01 7765e1e
recompute dyrel arrays per time step
