+++ .github/workflows/ci.yml
@@ -55,7 +55,7 @@ jobs:
             allow_failure: true
 
     steps:
-      - uses: actions/checkout@v5
+      - uses: actions/checkout@v6
       - uses: julia-actions/setup-julia@v2
         with:
           version: ${{ matrix.version }}
