### 2025-12-02 feccada
Apply suggestions from code review
Co-authored-by: Albert de Montserrat <58044444+albert-de-montserrat@users.noreply.github.com>
