+++ src/variational_stokes/StressKernels.jl
@@ -261,7 +261,7 @@ end
 
             dQdτyz = 0.5 * (τyzv_ij + dτyzv) / τIIv_ij
             ε_plxyv_ij = λv[1][I...] * dQdτyz
-            τyzv[I...] += dτyzv - 2.0 * ηv_ij * ε_plxyv_ij * dτ_rv
+            τyzv[I...] += @muladd dτyzv - 2.0 * ηv_ij * ε_plxyv_ij * dτ_rv
             ε_pl[4][I...] = ε_plxyv_ij
         else
             # stress correction @ vertex
@@ -336,7 +336,7 @@ end
 
             dQdτxz = 0.5 * (τxzv_ij + dτxzv) / τIIv_ij
             ε_plxzv_ij = λv[2][I...] * dQdτxz
-            τxzv[I...] += dτxzv - 2.0 * ηv_ij * ε_plxzv_ij * dτ_rv
+            τxzv[I...] += @muladd dτxzv - 2.0 * ηv_ij * ε_plxzv_ij * dτ_rv
             ε_pl[5][I...] = ε_plxzv_ij
         else
             # stress correction @ vertex
@@ -412,7 +412,7 @@ end
 
             dQdτxy = 0.5 * (τxyv_ij + dτxyv) / τIIv_ij
             ε_plxyv_ij = λv[3][I...] * dQdτxy
-            τxyv[I...] += dτxyv - 2.0 * ηv_ij * ε_plxyv_ij * dτ_rv
+            τxyv[I...] += @muladd dτxyv - 2.0 * ηv_ij * ε_plxyv_ij * dτ_rv
             ε_pl[6][I...] = ε_plxyv_ij
         else
             # stress correction @ vertex
@@ -576,7 +576,7 @@ end
                 relλ * (max(Fv, 0.0) / (ηv_ij * dτ_rv * dt + η_regv + volumev))
             dQdτxy = 0.5 * (τxyv[I...] + dτxyv) / τIIv_ij
             εij_pl = λv[I...] * dQdτxy
-            τxyv[I...] += dτxyv - 2.0 * ηv_ij * dt * εij_pl * dτ_rv
+            τxyv[I...] += @muladd dτxyv - 2.0 * ηv_ij * dt * εij_pl * dτ_rv
             ε_pl[3][I...] = εij_pl
         else
             # stress correction @ vertex
