### 2025-12-02 b17625d
update miniapps
