+++ src/stokes/PressureKernels.jl
@@ -95,7 +95,8 @@ end
         ::Nothing,
     ) where {N, C <: JustRelax.CellArray}
     K = fn_ratio(get_bulk_modulus, rheology, @cell(phase_ratio[I...]))
-    @inbounds RP[I...], P[I...] = _compute_P!(P[I...], P0[I...], ∇V[I...], Q[I...], η[I...], K, dt, r, θ_dτ)
+    Gdt = fn_ratio(get_shear_modulus, rheology, @cell(phase_ratio[I...])) * dt
+    @inbounds RP[I...], P[I...] = _compute_P!(P[I...], P0[I...], ∇V[I...], Q[I...], η[I...], K, Gdt, dt, r, θ_dτ)
     return nothing
 end
 
@@ -115,7 +116,8 @@ end
         melt_fraction,
     ) where {N, C <: JustRelax.CellArray}
     K = fn_ratio(get_bulk_modulus, rheology, @cell(phase_ratio[I...]))
-    @inbounds RP[I...], P[I...] = _compute_P!(P[I...], P0[I...], ∇V[I...], Q[I...], η[I...], K, dt, r, θ_dτ)
+    Gdt = fn_ratio(get_shear_modulus, rheology, @cell(phase_ratio[I...])) * dt
+    @inbounds RP[I...], P[I...] = _compute_P!(P[I...], P0[I...], ∇V[I...], Q[I...], η[I...], K, Gdt, dt, r, θ_dτ)
     return nothing
 end
 
@@ -136,9 +138,10 @@ end
     ) where {N, C <: JustRelax.CellArray}
     @inbounds phase_ratio_I = phase_ratio[I...]
     @inbounds K = fn_ratio(get_bulk_modulus, rheology, phase_ratio_I)
+    @inbounds Gdt = fn_ratio(get_shear_modulus, rheology, phase_ratio_I) * dt
     @inbounds α = fn_ratio(get_thermal_expansion, rheology, phase_ratio_I)
     @inbounds RP[I...], P[I...] = _compute_P!(
-        P[I...], P0[I...], ∇V[I...], Q[I...], ΔTc[I...], α, η[I...], K, dt, r, θ_dτ
+        P[I...], P0[I...], ∇V[I...], Q[I...], ΔTc[I...], α, η[I...], K, Gdt, dt, r, θ_dτ
     )
     return nothing
 end
@@ -159,9 +162,10 @@ end
         melt_fraction,
     ) where {N, C <: JustRelax.CellArray}
     @inbounds K = fn_ratio(get_bulk_modulus, rheology, @cell(phase_ratio[I...]))
+    @inbounds Gdt = fn_ratio(get_shear_modulus, rheology, @cell(phase_ratio[I...])) * dt
     @inbounds α = fn_ratio(get_thermal_expansion, rheology, @cell(phase_ratio[I...]), (; ϕ = melt_fraction[I...]))
     @inbounds RP[I...], P[I...] = _compute_P!(
-        P[I...], P0[I...], ∇V[I...], Q[I...], ΔTc[I...], α, η[I...], K, dt, r, θ_dτ
+        P[I...], P0[I...], ∇V[I...], Q[I...], ΔTc[I...], α, η[I...], K, Gdt, dt, r, θ_dτ
     )
     return nothing
 end
@@ -174,21 +178,21 @@ function _compute_P!(P, ∇V, Q, η, dt, r, θ_dτ)
     return RP, P
 end
 
[…]
