### 2025-12-02 418bbd8
pass G and not Gdt
