+++ src/stokes/PressureKernels.jl
@@ -7,21 +7,21 @@
 end
 
 ## Compressible
-@parallel_indices (I...) function compute_P!(P, P0, RP, ∇V, Q, η, K, Gdt, dt, r, θ_dτ)
+@parallel_indices (I...) function compute_P!(P, P0, RP, ∇V, Q, η, K, G, dt, r, θ_dτ)
     @inbounds RP[I...], P[I...] = _compute_P!(
-        P[I...], P0[I...], ∇V[I...], Q[I...], η[I...], K[I...], Gdt[I...], dt, r, θ_dτ
+        P[I...], P0[I...], ∇V[I...], Q[I...], η[I...], K[I...], G[I...], dt, r, θ_dτ
     )
     return nothing
 end
 
-# With GeoParams
+  # With GeoParams
 
 @parallel_indices (I...) function compute_P!(
         P, P0, RP, ∇V, Q, η, rheology::NTuple{N, MaterialParams}, phase, dt, r, θ_dτ, args
     ) where {N}
     K = get_bulk_modulus(rheology, phase[I...])
-    Gdt = get_shear_modulus(rheology, phase[I...]) * dt
-    @inbounds RP[I...], P[I...] = _compute_P!(P[I...], P0[I...], ∇V[I...], Q[I...], η[I...], K, Gdt, dt, r, θ_dτ)
+    G = get_shear_modulus(rheology, phase[I...])
+    @inbounds RP[I...], P[I...] = _compute_P!(P[I...], P0[I...], ∇V[I...], Q[I...], η[I...], K, G, dt, r, θ_dτ)
     return nothing
 end
 
@@ -96,8 +96,8 @@ end
         ::Nothing,
     ) where {N, C <: JustRelax.CellArray}
     K = fn_ratio(get_bulk_modulus, rheology, @cell(phase_ratio[I...]))
-    Gdt = fn_ratio(get_shear_modulus, rheology, @cell(phase_ratio[I...])) * dt
-    @inbounds RP[I...], P[I...] = _compute_P!(P[I...], P0[I...], ∇V[I...], Q[I...], η[I...], K, Gdt, dt, r, θ_dτ)
+    G = fn_ratio(get_shear_modulus, rheology, @cell(phase_ratio[I...]))
+    @inbounds RP[I...], P[I...] = _compute_P!(P[I...], P0[I...], ∇V[I...], Q[I...], η[I...], K, G, dt, r, θ_dτ)
     return nothing
 end
 
@@ -117,8 +117,8 @@ end
         melt_fraction,
     ) where {N, C <: JustRelax.CellArray}
     K = fn_ratio(get_bulk_modulus, rheology, @cell(phase_ratio[I...]))
-    Gdt = fn_ratio(get_shear_modulus, rheology, @cell(phase_ratio[I...])) * dt
-    @inbounds RP[I...], P[I...] = _compute_P!(P[I...], P0[I...], ∇V[I...], Q[I...], η[I...], K, Gdt, dt, r, θ_dτ)
+    G = fn_ratio(get_shear_modulus, rheology, @cell(phase_ratio[I...]))
+    @inbounds RP[I...], P[I...] = _compute_P!(P[I...], P0[I...], ∇V[I...], Q[I...], η[I...], K, G, dt, r, θ_dτ)
     return nothing
 end
 
@@ -139,10 +139,10 @@ end
     ) where {N, C <: JustRelax.CellArray}
     @inbounds phase_ratio_I = phase_ratio[I...]
[…]
