+++ src/stokes/Stokes3D.jl
@@ -81,7 +81,7 @@ function _solve!(
                 stokes.Q,
                 η,
                 K,
-                G*dt,
+                G,
                 dt,
                 pt_stokes.r,
                 pt_stokes.θ_dτ,
