### 2025-12-04 1b45828
remove tensile cap & add a couple of improvements
