### 2025-12-05 409769e
update rift miniapps
