+++ Rift2D_Albert/Rift2D_Albert/Rift2D.jl
@@ -1,0 +1,565 @@
+# const isCUDA = false
+const isCUDA = true
+
+@static if isCUDA
+    using CUDA
+end
+
+using JustRelax, JustRelax.JustRelax2D, JustRelax.DataIO
+import JustRelax.@cell
+
+const backend_JR = @static if isCUDA
+    CUDABackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+else
+    JustRelax.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+end
+
+using ParallelStencil, ParallelStencil.FiniteDifferences2D
+
+@static if isCUDA
+    @init_parallel_stencil(CUDA, Float64, 2)
+else
+    @init_parallel_stencil(Threads, Float64, 2)
+end
+
+using JustPIC, JustPIC._2D
+# Threads is the default backend_JR,
+# to run on a CUDA GPU load CUDA.jl (i.e. "using CUDA") at the beginning of the script,
+# and to run on an AMD GPU load AMDGPU.jl (i.e. "using AMDGPU") at the beginning of the script.
+const backend_JP = @static if isCUDA
+    CUDABackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+else
+    JustPIC.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+end
+
+# Load script dependencies
+using GeoParams, GLMakie
+using JLD2
+
+# Load file with all the rheology configurations
+include("Rift2DSetupFaultInclusion.jl")
+include("RiftRheology.jl")
+
+## SET OF HELPER FUNCTIONS PARTICULAR FOR THIS SCRIPT --------------------------------
+
+import ParallelStencil.INDICES
+const idx_k = INDICES[2]
+macro all_k(A)
+    esc(:($A[$idx_k]))
+end
+
+function copyinn_x!(A, B)
+    @parallel function f_x(A, B)
+        @all(A) = @inn_x(B)
+        return nothing
+    end
+
+    @parallel f_x(A, B)
+end
+# Initial pressure profile - not accurate
+@parallel function init_P!(P, ρg, z)
+    @all(P) = abs(@all(ρg) * @all_k(z)) #* <(@all_k(z), 0.0)
+    return nothing
+end
+
+
+# Initial thermal profile
+function thermal_anomaly!(T, xvi, anomaly_center, anomaly_radius, ΔT)
+
+    @parallel_indices (i, j) function f_anomaly!(T, xci, anomaly_center, anomaly_radius, ΔT)
+        r = sqrt(( xci[1][i] - anomaly_center[1])^2 + ( xci[2][j] - anomaly_center[2])^2)
+        if r ≤ anomaly_radius
+            T[i+1, j] +=  ΔT
+        end
+        return nothing
+    end
+    ni = length.(xvi)
+    @parallel (@idx ni) f_anomaly!(T, xvi, anomaly_center, anomaly_radius, ΔT)
+
+    return nothing
+end
+## END OF HELPER FUNCTION ------------------------------------------------------------
+
+## BEGIN OF MAIN SCRIPT --------------------------------------------------------------
+function main(li, origin, T_GMG, phases_GMG, igg; nx=16, ny=16, figdir="figs2D", do_vtk =false)
+
+    thickness = li[1] * m
+    η0        = 1.0e22 * Pa * s
+    CharDim   = GEO_units(;
+        length = thickness, viscosity = η0, temperature = maximum(T_GMG)*K
+    )
+    # Physical domain ------------------------------------
+    li_nd        = nondimensionalize(li .* m, CharDim)
+    origin_nd    = nondimensionalize(origin .* m, CharDim)
[…]
