+++ Rift2D_Albert/Rift2D_Albert/RiftRheology.jl
@@ -1,0 +1,321 @@
+function init_rheologies(CharDim; is_plastic = true)
+
+    # Dislocation and Diffusion creep
+    # disl_upper_crust = DislocationCreep(
+    #     A = 5.07e-18Pa^(-23 // 10) / s, # units are Pa^(-n) / s
+    #     n = 2.3NoUnits,
+    #     E = 154.0e3J / mol,
+    #     V = 6.0e-6m^3 / mol,
+    #     r = 0.0NoUnits,
+    #     R = 8.3145J / mol / K
+    # )
+    # disl_lower_crust = DislocationCreep(
+    #     A = 2.08e-23Pa^(-32 // 10) / s, # units are Pa^(-n) / s
+    #     n = 3.2NoUnits,
+    #     E = 238.0e3J / mol,
+    #     V = 6.0e-6m^3 / mol,
+    #     r = 0.0NoUnits,
+    #     R = 8.3145J / mol / K,
+    # )
+    # disl_lithospheric_mantle = DislocationCreep(
+    #     A = 2.51e-17Pa^(-35 // 10) / s, # units are Pa^(-n) / s
+    #     n = 3.5NoUnits,
+    #     E = 530.0e3J / mol,
+    #     V = 6.0e-6m^3 / mol,
+    #     r = 0.0NoUnits,
+    #     R = 8.3145J / mol / K
+    # )
+    # disl_sublithospheric_mantle = DislocationCreep(
+    #     A = 2.51e-17Pa^(-35 // 10) / s, # units are Pa^(-n) / s
+    #     n = 3.5NoUnits,
+    #     E = 530.0e3J / mol,
+    #     V = 6.0e-6m^3 / mol,
+    #     r = 0.0NoUnits,
+    #     R = 8.3145J / mol / K
+    # )
+    # diff_lithospheric_mantle = DiffusionCreep(
+    #     A = 2.51e-17Pa^(-1) / s, # units are Pa^(-n - r) / s * m^(-p)
+    #     n = 1.0NoUnits,
+    #     E = 530.0e3J / mol,
+    #     V = 6.0e-6m^3 / mol,
+    #     r = 0.0NoUnits,
+    #     p = 0.0NoUnits,
+    #     R = 8.3145J / mol / K
+    # )
+    # diff_sublithospheric_mantle = DiffusionCreep(
+    #     A = 2.51e-17Pa^(-1) / s, # units are Pa^(-n - r) / s * m^(-p)
+    #     n = 1.0NoUnits,
+    #     E = 530.0e3J / mol,
+    #     V = 6.0e-6m^3 / mol,
+    #     r = 0.0NoUnits,
+    #     p = 0.0NoUnits,
+    #     R = 8.3145J / mol / K
+    # )
+
+    # disl_upper_crust            = SetDislocationCreep(GeoParams.Dislocation.wet_quartzite_Hirth_2001)
+    # disl_lower_crust            = SetDislocationCreep(GeoParams.Dislocation.plagioclase_An75_Ji_1993)
+    # disl_lithospheric_mantle    = SetDislocationCreep(GeoParams.Dislocation.wet_olivine_Hirth_2003)
+    # disl_sublithospheric_mantle = SetDislocationCreep(GeoParams.Dislocation.dry_olivine_Hirth_2003)
+    # diff_lithospheric_mantle    = SetDiffusionCreep(GeoParams.Diffusion.dry_olivine_Hirth_2003)
+    # diff_sublithospheric_mantle = SetDiffusionCreep(GeoParams.Diffusion.wet_olivine_Hirth_2003)
+    
+    # disl_upper_crust = DislocationCreep(
[…]
