### 2025-12-05 895911f
update DYREL solver
