+++ src/DYREL/DYREL2D.jl
@@ -22,11 +22,13 @@ function _solve_DYREL!(
         igg::IGG;
         viscosity_cutoff = (-Inf, Inf),
         viscosity_relaxation = 1.0e-2,
-        λ_relaxation = 0.2,
+        λ_relaxation_DR = 1,
+        λ_relaxation_PH = 1,
         iterMax = 50.0e3,
         iterMin = 1.0e2,
         free_surface = false,
         nout = 100,
+        rel_drop = 1e-2,
         b_width = (4, 4, 0),
         verbose = true,
         kwargs...,
@@ -60,13 +62,13 @@ function _solve_DYREL!(
     γfact = 20
 
     # errors
-    err = 1.0
-    iter = 0
+    err      = 1.0
+    iter     = 0
     err_evo1 = Float64[]
     err_evo2 = Float64[]
-    norm_Rx = Float64[]
-    norm_Ry = Float64[]
-    norm_∇V = Float64[]
+    norm_Rx  = Float64[]
+    norm_Ry  = Float64[]
+    norm_∇V  = Float64[]
 
     # solver loop
     @copy stokes.P0 stokes.P
@@ -81,16 +83,13 @@ function _solve_DYREL!(
     stokes.λ  .= 0.0
     stokes.λv .= 0.0
 
-    # compute buoyancy forces and viscosity
-    compute_ρg!(ρg, phase_ratios, rheology, args)
-    # viscosity guess based on strain rate
-    compute_viscosity!(stokes, phase_ratios, args, rheology, viscosity_cutoff)
-    center2vertex!(stokes.viscosity.ηv, stokes.viscosity.η)
-    
-    # recompute all the DYREL variables
-    DYREL!(dyrel, stokes, rheology, phase_ratios, di, dt)
+    # # compute buoyancy forces and viscosity
+    # compute_ρg!(ρg, phase_ratios, rheology, args)
+    # # viscosity guess based on strain rate
+    # compute_viscosity!(stokes, phase_ratios, args, rheology, viscosity_cutoff)
+    # center2vertex!(stokes.viscosity.ηv, stokes.viscosity.η)
 
-    rel_drop   = 1e-2 # relative drop of velocity residual per PH iteration
+    # rel_drop   = 1e-2 # relative drop of velocity residual per PH iteration
     # Iteration loop
     errVx0     = 1.0
     errVy0     = 1.0
@@ -108,6 +107,10 @@ function _solve_DYREL!(
     # Powell-Hestenes iterations
     for itPH in 1:250
 
+        # reset plastic multiplier at the beginning of the time step
+        stokes.λ  .= 0.0
+        stokes.λv .= 0.0
+        
         # compute divergence
         @parallel (@idx ni) compute_∇V!(stokes.∇V, @velocity(stokes), _di)
 
@@ -117,7 +120,7 @@ function _solve_DYREL!(
         )
 
         # compute deviatoric stress
-        compute_stress_DRYEL!(stokes, rheology, phase_ratios, λ_relaxation, dt) # not resetting λ in every PH iteration seems to work better
+        compute_stress_DRYEL!(stokes, rheology, phase_ratios, λ_relaxation_PH, dt) # not resetting λ in every PH iteration seems to work better
         # compute_stress_DRYEL!(stokes, rheology, phase_ratios, 1, dt) # λ_relaxation = 1 to reset λ
         # update_halo!(stokes.τ.xy)
 
@@ -168,6 +171,7 @@ function _solve_DYREL!(
         while (err > ϵ_vel && itPT ≤ iterMax)
             itPT   += 1
             itg    += 1
+            iter   += 1 
 
             # Pseudo-old dudes 
             copyto!(Rx0, stokes.R.Rx)
[…]
