### 2025-12-05 bf08d2e
last updates of the week
