### 2025-12-08 0968e50
use MQS instead of bilinear interp in miniapps
