+++ miniapps/benchmarks/stokes2D/shear_band/ShearApparatus2D_nonDim.jl
@@ -1,0 +1,279 @@
+using GeoParams, GLMakie, CellArrays
+using JustRelax, JustRelax.JustRelax2D
+using ParallelStencil
+@init_parallel_stencil(Threads, Float64, 2)
+
+const backend = CPUBackend
+
+using JustPIC, JustPIC._2D
+import JustPIC._2D.GridGeometryUtils as GGU
+
+const backend_JP = JustPIC.CPUBackend
+
+# HELPER FUNCTIONS ---------------------------------------------------------------
+function init_phases!(pPhases, particles, xvi, width, height, fault_thickness, buffer, fault_dip)
+    ni      = size(particles.index)
+    lx      = xvi[1][end] - xvi[1][1]
+
+    fault   = GGU.Rectangle((0e0, height / 2), Inf, fault_thickness; θ = deg2rad(fault_dip))
+    sample  = GGU.Rectangle((0e0, height / 2), width, height)
+    buffer1 = GGU.Rectangle((xvi[1][1]   + buffer/2, height / 2), buffer, height)
+    buffer2 = GGU.Rectangle((xvi[1][end] - buffer/2, height / 2), buffer, height)
+
+    @parallel_indices (i, j) function init_phases!(pPhases, px, py, index, sample, fault, buffer1, buffer2)
+        
+        for ip in cellaxes(index)
+            
+            x        = @index px[ip, i, j]
+            y        = @index py[ip, i, j]
+            p        = GGU.Point(x, y)
+            infault  = GGU.inside(p, fault)
+            insample = GGU.inside(p, sample)
+            inbuffer = GGU.inside(p, buffer1) || GGU.inside(p, buffer2)
+
+            if insample
+                @index pPhases[ip, i, j] = 1e0
+            end
+            if infault
+                @index pPhases[ip, i, j] = 2e0
+            end
+            if inbuffer
+                @index pPhases[ip, i, j] = 3e0
+            end
+
+        end
+        return nothing
+    end
+
+    @parallel (@idx ni) init_phases!(pPhases, particles.coords..., particles.index, sample, fault, buffer1, buffer2)
+
+    return nothing
+end
+
+# MAIN SCRIPT --------------------------------------------------------------------
+function main(igg; nx = 64, ny = 64, figdir = "model_figs")
+
+    # domain
+    height          = 3e0
+    width           = height / 3
+    buffer          = 0.5
+    fault_thickness = 0.2
+    fault_dip       = 30
+    # Physical domain ------------------------------------
+    ly               = height              # domain length in y
+    lx               = width + 2 * buffer  # domain length in x
+    li               = lx, ly       # domain length in x- and y-
+    # aspect_ratio     = ceil(Int, ly / lx)
+    # aspect_ratio     = 1
+    # ny               = nx * aspect_ratio # number of cells
+    ni               = nx, ny       # number of cells
+    di               = @. li / ni   # grid step in x- and -y
+    origin           = -lx / 2, 0.0 # origin coordinates
+    grid             = Geometry(ni, li; origin = origin)
+    (; xci, xvi)     = grid # nodes at the center and vertices of the cells
+
+    # Physical properties using GeoParams ----------------
+    τ_y         = 1.6               # yield stress. If do_DP=true, τ_y stand for the cohesion: c*cos(ϕ)
+    ϕ           = 30                # friction angle
+    C           = τ_y               # Cohesion
[…]
