+++ miniapps/benchmarks/stokes2D/shear_band/ShearApparatus2D.jl
@@ -1,0 +1,258 @@
+using GeoParams, GLMakie, CellArrays
+using JustRelax, JustRelax.JustRelax2D
+using ParallelStencil
+@init_parallel_stencil(Threads, Float64, 2)
+
+const backend = CPUBackend
+
+using JustPIC, JustPIC._2D
+import JustPIC._2D.GridGeometryUtils as GGU
+
+const backend_JP = JustPIC.CPUBackend
+
+# Initialize phases on the particles
+function init_phases!(phase_ratios, xci, xvi, width, height, fault_thickness, fault_dip)
+    ni      = size(phase_ratios.center)
+    lx      = xvi[1][end] - xvi[1][1]
+
+    fault   = GGU.Rectangle((0e0, height / 2), Inf, fault_thickness; θ = deg2rad(fault_dip))
+    sample  = GGU.Rectangle((0e0, height / 2), width, height)
+    buffer1 = GGU.Rectangle((xvi[1][1]   + buffer/2, height / 2), buffer, height)
+    buffer2 = GGU.Rectangle((xvi[1][end] - buffer/2, height / 2), buffer, height)
+
+    @parallel_indices (i, j) function init_phases!(phases, xc, yc, sample, fault, buffer1, buffer2)
+        x, y     = xc[i], yc[j]
+        p        = GGU.Point(x, y)
+        infault  = GGU.inside(p, fault)
+        insample = GGU.inside(p, sample)
+        inbuffer = GGU.inside(p, buffer1) || GGU.inside(p, buffer2)
+        
+        if insample
+            @index phases[1, i, j] = 1.0
+            @index phases[2, i, j] = 0.0
+            @index phases[3, i, j] = 0.0
+        end
+        if infault
+            @index phases[1, i, j] = 0.0
+            @index phases[2, i, j] = 1.0
+            @index phases[3, i, j] = 0.0
+        end
+        if inbuffer
+            @index phases[1, i, j] = 0.0
+            @index phases[2, i, j] = 0.0
+            @index phases[3, i, j] = 1.0
+        end
+        
+        
+        return nothing
+    end
+
+    @parallel (@idx ni) init_phases!(phase_ratios.center, xci..., sample, fault, buffer1, buffer2)
+    @parallel (@idx ni .+ 1) init_phases!(phase_ratios.vertex, xvi..., sample, fault, buffer1, buffer2)
+    return nothing
+end
+
+ # domain
+ height          = 3e0
+    width           = height / 3
+    buffer          = 0.2
+    fault_thickness = 0.1
+    fault_dip       = 45
+
+
+# # Physical domain ------------------------------------
+# lx = width + 2 * buffer  # domain length in x
+# ly = height              # domain length in y
+# ni = nx, ny              # number of cells
+# li = lx, ly              # domain length in x- and y-
+# di = @. li / ni          # grid step in x- and -y
+# origin = -lx / 2, 0.0    # origin coordinates
+# grid = Geometry(ni, li; origin = origin)
+# (; xci, xvi) = grid # nodes at the center and vertices of the cells
+
+# # Initialize phase ratios -------------------------------
+# phase_ratios = PhaseRatios(backend_JP, length(rheology), ni)
+# init_phases!(phase_ratios, xci, xvi, width, height, fault_thickness, fault_dip)
+
+# p = [argmax(p) for p in phase_ratios.center]
+# display(heatmap(p))
+
+# MAIN SCRIPT --------------------------------------------------------------------
+function main(igg; nx = 64, ny = 64, figdir = "model_figs")
+
+    # domain
+    height          = 3e0
+    width           = height / 3
+    buffer          = 0.1
+    fault_thickness = 0.1
+    fault_dip       = 45
+
[…]
