### 2025-12-08 5517da1
remove aparatus miniapps
