### 2025-12-08 PTsolvers/JustRelax.jl#394 @gassmoeller
Include license information in source files
Part of https://github.com/openjournals/joss-reviews/issues/9365

JustRelax is published under MIT license, but this is only clear because of the main License file in the root folder. For several reasons it can be beneficial to add a license header to all source files, in particular if you expect users to copy and separate individual files from the repo (e.g. scientists using examples to start their own research model). The most modern version of this is to follow the REUSE software guidance and use SPDX compliant license headers: https://reuse.software/faq/#quick-summary

In short it can be as simple as adding the following 3 lines to the start of all source files:
```
# SPDX-FileCopyrightText: 2025 The JustRelax authors
#
# SPDX-License-Identifier: MIT
```

Then there is no ambiguity over the license even for downstream users (say students receiving a modified file from another student or professor), and the license is both human readable and machine readable (e.g. for automatic license compliance checks).

This is not a hard requirement for publication in JOSS, but I would recommend it, since it can be quickly done with a small script and may save you trouble in the future.
