### 2025-12-08 PTsolvers/JustRelax.jl#395 @gassmoeller
Tests use packages that are not installed as part of JustRelax
Related to https://github.com/openjournals/joss-reviews/issues/9365

The test `test_shearband2D.jl` uses the package `Suppressor` which at least for me was not installed when I installed `JustRelax` using `Pkg`. Is this an oversight or intended behavior? In other words do I need to manually install additional packages to run the test examples?

The other question I had when running into this: I did execute the test instructions from the github repo page:
```
using JustRelax

julia> ]

(@v1.xx) pkg> test JustRelax
```
and everything finished correctly (including the test_shearband2D). Does that mean the test somehow uses different packages than my environment when I start it manually?

Maybe these are trivial questions, I am just not that familiar with Julia.
