+++ previews/PR393/assets/man_installation.md.O9pO9bSm.lean.js
@@ -1,0 +1,1 @@
+import{_ as a,c as i,o as t,aA as n}from"./chunks/framework.2ZcTfwAw.js";const g=JSON.parse('{"title":"Installation","description":"","frontmatter":{},"headers":[],"relativePath":"man/installation.md","filePath":"man/installation.md","lastUpdated":null}'),l={name:"man/installation.md"};function e(p,s,h,k,d,o){return t(),i("div",null,[...s[0]||(s[0]=[n("",11)])])}const E=a(l,[["render",e]]);export{g as __pageData,E as default};
