+++ previews/PR393/assets/man_contributing.md.ClF-JPk7.js
@@ -1,0 +1,37 @@
+import{_ as s,c as a,o as e,aA as t}from"./chunks/framework.2ZcTfwAw.js";const d=JSON.parse('{"title":"Contributing","description":"","frontmatter":{},"headers":[],"relativePath":"man/contributing.md","filePath":"man/contributing.md","lastUpdated":null}'),i={name:"man/contributing.md"};function p(r,n,o,l,c,h){return e(),a("div",null,[...n[0]||(n[0]=[t(`<h1 id="contributing" tabindex="-1">Contributing <a class="header-anchor" href="#contributing" aria-label="Permalink to &quot;Contributing&quot;">​</a></h1><p><a href="https://github.com/PTsolvers/JustRelax.jl" target="_blank" rel="noreferrer">JustRelax.jl</a> is an open-source project and we are very happy to accept contributions from the community. Please feel free to <a href="https://github.com/PTsolvers/JustRelax.jl/issues/new" target="_blank" rel="noreferrer">open issues</a> or submit patches (preferably as <a href="https://github.com/PTsolvers/JustRelax.jl/pulls" target="_blank" rel="noreferrer">pull requests</a>) any time. For planned larger contributions, it is often beneficial to get in contact with one of the principal developers first (see <a href="/JustRelax.jl/previews/PR393/man/authors#Authors">Authors</a>).</p><p><a href="https://github.com/PTsolvers/JustRelax.jl" target="_blank" rel="noreferrer">JustRelax.jl</a> and its contributions are licensed under the MIT license. As a contributor, you certify that all your contributions are in conformance with the <em>Developer Certificate of Origin (Version 1.1)</em>, which is reproduced below.</p><h2 id="Developer-Certificate-of-Origin-Version-1.1" tabindex="-1">Developer Certificate of Origin (Version 1.1) <a class="header-anchor" href="#Developer-Certificate-of-Origin-Version-1.1" aria-label="Permalink to &quot;Developer Certificate of Origin (Version 1.1) {#Developer-Certificate-of-Origin-Version-1.1}&quot;">​</a></h2><p>The following text was taken from <a href="https://developercertificate.org" target="_blank" rel="noreferrer">https://developercertificate.org</a>:</p><div class="language- vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang"></span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span>Developer Certificate of Origin</span></span>
+<span class="line"><span>Version 1.1</span></span>
+<span class="line"><span></span></span>
+<span class="line"><span>Copyright (C) 2004, 2006 The Linux Foundation and its contributors.</span></span>
+<span class="line"><span>1 Letterman Drive</span></span>
+<span class="line"><span>Suite D4700</span></span>
+<span class="line"><span>San Francisco, CA, 94129</span></span>
+<span class="line"><span></span></span>
+<span class="line"><span>Everyone is permitted to copy and distribute verbatim copies of this</span></span>
+<span class="line"><span>license document, but changing it is not allowed.</span></span>
+<span class="line"><span></span></span>
+<span class="line"><span></span></span>
+<span class="line"><span>Developer&#39;s Certificate of Origin 1.1</span></span>
+<span class="line"><span></span></span>
+<span class="line"><span>By making a contribution to this project, I certify that:</span></span>
+<span class="line"><span></span></span>
+<span class="line"><span>(a) The contribution was created in whole or in part by me and I</span></span>
+<span class="line"><span>    have the right to submit it under the open source license</span></span>
+<span class="line"><span>    indicated in the file; or</span></span>
+<span class="line"><span></span></span>
[…]
