+++ previews/PR374/assets/index.md.D4ZaTm4m.lean.js
@@ -1,0 +1,1 @@
+import{_ as s,c as i,o as e,aA as t}from"./chunks/framework.BqrsUk5L.js";const k=JSON.parse('{"title":"","description":"","frontmatter":{"layout":"home","hero":{"name":"JustRelax.jl","text":"Solving multi-physical geodynamic problems","tagline":"A collection of iterative accelerated pseudo-transient solvers using MPI for distributed computing on GPUs or CPUs.","actions":[{"theme":"brand","text":"Getting Started","link":"/man/installation"},{"theme":"alt","text":"API Reference 📚","link":"/man/listfunctions"},{"theme":"alt","text":"View on GitHub","link":"https://github.com/PTsolvers/JustRelax.jl"}],"image":{"src":"/logo.png","alt":"JustRelax.jl"}},"features":[{"icon":"🚀","title":"Backend Agnostic","details":"Effortlessly execute your code on CPUs and GPUs with ParallelStencils.jl.","link":"/man/backend"},{"icon":"🛠️","title":"Governing equations","details":"Governing equations of the solvers using the accelerated pseudo-transient method","link":"/man/equations_basic"},{"icon":"⚡","title":"Examples","details":"An overview of available examples from shear bands to 2d subduction","link":"man/subduction2D/setup"},{"icon":"🧩","title":"Extensibility","details":"Provides a natural repository for contributions of new solvers for use by the larger community.","link":"/man/contributing"}]},"headers":[],"relativePath":"index.md","filePath":"index.md","lastUpdated":null}'),l={name:"index.md"};function n(r,a,o,h,p,d){return e(),i("div",null,[...a[0]||(a[0]=[t("",14)])])}const c=s(l,[["render",n]]);export{k as __pageData,c as default};
