+++ previews/PR366/assets/man_code_of_conduct.md.BlggnQLv.lean.js
@@ -1,0 +1,1 @@
+import{_ as o,c as a,o as n,aA as t}from"./chunks/framework.DSrZ4BQQ.js";const h=JSON.parse('{"title":"Code of Conduct","description":"","frontmatter":{},"headers":[],"relativePath":"man/code_of_conduct.md","filePath":"man/code_of_conduct.md","lastUpdated":null}'),i={name:"man/code_of_conduct.md"};function r(s,e,c,l,d,p){return n(),a("div",null,[...e[0]||(e[0]=[t("",2)])])}const m=o(i,[["render",r]]);export{h as __pageData,m as default};

[…]
