### 2025-12-09 9c7e11f
update init_particles call
