+++ miniapps/benchmarks/stokes2D/elastic_buildup/Elastic_BuildUp_phases_incompressible.jl
@@ -1,10 +1,28 @@
+const isCUDA = false
+# const isCUDA = true
+
+@static if isCUDA
+    using CUDA
+end
+
 using ParallelStencil
 @init_parallel_stencil(Threads, Float64, 2)
 
 using JustRelax, JustRelax.JustRelax2D, JustRelax.DataIO
-const backend_JR = CPUBackend
+const backend_JR = @static if isCUDA
+    CUDABackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+else
+    JustRelax.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+end
 
-using GeoParams, CairoMakie, CellArrays
+using JustPIC,JustPIC._2D
+const backend = @static if isCUDA
+    JustPIC.CUDABackend
+else
+    JustPIC.CPUBackend
+end
+
+using GeoParams, CairoMakie
 
 # HELPER FUNCTIONS ---------------------------------------------------------------
 solution(ε, t, G, η) = 2 * ε * η * (1 - exp(-G * t / η))
@@ -19,55 +37,61 @@ function init_phases!(phase_ratios)
         return nothing
     end
 
-    return @parallel (@idx ni) init_phases!(phase_ratios.center)
+    @parallel (@idx ni) init_phases!(phase_ratios.center)
+    @parallel (@idx ni.+1) init_phases!(phase_ratios.vertex)
+    return nothing
 end
+
+const yr      = 365.25 * 3600 * 24
+const kyr     = 1.0e3 * yr
 
 # MAIN SCRIPT --------------------------------------------------------------------
 function main(igg; nx = 64, ny = 64, figdir = "model_figs")
 
     # Physical domain ------------------------------------
-    ly = 1.0e0          # domain length in y
-    lx = ly           # domain length in x
-    ni = nx, ny       # number of cells
-    li = lx, ly       # domain length in x- and y-
-    di = @. li / ni   # grid step in x- and -y
-    origin = 0.0, 0.0     # origin coordinates
-    grid = Geometry(ni, li; origin = origin)
+    ly           = 100e3        # domain length in y
+    lx           = ly           # domain length in x
+    ni           = nx, ny       # number of cells
+    li           = lx, ly       # domain length in x- and y-
+    di           = @. li / ni   # grid step in x- and -y
+    origin       = 0.0, 0.0     # origin coordinates
+    grid         = Geometry(ni, li; origin = origin)
     (; xci, xvi) = grid # nodes at the center and vertices of the cells
-    dt = Inf
+
+    ## (Physical) Time domain and discretization
+    endtime = 500
+    ttot    = endtime * kyr # total simulation time
 
     # Physical properties using GeoParams ----------------
-    η0 = 1.0e22           # viscosity
-    G0 = 10^10           # elastic shear modulus
-    εbg = 1.0e-14           # background strain-rate
-    dt = 1.0e11
-    el_bg = SetConstantElasticity(; G = G0, ν = 0.5)
-    visc = LinearViscous(; η = η0)
+    η0    = 1.0e22  # viscosity
+    G0    = 1e10    # elastic shear modulus
+    εbg   = 1.0e-14 # background strain-rate
+    dt    = 1.0e11
+    el_bg = SetConstantElasticity(; G = G0, ν = 0.49)
+    visc  = LinearViscous(; η = η0)
     rheology = (
         # Low density phase
         SetMaterialParams(;
-            Phase = 1,
-            Density = ConstantDensity(; ρ = 2700.0),
[…]
