### 2025-12-15 8ce9ac3
attempt to remove `minapps` from format checker
