+++ miniapps/benchmarks/stokes2D/elastic_buildup/Elastic_BuildUp_phases_incompressible.jl
@@ -15,7 +15,7 @@ else
     JustRelax.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
 end
 
-using JustPIC,JustPIC._2D
+using JustPIC, JustPIC._2D
 const backend = @static if isCUDA
     JustPIC.CUDABackend
 else
@@ -38,45 +38,45 @@ function init_phases!(phase_ratios)
     end
 
     @parallel (@idx ni) init_phases!(phase_ratios.center)
-    @parallel (@idx ni.+1) init_phases!(phase_ratios.vertex)
+    @parallel (@idx ni .+ 1) init_phases!(phase_ratios.vertex)
     return nothing
 end
 
-const yr      = 365.25 * 3600 * 24
-const kyr     = 1.0e3 * yr
+const yr = 365.25 * 3600 * 24
+const kyr = 1.0e3 * yr
 
 # MAIN SCRIPT --------------------------------------------------------------------
 function main(igg; nx = 64, ny = 64, figdir = "model_figs")
 
     # Physical domain ------------------------------------
-    ly           = 100e3        # domain length in y
-    lx           = ly           # domain length in x
-    ni           = nx, ny       # number of cells
-    li           = lx, ly       # domain length in x- and y-
-    di           = @. li / ni   # grid step in x- and -y
-    origin       = 0.0, 0.0     # origin coordinates
-    grid         = Geometry(ni, li; origin = origin)
+    ly = 100.0e3        # domain length in y
+    lx = ly           # domain length in x
+    ni = nx, ny       # number of cells
+    li = lx, ly       # domain length in x- and y-
+    di = @. li / ni   # grid step in x- and -y
+    origin = 0.0, 0.0     # origin coordinates
+    grid = Geometry(ni, li; origin = origin)
     (; xci, xvi) = grid # nodes at the center and vertices of the cells
 
     ## (Physical) Time domain and discretization
     endtime = 500
-    ttot    = endtime * kyr # total simulation time
+    ttot = endtime * kyr # total simulation time
 
     # Physical properties using GeoParams ----------------
-    η0    = 1.0e22  # viscosity
-    G0    = 1e10    # elastic shear modulus
-    εbg   = 1.0e-14 # background strain-rate
-    dt    = 1.0e11
+    η0 = 1.0e22  # viscosity
+    G0 = 1.0e10    # elastic shear modulus
+    εbg = 1.0e-14 # background strain-rate
+    dt = 1.0e11
     el_bg = SetConstantElasticity(; G = G0, ν = 0.49)
-    visc  = LinearViscous(; η = η0)
+    visc = LinearViscous(; η = η0)
     rheology = (
         # Low density phase
         SetMaterialParams(;
-            Phase             = 1,
-            Density           = ConstantDensity(; ρ = 0e0),
-            Gravity           = ConstantGravity(; g = 0e0),
+            Phase = 1,
+            Density = ConstantDensity(; ρ = 0.0e0),
+            Gravity = ConstantGravity(; g = 0.0e0),
             CompositeRheology = CompositeRheology((visc, el_bg)),
-            Elasticity        = el_bg,
+            Elasticity = el_bg,
         ),
     )
 
@@ -86,11 +86,11 @@ function main(igg; nx = 64, ny = 64, figdir = "model_figs")
 
     # STOKES ---------------------------------------------
     # Allocate arrays needed for every Stokes problem
-    stokes    = StokesArrays(backend_JR, ni)
+    stokes = StokesArrays(backend_JR, ni)
[…]
