+++ .github/workflows/format_check.yml
@@ -45,11 +45,8 @@ jobs:
         id: runic
         run: |
           set +e
-          EXCLUDES=(
-            ':(exclude)miniapps/**'
-          )
           MERGE_BASE=$(git merge-base upstream/${{ github.base_ref }} HEAD) || exit 1
-          DIFF=$(git runic --diff "$MERGE_BASE" -- . "${EXCLUDES[@]}")
+          DIFF=$(git runic --diff $MERGE_BASE)
           EXIT_CODE=$?
 
           echo "exit_code=$EXIT_CODE" >> $GITHUB_OUTPUT
