### 2025-12-15 128fd33
update `installation.md`
