### 2025-12-15 70b5d27
Update docs/src/man/installation.md
Co-authored-by: Pascal Aellig <93140290+aelligp@users.noreply.github.com>
