### 2025-12-16 4c4c212
Implement most of Rene's suggestions (#399) & some grammar/typo fixes
