### 2025-12-17 c589af8
add citation.cff
