+++ CITATION.CFF
@@ -30,8 +30,6 @@ authors:
   family-names: "Kaus"
   given-names: "Boris J. P. "
   orcid: https://orcid.org/0000-0002-0247-8660
-- family-names: Patrick Sanan
-- family-names: Sidney Beeler
 - affiliation: Johannes Gutenberg University Mainz
   family-names: "Dominguez"
   given-names: "Hugo"
