+++ CITATION.CFF
@@ -28,7 +28,7 @@ authors:
   orcid: https://orcid.org/0000-0002-9165-6384
 - affiliation: Johannes Gutenberg University Mainz
   family-names: "Kaus"
-  given-names: "Boris J. P. "
+  given-names: "Boris J. P."
   orcid: https://orcid.org/0000-0002-0247-8660
 - affiliation: Johannes Gutenberg University Mainz
   family-names: "Dominguez"
