+++ docs/paper/paper.bib
@@ -147,8 +147,6 @@
   doi = {10.1016/j.pepi.2008.08.005},
   publisher={Elsevier}
 }
-
-
 @inproceedings{lamem,
   title={Forward and inverse modelling of lithospheric deformation on geological timescales},
   author={Kaus, Boris JP and Popov, Anton A and Baumann, T and Pusok, A and Bauville, Arthur and Fernandez, Naiara and Collignon, Marine},
@@ -159,6 +157,12 @@
   organization={John von Neumann Institute for Computing (NIC), NIC Series}
 }
 
+@inproceedings{lamem2,
+  title={Scalable hybrid multigrid for staggered grid discretizations in geodynamics},
+  author={Popov, A. A., Kaus, B. J. P., and Ibragimov, I.},
+  doi = {doi.org/10.5194/egusphere-egu25-11825},
+  organization={EGU General Assembly 2025, Vienna, Austria, 27 Apr–2 May 2025, EGU25-11825}
+}
 @article{citcom,
   title={Role of temperature-dependent viscosity and surface plates in spherical shell models of mantle convection},
   author={Zhong, Shijie and Zuber, Maria T and Moresi, Louis and Gurnis, Michael},
