### 2025-12-17 7af2917
add silver badge of SQAaas; add template files
