+++ .github/PULL_REQUEST_TEMPLATE/documentation_fix.md
@@ -1,0 +1,8 @@
+**Checklist**
+
+- [] The improved documentation only uses public API
+- [] The new code follows the [contributor guidelines](https://github.com/PTsolvers/JustRelax.jl/blob/main/CONTRIBUTING.md), in particular the [Runic Style](https://github.com/fredrikekre/Runic.jl)
+
+**Additional context**
+
+Add any other context about the problem here.
