+++ dev/assets/man_ShearBands.md.DbmgfaRH.lean.js
@@ -1,0 +1,1 @@
+import{_ as i,c as a,o as h,aA as n}from"./chunks/framework.C0SHoJ8-.js";const k="/JustRelax.jl/dev/assets/DP_nx2058_2D.CSWFAzLr.gif",y=JSON.parse('{"title":"ShearBand benchmark","description":"","frontmatter":{},"headers":[],"relativePath":"man/ShearBands.md","filePath":"man/ShearBands.md","lastUpdated":null}'),l={name:"man/ShearBands.md"};function t(p,s,e,E,r,d){return h(),a("div",null,[...s[0]||(s[0]=[n("",49)])])}const F=i(l,[["render",t]]);export{y as __pageData,F as default};
