### 2025-12-18 ab9444e
update thermal stress miniapp
