+++ miniapps/benchmarks/thermal_stress/Thermal_Stress_Magma_Chamber_nondim.jl
@@ -12,11 +12,7 @@ using JustPIC._2D
 # and to run on an AMD GPU load AMDGPU.jl (i.e. "using AMDGPU") at the beginning of the script.
 const backend = JustPIC.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
 
-using Printf, Statistics, LinearAlgebra, GeoParams, CairoMakie, CellArrays
-using StaticArrays
-using ImplicitGlobalGrid
-using MPI: MPI
-using WriteVTK
+using Printf, Statistics, LinearAlgebra, GeoParams, GLMakie
 
 # -----------------------------------------------------------------------------------------
 ## SET OF HELPER FUNCTIONS PARTICULAR FOR THIS SCRIPT --------------------------------
@@ -35,8 +31,8 @@ macro all_j(A)
     return esc(:($A[$idx_j]))
 end
 
-@parallel function init_P!(P, ρg, z)
-    @all(P) = abs(@all(ρg) * @all_j(z)) * <(@all_j(z), 0.0)
+@parallel function init_P!(P, ρg, z, sticky_air)
+    @all(P) = abs(@all(ρg) * (@all_j(z)- sticky_air)) #* <(@all_j(z), 0.0)
     return nothing
 end
 
@@ -51,7 +47,7 @@ function init_phases!(phases, particles, xc_anomaly, yc_anomaly, r_anomaly, stic
             @index(index[ip, i, j]) == 0 && continue
 
             x = @index px[ip, i, j]
-            y = -(@index py[ip, i, j]) - sticky_air
+            y = -(@index py[ip, i, j]) #- sticky_air
             if top ≤ y ≤ bottom
                 @index phases[ip, i, j] = 1.0 # crust
             end
@@ -83,15 +79,15 @@ end
 
 # Initial thermal profile
 @parallel_indices (i, j) function init_T!(T, y, sticky_air, top, bottom, dTdz, offset)
-    depth = -y[j] - sticky_air
+    depth = y[j]
 
-    if depth < top
+    if depth ≥ 0e0
         T[i + 1, j] = offset
 
-    elseif top ≤ (depth) < bottom
-        dTdZ = dTdz
+    else # if top ≤ (depth) < bottom
+        dTdZ   = dTdz
         offset = offset
-        T[i + 1, j] = (depth) * dTdZ + offset
+        T[i + 1, j] = abs(depth) * dTdZ + offset
 
     end
 
@@ -102,7 +98,7 @@ function circular_perturbation!(T, δT, xc_anomaly, yc_anomaly, r_anomaly, xvi, 
     @parallel_indices (i, j) function _circular_perturbation!(
             T, δT, xc_anomaly, yc_anomaly, r_anomaly, x, y, sticky_air
         )
-        depth = -y[j] - sticky_air
+        depth = -y[j] #- sticky_air
         if ((x[i] - xc_anomaly)^2 + (depth[j] + yc_anomaly)^2 ≤ r_anomaly^2)
             # T[i + 1, j] *= δT / 100 + 1
             T[i + 1, j] = δT
@@ -117,13 +113,27 @@ function circular_perturbation!(T, δT, xc_anomaly, yc_anomaly, r_anomaly, xvi, 
     )
 end
 
-function init_rheology(CharDim; is_compressible = false, steady_state = true)
+function linear_creep_models()
+    creep_rock = LinearViscous(; η = 1.0e23 * Pa * s)
+    creep_magma = LinearViscous(; η = 1.0e18 * Pa * s)
+    creep_air = LinearViscous(; η = 1.0e18 * Pa * s)
[…]
