+++ src/stokes/Stokes2D.jl
@@ -534,6 +534,8 @@ function _solve!(
     ni = size(stokes.P)
 
     # ~preconditioner
+    @copy stokes.P0 stokes.P
+
     ητ = deepcopy(η)
     # @hide_communication b_width begin # communication/computation overlap
     compute_maxloc!(ητ, η; window = (1, 1))
@@ -556,7 +558,6 @@ function _solve!(
     sizehint!(err_evo2, Int(iterMax))
 
     # solver loop
-    @copy stokes.P0 stokes.P
     wtime0 = 0.0
     relλ = λ_relaxation
     θ = deepcopy(stokes.P)
@@ -571,7 +572,6 @@ function _solve!(
 
     # compute buoyancy forces and viscosity
     compute_ρg!(ρg, phase_ratios, rheology, args)
-    compute_viscosity!(stokes, phase_ratios, args, rheology, viscosity_cutoff)
     displacement2velocity!(stokes, dt, flow_bcs)
 
     while iter ≤ iterMax
@@ -618,16 +618,7 @@ function _solve!(
                 )
             end
 
-            update_viscosity_τII!(
-                stokes,
-                phase_ratios,
-                args,
-                rheology,
-                viscosity_cutoff;
-                relaxation = viscosity_relaxation,
-            )
-            # end
-
+            
             if strain_increment
                 @parallel (@idx ni .+ 1) update_stresses_center_vertex_ps!(
                     @strain(stokes),
@@ -679,9 +670,17 @@ function _solve!(
                     phase_ratios.vertex,
                 )
             end
-
             update_halo!(stokes.τ.xy)
-
+            
+            update_viscosity_τII!(
+                stokes,
+                phase_ratios,
+                args,
+                rheology,
+                viscosity_cutoff;
+                relaxation = viscosity_relaxation,
+            )
+            
             @hide_communication b_width begin # communication/computation overlap
                 @parallel compute_V!(
                     @velocity(stokes)...,
@@ -768,6 +767,10 @@ function _solve!(
 
     @parallel (@idx ni .+ 1) multi_copy!(@tensor(stokes.τ_o), @tensor(stokes.τ))
     @parallel (@idx ni) multi_copy!(@tensor_center(stokes.τ_o), @tensor_center(stokes.τ))
+
+    
+    tensor_invariant!(stokes.ε)
+    tensor_invariant!(stokes.ε_pl)
 
     return (
         iter = iter,
