+++ previews/PR407/assets/man_constitutive_equations.md.DOEGO3UY.lean.js
@@ -1,0 +1,1 @@
+import{_ as l,c as a,o as T,j as t,aA as e,a as s}from"./chunks/framework.D78tWPYL.js";const L3=JSON.parse('{"title":"Constitutive equation","description":"","frontmatter":{},"headers":[],"relativePath":"man/constitutive_equations.md","filePath":"man/constitutive_equations.md","lastUpdated":null}'),n={name:"man/constitutive_equations.md"},i={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},o={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.74ex"},xmlns:"http://www.w3.org/2000/svg",width:"27.753ex",height:"2.61ex",role:"img",focusable:"false",viewBox:"0 -826.9 12267 1153.7","aria-hidden":"true"},d={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},r={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-2.028ex"},xmlns:"http://www.w3.org/2000/svg",width:"30.982ex",height:"5.188ex",role:"img",focusable:"false",viewBox:"0 -1396.5 13693.9 2293","aria-hidden":"true"},m={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},h={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.489ex"},xmlns:"http://www.w3.org/2000/svg",width:"3.002ex",height:"1.489ex",role:"img",focusable:"false",viewBox:"0 -442 1326.7 658","aria-hidden":"true"},p={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},H={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.05ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.778ex",height:"1.645ex",role:"img",focusable:"false",viewBox:"0 -705 786 727","aria-hidden":"true"},g={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},c={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.027ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.319ex",height:"2.333ex",role:"img",focusable:"false",viewBox:"0 -1019 583 1031","aria-hidden":"true"},k={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},x={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.439ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.79ex",height:"2.032ex",role:"img",focusable:"false",viewBox:"0 -704 791 898","aria-hidden":"true"},u={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},V={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.566ex"},xmlns:"http://www.w3.org/2000/svg",width:"16.695ex",height:"2.262ex",role:"img",focusable:"false",viewBox:"0 -750 7379.1 1000","aria-h
[…]
