+++ previews/PR407/assets/man_subduction2D_subduction2D.md.CuzqTqpd.js
@@ -1,0 +1,140 @@
+import{_ as i,c as a,o as n,aA as h}from"./chunks/framework.D78tWPYL.js";const l="/JustRelax.jl/previews/PR407/assets/Subduction2D_990.BLKNEd_w.png",y=JSON.parse('{"title":"2D subduction","description":"","frontmatter":{},"headers":[],"relativePath":"man/subduction2D/subduction2D.md","filePath":"man/subduction2D/subduction2D.md","lastUpdated":null}'),t={name:"man/subduction2D/subduction2D.md"};function k(e,s,p,r,E,d){return n(),a("div",null,[...s[0]||(s[0]=[h(`<h1 id="2D-subduction" tabindex="-1">2D subduction <a class="header-anchor" href="#2D-subduction" aria-label="Permalink to &quot;2D subduction {#2D-subduction}&quot;">​</a></h1><p>Model setups taken from <a href="https://doi.org/10.5194/se-15-567-2024" target="_blank" rel="noreferrer">Hummel et al 2024</a>.</p><h1 id="Initialize-packages" tabindex="-1">Initialize packages <a class="header-anchor" href="#Initialize-packages" aria-label="Permalink to &quot;Initialize packages {#Initialize-packages}&quot;">​</a></h1><p>Load JustRelax necessary modules and define backend.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> CUDA </span><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># comment this out if you are not using CUDA; or load AMDGPU.jl if you are using an AMD GPU</span></span>
+<span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> JustRelax, JustRelax</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">.</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">JustRelax2D, JustRelax</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">.</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">DataIO</span></span>
[…]
