+++ previews/PR407/assets/man_advection.md.4I44YVLy.lean.js
@@ -1,0 +1,1 @@
+import{_ as a,c as t,o as i,aA as l}from"./chunks/framework.D78tWPYL.js";const p=JSON.parse('{"title":"Field advection","description":"","frontmatter":{},"headers":[],"relativePath":"man/advection.md","filePath":"man/advection.md","lastUpdated":null}'),r={name:"man/advection.md"};function n(o,e,c,s,d,_){return i(),t("div",null,[...e[0]||(e[0]=[l("",3)])])}const h=a(r,[["render",n]]);export{p as __pageData,h as default};
