+++ previews/PR407/assets/man_installation.md.BmEdnyRQ.lean.js
@@ -1,0 +1,1 @@
+import{_ as i,c as a,o as n,aA as t}from"./chunks/framework.D78tWPYL.js";const g=JSON.parse('{"title":"Installation","description":"","frontmatter":{},"headers":[],"relativePath":"man/installation.md","filePath":"man/installation.md","lastUpdated":null}'),e={name:"man/installation.md"};function l(p,s,h,k,d,r){return n(),a("div",null,[...s[0]||(s[0]=[t("",17)])])}const E=i(e,[["render",l]]);export{g as __pageData,E as default};

[…]
