+++ miniapps/benchmarks/thermal_stress/Thermal_Stress_Magma_Chamber_nondim.jl
@@ -32,7 +32,7 @@ macro all_j(A)
 end
 
 @parallel function init_P!(P, ρg, z, sticky_air)
-    @all(P) = abs(@all(ρg) * (@all_j(z)- sticky_air)) #* <(@all_j(z), 0.0)
+    @all(P) = abs(@all(ρg) * (@all_j(z) - sticky_air)) #* <(@all_j(z), 0.0)
     return nothing
 end
 
@@ -81,11 +81,11 @@ end
 @parallel_indices (i, j) function init_T!(T, y, sticky_air, top, bottom, dTdz, offset)
     depth = y[j]
 
-    if depth ≥ 0e0
+    if depth ≥ 0.0e0
         T[i + 1, j] = offset
 
     else # if top ≤ (depth) < bottom
-        dTdZ   = dTdz
+        dTdZ = dTdz
         offset = offset
         T[i + 1, j] = abs(depth) * dTdZ + offset
 
@@ -121,19 +121,19 @@ function linear_creep_models()
 end
 
 function nonlinear_creep_models()
-    creep_rock  = DislocationCreep(; A = 1.67e-24Pa^(-(35//10)) / s, n = 3.5, E = 1.87e5J / mol, V = 0*6.0e-6m^3 / mol, r = 0.0, R = 8.3145J / mol / K)
-    creep_magma = DislocationCreep(; A = 1.67e-24Pa^(-(35//10)) / s, n = 3.5, E = 1.87e5J / mol, V = 0*6.0e-6m^3 / mol, r = 0.0, R = 8.3145J / mol / K)
-    creep_air   = LinearViscous(; η = 1.0e18 * Pa * s)
+    creep_rock = DislocationCreep(; A = 1.67e-24Pa^(-(35 // 10)) / s, n = 3.5, E = 1.87e5J / mol, V = 0 * 6.0e-6m^3 / mol, r = 0.0, R = 8.3145J / mol / K)
+    creep_magma = DislocationCreep(; A = 1.67e-24Pa^(-(35 // 10)) / s, n = 3.5, E = 1.87e5J / mol, V = 0 * 6.0e-6m^3 / mol, r = 0.0, R = 8.3145J / mol / K)
+    creep_air = LinearViscous(; η = 1.0e18 * Pa * s)
     return creep_rock, creep_magma, creep_air
 end
 
 function init_rheology(creep_rock, creep_magma, creep_air, CharDim; is_compressible = false, steady_state = true)
     # plasticity setup
-    do_DP    = true          # do_DP=false: Von Mises, do_DP=true: Drucker-Prager (friction angle)
-    η_reg    = 1.0e20Pa * s  # regularisation "viscosity" for Drucker-Prager
-    Coh      = 10.0MPa       # yield stress. If do_DP=true, τ_y stand for the cohesion: c*cos(ϕ)
-    ϕ        = 30.0 * do_DP  # friction angle
-    G0       = 6.0e11Pa      # elastic shear modulus
+    do_DP = true          # do_DP=false: Von Mises, do_DP=true: Drucker-Prager (friction angle)
+    η_reg = 1.0e20Pa * s  # regularisation "viscosity" for Drucker-Prager
+    Coh = 10.0MPa       # yield stress. If do_DP=true, τ_y stand for the cohesion: c*cos(ϕ)
+    ϕ = 30.0 * do_DP  # friction angle
+    G0 = 6.0e11Pa      # elastic shear modulus
     G_magma = 6.0e11Pa      # elastic shear modulus perturbation
 
     soft_C = NonLinearSoftening(; ξ₀ = ustrip(Coh), Δ = ustrip(Coh) / 2) # softening law
[…]
