### 2025-12-19 PTsolvers/JustRelax.jl#395 @gassmoeller
Thanks for the explanation and adding the documentation! I guess that a number of users will be new to Julia like me and this instruction is going to be helpful.

However, when I follow exactly the instructions for the miniapps you added in #403 I still run into problems, because the dependencies of the miniapps are not installed. I followed your instructions to the letter in a cloned JustRelax repo, but it doesnt look like your instructions are installing/activating any packages. Did I do something wrong?

If I follow the instructions for the tests you provide above everything works as expected.

Here is my screen output for the miniapps:
```
rene@rene-laptop:~/work/reviews/2025-joss-justrelax/JustRelax.jl$ julia
        Info The latest version of Julia in the `release` channel is 1.12.3+0.x64.linux.gnu. You currently have `1.12.2+0.x64.linux.gnu` installed. Run:

  juliaup update

in your terminal shell to install Julia 1.12.3+0.x64.linux.gnu and update the `release` channel to that version.
               _
   _       _ _(_)_     |  Documentation: https://docs.julialang.org
  (_)     | (_) (_)    |
   _ _   _| |_  __ _   |  Type "?" for help, "]?" for Pkg help.
  | | | | | | |/ _` |  |
  | | |_| | | | (_| |  |  Version 1.12.2 (2025-11-20)
 _/ |\__'_|_|_|\__'_|  |  Official https://julialang.org release
|__/                   |

(@v1.12) pkg> activate miniapps
  Activating project at `~/work/reviews/2025-joss-justrelax/JustRelax.jl/miniapps`

(miniapps) pkg> instantiate

(miniapps) pkg> activate .
  Activating project at `~/work/reviews/2025-joss-justrelax/JustRelax.jl`

julia> using JustRelax

     ██╗██╗   ██╗███████╗████████╗██████╗ ███████╗██╗      █████╗ ██╗  ██╗
     ██║██║   ██║██╔════╝╚══██╔══╝██╔══██╗██╔════╝██║     ██╔══██╗╚██╗██╔╝
     ██║██║   ██║███████╗   ██║   ██████╔╝█████╗  ██║     ███████║ ╚███╔╝
██   ██║██║   ██║╚════██║   ██║   ██╔══██╗██╔══╝  ██║     ██╔══██║ ██╔██╗
╚█████╔╝╚██████╔╝███████║   ██║   ██║  ██║███████╗███████╗██║  ██║██╔╝ ██╗
 ╚════╝  ╚═════╝ ╚══════╝   ╚═╝   ╚═╝  ╚═╝╚══════╝╚══════╝╚═╝  ╚═╝╚═╝  ╚═╝

Version: 0.4.2
Latest commit: Paper comments
Commit date: Mon Dec 8 15:21:47 2025 +0100

julia> include("miniapps/benchmarks/stokes2D/shear_band/ShearBand2D.jl")
ERROR: LoadError: ArgumentError: Package GLMakie not found in current path.
- Run `import Pkg; Pkg.add("GLMakie")` to install the GLMakie package.
Stacktrace:
 [1] macro expansion
   @ ./loading.jl:2400 [inlined]
 [2] macro expansion
   @ ./lock.jl:376 [inlined]
 [3] __require(into::Module, mod::Symbol)
   @ Base ./loading.jl:2383
 [4] require(into::Module, mod::Symbol)
   @ Base ./loading.jl:2359
 [5] include(mapexpr::Function, mod::Module, _path::String)
   @ Base ./Base.jl:307
 [6] top-level scope
   @ REPL[5]:1
in expression starting at /home/rene/work/reviews/2025-joss-justrelax/JustRelax.jl/miniapps/benchmarks/stokes2D/shear_band/ShearBand2D.jl:1
```
