+++ previews/PR408/assets/man_boundary_conditions.md.DOPve5aY.lean.js
@@ -1,0 +1,1 @@
+import{_ as e,c as a,o as t,j as i,aA as l,a as n}from"./chunks/framework.BIqLJ5tb.js";const A=JSON.parse('{"title":"Flow boundary conditions","description":"","frontmatter":{},"headers":[],"relativePath":"man/boundary_conditions.md","filePath":"man/boundary_conditions.md","lastUpdated":null}'),h={name:"man/boundary_conditions.md"},p={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},k={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-1.034ex"},xmlns:"http://www.w3.org/2000/svg",width:"7.487ex",height:"3.2ex",role:"img",focusable:"false",viewBox:"0 -957.1 3309.4 1414.3","aria-hidden":"true"},r={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},d={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"0"},xmlns:"http://www.w3.org/2000/svg",width:"1.414ex",height:"1.538ex",role:"img",focusable:"false",viewBox:"0 -680 625 680","aria-hidden":"true"},o={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},g={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.357ex"},xmlns:"http://www.w3.org/2000/svg",width:"6.182ex",height:"1.864ex",role:"img",focusable:"false",viewBox:"0 -666 2732.5 823.8","aria-hidden":"true"},E={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},y={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"0"},xmlns:"http://www.w3.org/2000/svg",width:"1.414ex",height:"1.538ex",role:"img",focusable:"false",viewBox:"0 -680 625 680","aria-hidden":"true"},Q={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},c={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-1.034ex"},xmlns:"http://www.w3.org/2000/svg",width:"7.487ex",height:"3.07ex",role:"img",focusable:"false",viewBox:"0 -899.6 3309.4 1356.7","aria-hidden":"true"},T={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},m={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"0"},xmlns:"http://www.w3.org/2000/svg",width:"1.414ex",height:"1.538ex",role:"img",focusable:"false",viewBox:"0 -680 625 680","aria-hidden":"true"},u={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},F={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"0"},xmlns:"http://www.w3.org/2000/svg",width:"1.633ex",height:"1.593ex",role:"img",focusable:"false",viewBox:"0 -704 722 704","aria-hidden":"true"};function b(x,s,f,C,w,D){return t(),a("div",null,[s[21]||(s[21]=i("h1",{id:"Flow-boundary-conditions",tabindex:"-1"},[n("Flow boundary conditions "),i("a",{class:"
[…]
