+++ previews/PR408/assets/man_equations_APT.md.fwF3f4MX.lean.js
@@ -1,0 +1,1 @@
+import{_ as o,c as e,o as a,j as t,a as T,aA as l}from"./chunks/framework.BIqLJ5tb.js";const F2=JSON.parse('{"title":"","description":"","frontmatter":{},"headers":[],"relativePath":"man/equations_APT.md","filePath":"man/equations_APT.md","lastUpdated":null}'),s={name:"man/equations_APT.md"},n={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},r={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.464ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.473ex",height:"2.034ex",role:"img",focusable:"false",viewBox:"0 -694 651 899","aria-hidden":"true"},m={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},d={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-2.016ex"},xmlns:"http://www.w3.org/2000/svg",width:"36.232ex",height:"5.163ex",role:"img",focusable:"false",viewBox:"0 -1391 16014.7 2282","aria-hidden":"true"},i={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},h={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-2.016ex"},xmlns:"http://www.w3.org/2000/svg",width:"18.174ex",height:"5.163ex",role:"img",focusable:"false",viewBox:"0 -1391 8033 2282","aria-hidden":"true"},p={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},g={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-2.016ex"},xmlns:"http://www.w3.org/2000/svg",width:"22.869ex",height:"5.163ex",role:"img",focusable:"false",viewBox:"0 -1391 10107.9 2282","aria-hidden":"true"},x={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},w={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-2.109ex"},xmlns:"http://www.w3.org/2000/svg",width:"31.878ex",height:"5.348ex",role:"img",focusable:"false",viewBox:"0 -1432 14089.9 2364","aria-hidden":"true"},u={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},c={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-2.158ex"},xmlns:"http://www.w3.org/2000/svg",width:"28.804ex",height:"5.448ex",role:"img",focusable:"false",viewBox:"0 -1454 12731.4 2408","aria-hidden":"true"},H={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},f={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.464ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.473ex",height:"2.034
[…]
