+++ previews/PR408/assets/man_license.md.BfdEC6Kw.js
@@ -1,0 +1,1 @@
+import{_ as s,c as n,o as a,j as e,a as t}from"./chunks/framework.BIqLJ5tb.js";const R=JSON.parse('{"title":"License","description":"","frontmatter":{},"headers":[],"relativePath":"man/license.md","filePath":"man/license.md","lastUpdated":null}'),i={name:"man/license.md"};function r(l,o,c,T,O,E){return a(),n("div",null,[...o[0]||(o[0]=[e("h1",{id:"license",tabindex:"-1"},[t("License "),e("a",{class:"header-anchor",href:"#license","aria-label":'Permalink to "License {#license}"'},"​")],-1),e("blockquote",null,[e("p",null,"MIT License"),e("p",null,[t("Copyright (c) 2021 Albert de Montserrat, Pascal Aellig and Authors (see "),e("a",{href:"/JustRelax.jl/previews/PR408/man/authors#Authors"},"Authors"),t(")")]),e("p",null,'Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:'),e("p",null,"The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software."),e("p",null,'THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.')],-1)])])}const d=s(i,[["render",r]]);export{R as __pageData,d as default};
