+++ .github/PULL_REQUEST_TEMPLATE/new_feature.md
@@ -1,14 +1,0 @@
-**What feature report was addressed?**
-
-Fixes ...
-
-**Checklist**
-
-- [] Appropriate tests were added
-- [] The new feature was added in a way that does not break public API
-- [] New documentation related to the new feature was added
-- [] The new code follows the [contributor guidelines](https://github.com/PTsolvers/JustRelax.jl/blob/main/CONTRIBUTING.md), in particular the [Runic Style](https://github.com/fredrikekre/Runic.jl)
-
-**Additional context**
-
-Add any other context about the problem here.
