+++ previews/PR408/assets/chunks/@localSearchIndexroot.CGHa5wET.js
@@ -1,1 +1,0 @@
-const e='{"documentCount":153,"nextId":153,"documentIds":{"0":"/JustRelax.jl/previews/PR408/#What-is-JustRelax.jl?","1":"/JustRelax.jl/previews/PR408/#How-to-Install-JustRelax.jl?","2":"/JustRelax.jl/previews/PR408/#Example-miniapps","3":"/JustRelax.jl/previews/PR408/#funding","4":"/JustRelax.jl/previews/PR408/man/Blankenbach#Blankenbach-benchmark","5":"/JustRelax.jl/previews/PR408/man/Blankenbach#Initialize-packages","6":"/JustRelax.jl/previews/PR408/man/Blankenbach#script","7":"/JustRelax.jl/previews/PR408/man/Blankenbach#Model-domain","8":"/JustRelax.jl/previews/PR408/man/Blankenbach#rheology","9":"/JustRelax.jl/previews/PR408/man/Blankenbach#Initialize-particles","10":"/JustRelax.jl/previews/PR408/man/Blankenbach#Temperature-anomaly","11":"/JustRelax.jl/previews/PR408/man/Blankenbach#Stokes-and-heat-diffusion-arrays","12":"/JustRelax.jl/previews/PR408/man/Blankenbach#Initialize-thermal-profile-and-viscosity-fields","13":"/JustRelax.jl/previews/PR408/man/Blankenbach#Boundary-conditions","14":"/JustRelax.jl/previews/PR408/man/Blankenbach#Pseuo-transient-coefficients","15":"/JustRelax.jl/previews/PR408/man/Blankenbach#Just-before-solving-the-problem...","16":"/JustRelax.jl/previews/PR408/man/Blankenbach#Advancing-one-time-step","17":"/JustRelax.jl/previews/PR408/man/Blankenbach#visualization","18":"/JustRelax.jl/previews/PR408/man/Blankenbach#fields","19":"/JustRelax.jl/previews/PR408/man/Blankenbach#Final-model","20":"/JustRelax.jl/previews/PR408/man/GeoParams#List-of-GeoParams-functions","21":"/JustRelax.jl/previews/PR408/man/ShearBands#ShearBand-benchmark","22":"/JustRelax.jl/previews/PR408/man/ShearBands#Initialize-packages","23":"/JustRelax.jl/previews/PR408/man/ShearBands#script","24":"/JustRelax.jl/previews/PR408/man/ShearBands#Model-domain","25":"/JustRelax.jl/previews/PR408/man/ShearBands#Physical-properties-using-GeoParams","26":"/JustRelax.jl/previews/PR408/man/ShearBands#rheology","27":"/JustRelax.jl/previews/PR408/man/ShearBands#Phase-anomaly","28":"/JustRelax.jl/previews/PR408/man/ShearBands#Stokes-arrays","29":"/JustRelax.jl/previews/PR408/man/ShearBands#Initialize-viscosity-fields","30":"/JustRelax.jl/previews/PR408/man/ShearBands#Boundary-conditions","31":"/JustRelax.jl/previews/PR408/man/ShearBands#Pseuo-transient-coefficients","32":"/JustRelax.jl/previews/PR408/man/ShearBands#Just-before-solving-the-problem...","33":"/JustRelax.jl/previews/PR408/man/ShearBands#Advancing-one-time-step","34":"/JustRelax.jl/previews/PR408/man/ShearBands#visualization","35":"/JustRelax.jl/previews/PR408/man/ShearBands#fields","36":"/JustRelax.jl/previews/PR408/man/ShearBands#Final-model","37":"/JustRelax.jl/previews/PR408/man/JustPIC#List-of-JustPIC-functions","38":"/Just
[…]
