+++ previews/PR408/assets/man_checkpointing.md.BC7TNOR8.js
@@ -1,11 +1,11 @@
-import{_ as s,c as a,o as t,aA as e}from"./chunks/framework.BIqLJ5tb.js";const c=JSON.parse('{"title":"Checkpointing","description":"","frontmatter":{},"headers":[],"relativePath":"man/checkpointing.md","filePath":"man/checkpointing.md","lastUpdated":null}'),n={name:"man/checkpointing.md"};function l(h,i,p,k,o,d){return t(),a("div",null,[...i[0]||(i[0]=[e(`<h1 id="checkpointing" tabindex="-1">Checkpointing <a class="header-anchor" href="#checkpointing" aria-label="Permalink to &quot;Checkpointing&quot;">​</a></h1><p>It is common to save the state of a simulation at regular intervals, especially for long-running simulations. This allows you to restart the simulation from the last saved state in case of interruptions or to continue the simulation at a later time without losing progress. JustRelax provides a simple way to save and load checkpoint files. Two checkpointing functions are available for the most common file extensions (HDF5 and JLD2). By loading the <code>DataIO</code> module, you gain access to these checkpointing functions as well as VTK saving functions for later visualization with <a href="https://www.paraview.org/" target="_blank" rel="noreferrer">ParaView</a>. For more details on the vtk, see the <a href="./subduction2D/subduction2D">here</a>.</p><p>!!! tip &quot;JustPIC checkpointing&quot; A similar checkpointing function is defined by <a href="https://juliageodynamics.github.io/JustPIC.jl/dev/IO/" target="_blank" rel="noreferrer">JustPIC.jl</a> to save the state of the particles.</p><div class="vp-code-group vp-adaptive-theme"><div class="tabs"><input type="radio" name="group-ROQyK" id="tab-fVCXnwT" checked><label data-title="2D module" for="tab-fVCXnwT">2D module</label><input type="radio" name="group-ROQyK" id="tab-i-k0CMc"><label data-title="3D module" for="tab-i-k0CMc">3D module</label></div><div class="blocks"><div class="language-julia vp-adaptive-theme active"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> JustRelax, JustRelax</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">.</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">JustRelax2D</span></span>
[…]
