+++ previews/PR408/assets/chunks/VPLocalSearchBox.HJ9gPAxR.js
@@ -1,8 +1,0 @@
-var Ft=Object.defineProperty;var Ot=(a,e,t)=>e in a?Ft(a,e,{enumerable:!0,configurable:!0,writable:!0,value:t}):a[e]=t;var Ae=(a,e,t)=>Ot(a,typeof e!="symbol"?e+"":e,t);import{V as Ct,D as le,h as ge,aB as tt,aC as Rt,aD as At,aE as Mt,q as je,aF as Lt,d as Dt,aG as st,p as he,aH as Pt,ac as zt,s as Vt,aI as $t,v as Me,P as fe,O as Se,aJ as jt,ak as Bt,W as Wt,R as Kt,$ as Jt,b as qt,o as H,j as x,a0 as Ut,a2 as Gt,k as L,a5 as Ht,ad as Qt,c as Z,e as Ee,n as nt,B as it,F as rt,a as pe,t as ve,aa as Yt,aK as at,aL as Zt,aq as Xt,av as es,aM as ts,_ as ss}from"./framework.BIqLJ5tb.js";import{u as ns,c as is}from"./theme.Voy5AS-Y.js";const rs={root:()=>Ct(()=>import("./@localSearchIndexroot.BXFFIAUf.js"),[])};/*!
-* tabbable 6.3.0
-* @license MIT, https://github.com/focus-trap/tabbable/blob/master/LICENSE
-*/var mt=["input:not([inert])","select:not([inert])","textarea:not([inert])","a[href]:not([inert])","button:not([inert])","[tabindex]:not(slot):not([inert])","audio[controls]:not([inert])","video[controls]:not([inert])",'[contenteditable]:not([contenteditable="false"]):not([inert])',"details>summary:first-of-type:not([inert])","details:not([inert])"],Ne=mt.join(","),gt=typeof Element>"u",re=gt?function(){}:Element.prototype.matches||Element.prototype.msMatchesSelector||Element.prototype.webkitMatchesSelector,Fe=!gt&&Element.prototype.getRootNode?function(a){var e;return a==null||(e=a.getRootNode)===null||e===void 0?void 0:e.call(a)}:function(a){return a==null?void 0:a.ownerDocument},ye=function(e,t){var s;t===void 0&&(t=!0);var n=e==null||(s=e.getAttribute)===null||s===void 0?void 0:s.call(e,"inert"),r=n===""||n==="true",i=r||t&&e&&ye(e.parentNode);return i},as=function(e){var t,s=e==null||(t=e.getAttribute)===null||t===void 0?void 0:t.call(e,"contenteditable");return s===""||s==="true"},bt=function(e,t,s){if(ye(e))return[];var n=Array.prototype.slice.apply(e.querySelectorAll(Ne));return t&&re.call(e,Ne)&&n.unshift(e),n=n.filter(s),n},Oe=function(e,t,s){for(var n=[],r=Array.from(e);r.length;){var i=r.shift();if(!ye(i,!1))if(i.tagName==="SLOT"){var o=i.assignedElements(),l=o.length?o:i.children,c=Oe(l,!0,s);s.flatten?n.push.apply(n,c):n.push({scopeParent:i,candidates:c})}else{var h=re.call(i,Ne);h&&s.filter(i)&&(t||!e.includes(i))&&n.push(i);var m=i.shadowRoot||typeof s.getShadowRoot=="function"&&s.getShadowRoot(i),f=!ye(m,!1)&&(!s.shadowRootFilter||s.shadowRootFilter(i));if(m&&f){var g=Oe(m===!0?i.children:m.children,!0,s);s.flatten?n.push.apply(n,g):n.push({scopeParent:i,candidates:g})}else r.unshift.apply(r,i.children)}}return n},
[…]
