### 2026-01-06 7f5e2f9
fix fig 3 caption
