+++ docs/paper/paper.md
@@ -59,7 +59,7 @@ bibliography: paper.bib
 
 <!-- Simulating Earth's thermo-mechanical evolution requires solving coupled, non-linear Stokes and heat‑diffusion problems across large domains with sharp material contrasts, complex rheologies, and long time scales. These calculations are computationally demanding and traditionally rely on monolithic, CPU‑oriented codes built around matrix‑based linear solvers. Such designs are hard to port efficiently to modern heterogeneous HPC systems (e.g. multi‑XPU nodes). -->
 
-Simulating Earth's thermo-mechanical evolution requires solving coupled, non-linear Stokes and heat‑diffusion problems across large domains with sharp material contrasts, complex rheologies, and long time scales. These calculations are computationally demanding and traditionally rely on CPU‑oriented codes often built around matrix‑based linear solvers (e.g. I3ELVIS [@gerya2015plate], StagYY [@stagyy], [CITCOM-S](https://github.com/geodynamics/citcoms) [@citcom], [Underworld](https://github.com/underworldcode/underworld2) [@underworld]), with designs that may be hard to port efficiently to modern heterogeneous HPC systems (e.g. multi‑XPU nodes). While other geodynamics code, such as [ASPECT](https://github.com/geodynamics/aspect) [@aspect], [TerraNeo](https://terraneo.fau.de/), and [LaMEM](https://github.com/UniMainzGeo/LaMEM/) [@lamem2], do support some matrix-free solvers, they remain CPU-oriented codes.
+Simulating Earth's thermo-mechanical evolution requires solving coupled, non-linear Stokes and heat‑diffusion problems across large domains with sharp material contrasts, complex rheologies, and long time scales. These calculations are computationally demanding and traditionally rely on CPU‑oriented codes often built around matrix‑based linear solvers (e.g. I3ELVIS [@gerya2015plate], StagYY [@stagyy], [CITCOM-S](https://github.com/geodynamics/citcoms) [@citcom], [Underworld](https://github.com/underworldcode/underworld2) [@underworld]), with designs that may be hard to port efficiently to modern heterogeneous HPC systems (e.g. multi‑XPU nodes). While other geodynamics code, such as [ASPECT](https://github.com/geodynamics/aspect) [@aspect], [TerraNeo](https://terraneo.fau.de/) [@TerraNeo], and [LaMEM](https://github.com/UniMainzGeo/LaMEM/) [@lamem2], do support some matrix-free solvers, they remain CPU-oriented codes as in their current state.
 
 With `JustRelax.jl` we aim to provide a compact, modular, and highly portable (CPU/GPU) toolkit written in Julia. It implements a matrix‑free Accelerated Pseudo‑Transient (APT) solver that reduces global linear‑algebra bottlenecks and can be ported efficiently to hardware accelerators, while outsourcing some physical calculations, parallelization, and I/O to other packages, as shown in \ref{fig:dependencies}. This keeps the core solver as contained as possible and makes the software more flexible. Julia’s high‑level syntax and package manager simplify installation and scripting, lowering the barrier for students and researchers to prototype and test new physics.
 
