+++ .github/PULL_REQUEST_TEMPLATE/pull_request_template.md
@@ -10,7 +10,7 @@
 **Checklist**
 - [ ] The PR title is descriptive and starts with the appropriate tag: `[BUGFIX]`, `[ADDITION]`, `[DOC]`, etc.
 - [ ] Appropriate tests were added or updated
-- [ ] Potentially effected miniapps have also been updated
+- [ ] Affected miniapps have also been updated
 - [ ] The new feature was added in a way that does not break public API
 - [ ] New documentation related to the new feature was added
 - [ ] The new code follows the [contributor guidelines](https://github.com/PTsolvers/JustRelax.jl/blob/main/CONTRIBUTING.md), in particular the [Runic Style](https://github.com/fredrikekre/Runic.jl)
