### 2026-01-06 04671c1
replace GLMakie by CairoMakie
