+++ docs/src/man/Blankenbach.md
@@ -349,7 +349,7 @@ push!(trms, t)
 ### Visualization
 We will use [Makie.jl](https://github.com/MakieOrg/Makie.jl) to visualize the results
 ```julia
-using GLMakie
+using CairoMakie
 ```
 
 ### Fields
