### 2026-01-06 6ea6bb4
update thermal stresses miniapp and test
