### 2026-01-06 PTsolvers/JustRelax.jl#381 @albert-de-montserrat
Thanks @shipengcheng1230 for all your review!

Regarding your comments:

1. References (to those with an available citation) and links to their repositories have been added to the manuscript.
2. I believe the sentence in the manuscript is correct. The y-axis is `time_N_processes / time_single_process`, `time_N_processes` being the "real" (not multiplied by the number of processes) wall time of a run with `N` MPI ranks. So the closer this number is to 1, the better.
3. Done
4. Done
5. Links point to the correct location.
6. There were indeed some things out-of-date in this example. I have addressed in [this commit](https://github.com/PTsolvers/JustRelax.jl/commit/e79e64346d43a646809dbfe6d742bc981ea80700), it should be up-to-date now.
