+++ src/Utils.jl
@@ -513,7 +513,8 @@ Do a continuation step `exp((1-ν)*log(x_old) + ν*log(x_new))` with damping par
     x_cont = exp((1 - ν) * log(x_old) + ν * log(x_new))
     return isnan(x_cont) ? 0.0 : x_cont
 end
-@inline continuation_linear(x_new, x_old, ν) = muladd((1 - ν), x_old, ν * x_new) # (1 - ν) * x_old + ν * x_new
+@inline continuation_linear(x_new, x_old, ν) = (1 - ν) * x_old + ν * x_new
+# @inline continuation_linear(x_new, x_old, ν) = muladd((1 - ν), x_old, ν * x_new) # (1 - ν) * x_old + ν * x_new
 
 # Others
 

[…]
