### 2026-01-07 ba2b225
power law prototypes
