+++ Stokes2D_DYREL_PowerLaw.jl
@@ -1,0 +1,419 @@
+# Initialisation
+using Plots, Printf, Statistics, LinearAlgebra
+# Macros
+@views    av(A) = 0.25*(A[1:end-1,1:end-1].+A[2:end,1:end-1].+A[1:end-1,2:end].+A[2:end,2:end])
+@views av_xa(A) =  0.5*(A[1:end-1,:].+A[2:end,:])
+@views av_ya(A) =  0.5*(A[:,1:end-1].+A[:,2:end])
+@views av4_harm(A) = 1.0./( 0.25.*(1.0./A[1:end-1,1:end-1].+1.0./A[2:end,1:end-1].+1.0./A[1:end-1,2:end].+1.0./A[2:end,2:end]) ) 
+
+# can be replaced by AD
+function Gershgorin_Stokes2D_SchurComplement(ηc, ηv, γ, Δx, Δy, ncx  ,ncy)
+        
+    ηN    = ones(ncx-1, ncy)
+    ηS    = ones(ncx-1, ncy)
+    ηN[:,1:end-1] .= ηv[2:end-1,2:end-1]
+    ηS[:,2:end-0] .= ηv[2:end-1,2:end-1]
+    ηW    = ηc[1:end-1,:]
+    ηE    = ηc[2:end-0,:]
+    ebW   = γ[1:end-1,:] 
+    ebE   = γ[2:end-0,:] 
+    Cxx   = ones(ncx-1, ncy)
+    Cxy   = ones(ncx-1, ncy)
+    @. Cxx = abs.(ηN ./ Δy .^ 2) + abs.(ηS ./ Δy .^ 2) + abs.(ebE ./ Δx .^ 2 + (4 // 3) * ηE ./ Δx .^ 2) + abs.(ebW ./ Δx .^ 2 + (4 // 3) * ηW ./ Δx .^ 2) + abs.(-(-ηN ./ Δy - ηS ./ Δy) ./ Δy + (ebE ./ Δx + ebW ./ Δx) ./ Δx + ((4 // 3) * ηE ./ Δx + (4 // 3) * ηW ./ Δx) ./ Δx)
+    @. Cxy = abs.(ebE ./ (Δx .* Δy) - 2 // 3 * ηE ./ (Δx .* Δy) + ηN ./ (Δx .* Δy)) + abs.(ebE ./ (Δx .* Δy) - 2 // 3 * ηE ./ (Δx .* Δy) + ηS ./ (Δx .* Δy)) + abs.(ebW ./ (Δx .* Δy) + ηN ./ (Δx .* Δy) - 2 // 3 * ηW ./ (Δx .* Δy)) + abs.(ebW ./ (Δx .* Δy) + ηS ./ (Δx .* Δy) - 2 // 3 * ηW ./ (Δx .* Δy))
+    
+    Dx  = ones(ncx-1, ncy)
+    @. Dx .= -(-ηN ./ Δy - ηS ./ Δy) ./ Δy + (ebE ./ Δx + ebW ./ Δx) ./ Δx + ((4 // 3) * ηE ./ Δx + (4 // 3) * ηW ./ Δx) ./ Δx
+
+    ηE    = ones(ncx, ncy-1)
+    ηW    = ones(ncx, ncy-1)
+    ηE[1:end-1,:] .= ηv[2:end-1,2:end-1]
+    ηW[2:end-0,:] .= ηv[2:end-1,2:end-1]
+    ηS    = ηc[:,1:end-1]
+    ηN    = ηc[:,2:end-0]
+    ebS  = γ[:,1:end-1] 
+    ebN  = γ[:,2:end-0] 
+    Cyy  = ones(ncx, ncy-1)
+    Cyx  = ones(ncx, ncy-1)
[…]
