### 2026-01-07 ac169e7
working power law solver
