+++ src/rheology/Viscosity.jl
@@ -166,7 +166,7 @@ end
 
         # compute and update stress viscosity
         ηi = fn_viscosity(rheology, AII, args_ij)                
-        ηi = continuation_linear(ηi, η[I...], ν)
+        ηi = continuation_log(ηi, η[I...], ν)
         η[I...] = clamp(ηi, cutoff...)
     end
 
@@ -198,7 +198,7 @@ end
         # compute and update stress viscosity
         ηi = fn_viscosity(rheology, AII_ij, args_ij)
                 
-        ηi = continuation_linear(ηi, η[I...], ν)
+        ηi = continuation_log(ηi, η[I...], ν)
         η[I...] = clamp(ηi, cutoff...)
     end
 
@@ -367,7 +367,7 @@ end
         # compute and update stress viscosity
         ηi = compute_phase_viscosity(rheology, ratio_ij, AII, fn_viscosity, args_ij)
                 
-        ηi = continuation_linear(ηi, η[I...], ν)
+        ηi = continuation_log(ηi, η[I...], ν)
         η[I...] = clamp(ηi, cutoff...)
     end
 
@@ -402,7 +402,7 @@ end
 
         # update stress and effective viscosity
         ηi = fn_viscosity(rheology, AII, args_ijk)
-        ηi = continuation_linear(ηi, η[I...], ν)
+        ηi = continuation_log(ηi, η[I...], ν)
         η[I...] = clamp(ηi, cutoff...)
     end
 
@@ -454,7 +454,7 @@ end
         # update stress and effective viscosity
         ηi = compute_phase_viscosity(rheology, ratio_ijk, AII, fn_viscosity, args_ijk)
         
-        ηi = continuation_linear(ηi, η[I...], ν)
+        ηi = continuation_log(ηi, η[I...], ν)
         η[I...] = clamp(ηi, cutoff...)
     end
 
