+++ src/DYREL/DYREL2D.jl
@@ -161,26 +161,27 @@ function _solve_DYREL!(
             args,
         )
 
+        # itPH == 2 && break
         # Residual check
-        errVx = norm(stokes.R.Rx)
-        errVy = norm(stokes.R.Ry)
-        errPt = norm(stokes.R.RP)
+        errVx = norm(stokes.R.Rx) / √(length(stokes.R.Rx))
+        errVy = norm(stokes.R.Ry) / √(length(stokes.R.Ry))
+        errPt = norm(stokes.R.RP) / √(length(stokes.R.RP))
         if isone(itPH)
             errVx0 = errVx
             errVy0 = errVy
             errPt0 = errPt
         end
         err = maximum(
-            (min(errVx/errVx0, errVx), min(errVy/errVy0, errVy))
-            # (min(errVx/errVx0, errVx), min(errVy/errVy0, errVy), min(errPt/errPt0, errPt))
-            # (errVx,  errVy)
+            # (min(errVx/errVx0, errVx), min(errVy/errVy0, errVy))
+            (min(errVx/errVx0, errVx), min(errVy/errVy0, errVy), min(errPt/errPt0, errPt))
         )
+        isnan(err) && error("NaN detected in outer loop")
         @printf("itPH = %02d iter = %06d iter/nx = %03d, err = %1.3e norm[Rx=%1.3e, Ry=%1.3e, Rp=%1.3e] \n", itPH, iter, iter/ni[1], err, min(errVx/errVx0, errVx), min(errVy/errVy0, errVy), min(errPt/errPt0, errPt))
         # @printf("itPH = %02d iter = %06d iter/nx = %03d, err = %1.3e norm[Rx=%1.3e, Ry=%1.3e, Rp=%1.3e] \n", itPH, iter, iter/ni[1], err, errVx, errVy, errPt)
         err < ϵ && break
 
         # Set tolerance of velocity solve proportional to residual
-        ϵ_vel = err * rel_drop
+        ϵ_vel = 1.5e-6 #err * rel_drop
         itPT  = 0
         while (err > ϵ_vel && itPT ≤ iterMax)
             itPT   += 1
@@ -250,9 +251,8 @@ function _solve_DYREL!(
             # Residual check
             if iszero(iter % nout)
                
-                errVx = norm(Dx .* stokes.R.Rx)
-                errVy = norm(Dy .* stokes.R.Ry)
-                # isnan(errVx) && error("NaN detected in velocity residuals")
+                errVx = norm(Dx .* stokes.R.Rx) / √(length(stokes.R.Rx))
+                errVy = norm(Dy .* stokes.R.Ry) / √(length(stokes.R.Ry))
                 
                 if iter == nout 
                     errVx00 = errVx
@@ -263,6 +263,8 @@ function _solve_DYREL!(
                     # (errVx, errVy) 
                     (errVx / errVx00, errVy / errVy00) 
                 )
+                isnan(err) && error("NaN detected in inner loop")
+
                 push!(err_evo_V, errVx/errVx00)
                 push!(err_evo_P, errPt/errPt0)
                 push!(err_evo_it, iter)
@@ -271,6 +273,7 @@ function _solve_DYREL!(
                 @. dVy = dVydτ * βVy * dτVy
 
                 # @printf("it = %d, iter = %d, ϵ_vel = %1.3e, err = %1.3e norm[Rx=%1.3e, Ry=%1.3e] \n", itPT, iter, ϵ_vel, err, errVx, errVy)
+                @printf("it = %d, iter = %d, err = %1.3e \n", itPT, iter, err)
 
                 λminV  = abs(sum(dVx .* (stokes.R.Rx .- Rx0)) + sum(dVy.*(stokes.R.Ry .- Ry0))) / 
                     (sum(dVx.^2) + sum(dVy.^2))
[…]
