### 2026-01-07 77e2ced
more miniapps
