+++ ShearBand2D_Duretz2020_DYREL.jl
@@ -1,0 +1,268 @@
+# Setup from Duretz et al 2020 https://agupubs.onlinelibrary.wiley.com/doi/epdf/10.1029/2019GL086027
+
+# using CUDA
+
+using GeoParams, GLMakie, CellArrays
+using JustRelax, JustRelax.JustRelax2D, JustRelax.DataIO
+using ParallelStencil
+# @init_parallel_stencil(CUDA, Float64, 2)
+@init_parallel_stencil(Threads, Float64, 2)
+
+# const backend = CUDABackend
+const backend = CPUBackend
+
+using JustPIC, JustPIC._2D
+import JustPIC._2D.GridGeometryUtils as GGU
+
+const backend_JP = JustPIC.CPUBackend
+# const backend_JP = CUDABackend
+
+# HELPER FUNCTIONS ----------------------------------- ----------------------------
+# Initialize phases on the particles
+function init_phases!(phase_ratios, xci, xvi, circle)
+    ni = size(phase_ratios.center)
+
+    @parallel_indices (i, j) function init_phases!(phases, xc, yc, circle)
+        x, y = xc[i], yc[j]
+        p = GGU.Point(x, y)
+        if GGU.inside(p, circle)
+            @index phases[1, i, j] = 0.0
+            @index phases[2, i, j] = 1.0
+
+        else
+            @index phases[1, i, j] = 1.0
+            @index phases[2, i, j] = 0.0
+
+        end
+        return nothing
+    end
+
+    @parallel (@idx ni) init_phases!(phase_ratios.center, xci..., circle)
+    @parallel (@idx ni .+ 1) init_phases!(phase_ratios.vertex, xvi..., circle)
+    return nothing
+end
+
+const year = 3600 * 24 * 365.25
+
+# MAIN SCRIPT --------------------------------------------------------------------
+function main(igg; nx = 64, ny = 64, figdir = "model_figs")
+
+    # Physical domain ------------------------------------
+    ly = 30e3            # domain length in y
+    lx = 100e3           # domain length in x
+    ni = nx, ny          # number of cells
+    li = lx, ly          # domain length in x- and y-
+    di = @. li / ni      # grid step in x- and -y
+    origin = -lx/2, -ly  # origin coordinates
+    grid = Geometry(ni, li; origin = origin)
+    (; xci, xvi) = grid # nodes at the center and vertices of the cells
+
+    # Physical properties using GeoParams ----------------
+    dt       = 10e3 * year     # assumes Maxwell time of 4
+    el_crust = ConstantElasticity(; G = 1e10, ν = 0.499)
+    el_inc   = ConstantElasticity(; G = 1e10, ν = 0.499)
+    
+    disl_crust = DislocationCreep(A = 3.1623e-26, n = 3.3, E = 186.5e3, V = 0e0, r = 0.0, R = 8.3145)
+    disl_incl  = DislocationCreep(A = 1e-20,      n = 1,   E = 0e0,     V = 0e0, r = 0.0, R = 8.3145)
+    
+    pl_crust = DruckerPrager_regularised(;
+        # non-regularized plasticity
+        C = 50e6,
+        ϕ = atand(0.6),
+        η_vp = 1e21,
+        Ψ = 0
+    )
+        
+    pl_incl  = DruckerPrager_regularised(;
+        # non-regularized plasticity
+        C = 0.1e6,
+        ϕ = 0e0,
+        η_vp = 1e21,
+        Ψ = 0
+    )
+
+    rheology = (
+        # Low density phase
+        SetMaterialParams(;
+            Phase = 1,
+            Density = ConstantDensity(; ρ = 2.7e3),
[…]
