### 2026-01-07 b1e20e9
folder with DYREL miniapps
