### 2026-01-09 d8be152
Update documentation badge link to stable version
