### 2026-01-09 c446280
couple of fixes
