+++ versions.js
@@ -1,0 +1,11 @@
+var DOC_VERSIONS = [
+  "dev",
+  "v0.4.3",
+  "v0.4.2",
+  "v0.4.1",
+  "v0.3.2",
+  "v0.3.1",
+  "v0.2.1",
+  "v0.2.0",
+];
+var DOCUMENTER_NEWEST = "0.4.3";
