+++ dev/assets/man_authors.md.3sYVCgKN.js
@@ -1,0 +1,1 @@
+import{_ as r,c as t,o as a,aA as o}from"./chunks/framework.OnVvzU9V.js";const d=JSON.parse('{"title":"Authors","description":"","frontmatter":{},"headers":[],"relativePath":"man/authors.md","filePath":"man/authors.md","lastUpdated":null}'),i={name:"man/authors.md"};function l(s,e,n,u,p,h){return a(),t("div",null,[...e[0]||(e[0]=[o('<h1 id="Authors" tabindex="-1">Authors <a class="header-anchor" href="#Authors" aria-label="Permalink to &quot;Authors {#Authors}&quot;">​</a></h1><p><a href="https://github.com/PTsolvers/JustRelax.jl" target="_blank" rel="noreferrer">JustRelax.jl</a>&#39;s development is coordinated by a group of <em>principal developers</em>, who are also its main contributors and who can be contacted in case of questions about <a href="https://github.com/PTsolvers/JustRelax.jl" target="_blank" rel="noreferrer">JustRelax.jl</a>. In addition, there are <em>contributors</em> who have provided substantial additions or modifications. Together, these two groups form &quot;The JustRelax.jl Authors&quot;.</p><h2 id="Principal-Developers" tabindex="-1">Principal Developers <a class="header-anchor" href="#Principal-Developers" aria-label="Permalink to &quot;Principal Developers {#Principal-Developers}&quot;">​</a></h2><ul><li><p><a href="https://github.com/albert-de-montserrat" target="_blank" rel="noreferrer">Albert de Montserrat</a>, Institute for Geophysics, ETH Zurich, Switzerland</p></li><li><p><a href="https://github.com/aelligp" target="_blank" rel="noreferrer">Pascal Aellig</a>, Institute for Geosciences, Johannes Gutenberg University Mainz, Germany</p></li></ul><h2 id="Contributors" tabindex="-1">Contributors <a class="header-anchor" href="#Contributors" aria-label="Permalink to &quot;Contributors {#Contributors}&quot;">​</a></h2><p>The following people contributed major additions or modifications to <a href="https://github.com/PTsolvers/JustRelax.jl" target="_blank" rel="noreferrer">JustRelax.jl</a> and are listed in alphabetical order:</p><ul><li><p>Sidney Beeler</p></li><li><p>Lukas Fuchs</p></li><li><p>Boris Kaus</p></li><li><p>Patrick Sanan</p></li><li><p>Hendrik Ranocha</p></li><li><p>Ludovic Räss</p></li></ul>',7)])])}const b=r(i,[["render",l]]);export{d as __pageData,b as default};
