+++ dev/assets/chunks/@localSearchIndexroot.C8bSyRle.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":153,"nextId":153,"documentIds":{"0":"/JustRelax.jl/dev/#What-is-JustRelax.jl?","1":"/JustRelax.jl/dev/#How-to-Install-JustRelax.jl?","2":"/JustRelax.jl/dev/#Example-miniapps","3":"/JustRelax.jl/dev/#Funding","4":"/JustRelax.jl/dev/man/Blankenbach#Blankenbach-benchmark","5":"/JustRelax.jl/dev/man/Blankenbach#Initialize-packages","6":"/JustRelax.jl/dev/man/Blankenbach#Script","7":"/JustRelax.jl/dev/man/Blankenbach#Model-domain","8":"/JustRelax.jl/dev/man/Blankenbach#Rheology","9":"/JustRelax.jl/dev/man/Blankenbach#Initialize-particles","10":"/JustRelax.jl/dev/man/Blankenbach#Temperature-anomaly","11":"/JustRelax.jl/dev/man/Blankenbach#Stokes-and-heat-diffusion-arrays","12":"/JustRelax.jl/dev/man/Blankenbach#Initialize-thermal-profile-and-viscosity-fields","13":"/JustRelax.jl/dev/man/Blankenbach#Boundary-conditions","14":"/JustRelax.jl/dev/man/Blankenbach#Pseuo-transient-coefficients","15":"/JustRelax.jl/dev/man/Blankenbach#Just-before-solving-the-problem...","16":"/JustRelax.jl/dev/man/Blankenbach#Advancing-one-time-step","17":"/JustRelax.jl/dev/man/Blankenbach#Visualization","18":"/JustRelax.jl/dev/man/Blankenbach#Fields","19":"/JustRelax.jl/dev/man/Blankenbach#Final-model","20":"/JustRelax.jl/dev/man/GeoParams#List-of-GeoParams-functions","21":"/JustRelax.jl/dev/man/JustPIC#List-of-JustPIC-functions","22":"/JustRelax.jl/dev/man/ShearBands#ShearBand-benchmark","23":"/JustRelax.jl/dev/man/ShearBands#Initialize-packages","24":"/JustRelax.jl/dev/man/ShearBands#Script","25":"/JustRelax.jl/dev/man/ShearBands#Model-domain","26":"/JustRelax.jl/dev/man/ShearBands#Physical-properties-using-GeoParams","27":"/JustRelax.jl/dev/man/ShearBands#Rheology","28":"/JustRelax.jl/dev/man/ShearBands#Phase-anomaly","29":"/JustRelax.jl/dev/man/ShearBands#Stokes-arrays","30":"/JustRelax.jl/dev/man/ShearBands#Pseuo-transient-coefficients","31":"/JustRelax.jl/dev/man/ShearBands#Initialize-viscosity-fields","32":"/JustRelax.jl/dev/man/ShearBands#Boundary-conditions","33":"/JustRelax.jl/dev/man/ShearBands#Just-before-solving-the-problem...","34":"/JustRelax.jl/dev/man/ShearBands#Solving-Stokes","35":"/JustRelax.jl/dev/man/ShearBands#Visualization","36":"/JustRelax.jl/dev/man/ShearBands#Fields","37":"/JustRelax.jl/dev/man/ShearBands#Final-model","38":"/JustRelax.jl/dev/man/advection#Field-advection","39":"/JustRelax.jl/dev/man/advection#Particles-in-Cell","40":"/JustRelax.jl/dev/man/authors#Authors","41":"/JustRelax.jl/dev/man/authors#Principal-Developers","42":"/JustRelax.jl/dev/man/authors#Contributors","43":"/JustRelax.jl/dev/man/backend#S
[…]
