+++ docs/paper/paper.bib
@@ -258,7 +258,7 @@
   doi={10.1109/TPDS.2018.2872064}
 }
 
-@article{MPI.jl,
+@article{@Byrne2021,
   doi = {10.21105/jcon.00068},
   url = {https://doi.org/10.21105/jcon.00068},
   year = {2021},
