### 2026-01-11 PTsolvers/JustRelax.jl#381 @albert-de-montserrat
Thanks for reporting it! The references are now fixed (see [here](https://github.com/openjournals/joss-papers/blob/joss.09365/joss.09365/10.21105.joss.09365.pdf))
