### 2026-01-12 #9365 @willgearty
Fantastic, I can now move forward with my final checks before recommending for acceptance!

@albert-de-montserrat could you please perform the tasks listed above under "Additional Author Tasks After Review is Complete"?
