### 2026-01-12 #9365 @willgearty
I noticed the following issues @albert-de-montserrat :
1. The title of the [Zenodo repository](https://zenodo.org/records/18224646) does not match the title of the JOSS paper. Please manually edit the Zenodo metadata accordingly.
2. The last entry in the BibTeX file is a duplicate and also has an invalid key. I believe the entry can safely be deleted.
