+++ docs/paper/paper.bib
@@ -43,6 +43,19 @@
   journal = {Proceedings of the JuliaCon Conferences} 
 } 
 
+@article{@Byrne2021,
+  doi = {10.21105/jcon.00068},
+  url = {https://doi.org/10.21105/jcon.00068},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {1},
+  number = {1},
+  pages = {68},
+  author = {Simon Byrne and Lucas C. Wilcox and Valentin Churavy},
+  title = {MPI.jl: Julia bindings for the Message Passing Interface},
+  journal = {Proceedings of the JuliaCon Conferences}
+}
+
 @article{besard2018juliagpu,
   author        = {Besard, Tim and Foket, Christophe and De Sutter, Bjorn},
   title         = {Effective Extensible Programming: Unleashing {Julia} on {GPUs}},
@@ -254,19 +267,5 @@
   volume={30},
   number={4},
   pages={827-841},
-  keywords={Graphics processing units;Programming;Hardware;High level languages;Libraries;Graphics processors;very high-level languages;code generation;retargetable compilers},
   doi={10.1109/TPDS.2018.2872064}
-}
-
-@article{@Byrne2021,
-  doi = {10.21105/jcon.00068},
-  url = {https://doi.org/10.21105/jcon.00068},
-  year = {2021},
-  publisher = {The Open Journal},
-  volume = {1},
-  number = {1},
-  pages = {68},
-  author = {Simon Byrne and Lucas C. Wilcox and Valentin Churavy},
-  title = {MPI.jl: Julia bindings for the Message Passing Interface},
-  journal = {Proceedings of the JuliaCon Conferences}
 }
