### 2026-01-12 1a0a556
fix bib syntax
