+++ docs/paper/paper.bib
@@ -101,7 +101,8 @@
   url = {https://doi.org/10.21105/jcon.00137},
   year = {2024},
   publisher = {The Open Journal},
-  volume = {6}, number = {65},
+  volume = {6}, 
+  number = {65},
   pages = {137},
   author = {Samuel Omlin and Ludovic Räss and Ivan Utkin},
   title = {Distributed Parallelization of xPU Stencil Computations in Julia},
@@ -171,29 +172,29 @@
 }
 
 @InProceedings{TerraNeo,
-  author="Bauer, Simon
-  and Bunge, Hans-Peter
-  and Drzisga, Daniel
-  and Ghelichkhan, Siavash
-  and Huber, Markus
-  and Kohl, Nils
-  and Mohr, Marcus
-  and R{\"u}de, Ulrich
-  and Th{\"o}nnes, Dominik
-  and Wohlmuth, Barbara",
-  editor="Bungartz, Hans-Joachim
-  and Reiz, Severin
-  and Uekermann, Benjamin
-  and Neumann, Philipp
-  and Nagel, Wolfgang E.",
-  title="TerraNeo---Mantle Convection Beyond a Trillion Degrees of Freedom",
-  booktitle="Software for Exascale Computing - SPPEXA 2016-2019",
-  year="2020",
-  publisher="Springer International Publishing",
-  address="Cham",
-  pages="569--610",
-  abstract="Simulation of mantle convection on planetary scales is considered a grand-challenge application even in the exascale era. The reason being the enormous spatial and temporal scales that must be resolved in the computation as well as the complexities of realistic models and the large parameter uncertainties that need to be handled by advanced numerical methods. This contribution reports on the TerraNeo project which delivered novel matrix-free geometric multigrid solvers for the Stokes system that forms the core of mantle convection models. In TerraNeo the hierarchical hybrid grids paradigm was employed to demonstrate that scalability can be achieved when solving the Stokes system with more than ten trillion (1.1{\thinspace}⋅{\thinspace}1013) degrees of freedom even on present-day peta-scale supercomputers. Novel concepts were developed to ensure resilience of algorithms even in case of hard faults and new scheduling algorithms proposed for ensemble runs arising in Multilevel Monte Carlo algorithms for uncertainty quantification. The prototype framework was used to investigate geodynamic questions such as high velocity asthenospheric channels and dynamic topography and to perform adjoint inversions. We also describe the redesign of our software to support more advanced discretizations, adaptivity, and highly asynchronous execution while ensuring sustainability and flexibility for future extensions.",
-  isbn="978-3-030-47956-5"
+  author={Bauer, Simon,
+  and Bunge, Hans-Peter,
+  and Drzisga, Daniel,
+  and Ghelichkhan, Siavash,
+  and Huber, Markus,
+  and Kohl, Nils,
+  and Mohr, Marcus,
+  and R{\"u}de, Ulrich,
+  and Th{\"o}nnes, Dominik,
+  and Wohlmuth, Barbara},
+  editor={Bungartz, Hans-Joachim,
+  and Reiz, Severin,
+  and Uekermann, Benjamin,
+  and Neumann, Philipp,
+  and Nagel, Wolfgang E.},
+  title={TerraNeo---Mantle Convection Beyond a Trillion Degrees of Freedom},
+  booktitle={Software for Exascale Computing - SPPEXA 2016-2019},
+  year={2020},
+  publisher={Springer International Publishing},
+  address={Cham},
+  pages={569--610},
[…]
