+++ docs/paper/paper.bib
@@ -30,20 +30,7 @@
   year = {2023},
   URL = {https://doi.org/10.5281/zenodo.7804590}
 }
-@article{Byrne2021, 
-  doi = {10.21105/jcon.00068}, 
-  url = {https://doi.org/10.21105/jcon.00068}, 
-  year = {2021}, 
-  publisher = {The Open Journal}, 
-  volume = {1}, 
-  number = {1}, 
-  pages = {68}, 
-  author = {Simon Byrne and Lucas C. Wilcox and Valentin Churavy}, 
-  title = {MPI.jl: Julia bindings for the Message Passing Interface}, 
-  journal = {Proceedings of the JuliaCon Conferences} 
-} 
-
-@article{@Byrne2021,
+@article{Byrne2021,
   doi = {10.21105/jcon.00068},
   url = {https://doi.org/10.21105/jcon.00068},
   year = {2021},
@@ -68,60 +55,17 @@
   primaryClass  = {cs.PL},
 }
 
-@article{larionov2017variational,
-  author = {Larionov, Egor and Batty, Christopher and Bridson, Robert}, 
-  title = {Variational stokes: a unified pressure-viscosity solver for accurate viscous liquids}, 
-  year = {2017}, 
-  issue_date = {August 2017}, 
-  publisher = {Association for Computing Machinery}, 
-  address = {New York, NY, USA}, 
-  volume = {36}, 
-  number = {4},  
-  issn = {0730-0301}, 
-  url = {https://doi.org/10.1145/3072959.3073628}, 
-  doi = {10.1145/3072959.3073628},
-}
-
-@article{Harlow1965,
-    title = {{Numerical Calculation of Time‐Dependent Viscous Incompressible Flow of Fluid with Free Surface}},
-    year = {1965},
-    journal = {The Physics of Fluids},
-    author = {Harlow, Francis H. and Welch, J. Eddie},
-    number = {12},
-    month = {12},
-    pages = {2182--2189},
-    volume = {8},
-    publisher = {AIP Publishing},
-    doi = {10.1063/1.1761178},
-    issn = {0031-9171}
-}
-
 @article{Omlin2024,
   doi = {10.21105/jcon.00137},
   url = {https://doi.org/10.21105/jcon.00137},
   year = {2024},
   publisher = {The Open Journal},
-  volume = {6}, 
+  volume = {6},
   number = {65},
   pages = {137},
   author = {Samuel Omlin and Ludovic Räss and Ivan Utkin},
   title = {Distributed Parallelization of xPU Stencil Computations in Julia},
   journal = {Proceedings of the JuliaCon Conferences}
-}
-
-@article{i3elvis,
-  title={Robust characteristics method for modelling multiphase visco-elasto-plastic thermo-mechanical problems},
-  author={Gerya, Taras V and Yuen, David A},
-  journal={Physics of the Earth and Planetary Interiors},
-  volume={163},
-  number={1-4},
-  pages={83--105},
-  year={2007},
-  note = {Computational Challenges in the Earth Sciences},
-  issn = {0031-9201},
-  doi = {https://doi.org/10.1016/j.pepi.2007.04.015},
-  url = {https://www.sciencedirect.com/science/article/pii/S0031920107000969},
-  publisher={Elsevier}
 }
 
 @article{gerya2015plate,
@@ -197,12 +141,6 @@
   isbn={978-3-030-47956-5}
 }
 
-@article{lamem2,
-  title={Scalable hybrid multigrid for staggered grid discretizations in geodynamics},
-  author={Popov, A. A., Kaus, B. J. P., and Ibragimov, I.},
-  doi = {doi.org/10.5194/egusphere-egu25-11825},
[…]
