### 2026-01-13 c497367
fix DOIs
