+++ docs/paper/paper.bib
@@ -14,6 +14,7 @@
   author = {Kaus, Boris and de Montserrat, Albert and Medinger, Nils and Aellig, Pascal and Dominguez, Hugo and Riel, Nicolas and Cosarinsky, Matias and Spang, Arne and Berlie, Nicolas and Kiss, Daniel and Ranocha, Hendrik and Fuchs, Lukas and Frasunkiewicz, Jacob and Räss, Ludovic and Lueder, Mona and Seiler, Alexander and Duretz, Thibault},
   title = {GeoParams.jl: v0.7.4},
   year = {2025},
+  doi = {10.5281/zenodo.8089230},
   URL = {https://doi.org/10.5281/zenodo.15719680}
 }
 
@@ -21,6 +22,7 @@
   author = {de Montserrat, Albert and Aellig, Pascal and Räss, Ludovic and Duretz, Thibault and Utkin, Ivan and Kaus, Boris and Ranocha, Hendrik},
   title = {JustPIC.jl: v0.5.10},
   year = {2025},
+  doi = {10.5281/zenodo.10212675},
   URL = {https://doi.org/10.5281/zenodo.15704602}
 }
 
@@ -28,6 +30,7 @@
   author = {Polanco, J. I.},
   title = {WriteVTK.jl: a Julia package for writing VTK XML files (v1.18.0)},
   year = {2023},
+  doi = {doi.org/10.5281/zenodo.7804590},
   URL = {https://doi.org/10.5281/zenodo.7804590}
 }
 @article{Byrne2021,
@@ -114,6 +117,15 @@
   year={2016},
   publisher={John von Neumann Institute for Computing (NIC), NIC Series}
 }
+@software{lamem,
+  author = {Popov, Anton A. and Kaus, Boris J. P.},
+  title = {LaMEM - Lithosphere and Mantle Evolution Model},
+  version = {2.2.0},
+  date = {2025-01-15},
+  publisher = {Zenodo},
+  doi = {10.5281/zenodo.15524303},
+  url = {https://doi.org/10.5281/zenodo.15524303}
+}
 
 @InProceedings{TerraNeo,
   author={Bauer, Simon,
@@ -137,8 +149,8 @@
   publisher={Springer International Publishing},
   address={Cham},
   pages={569--610},
-  abstract={Simulation of mantle convection on planetary scales is considered a grand-challenge application even in the exascale era. The reason being the enormous spatial and temporal scales that must be resolved in the computation as well as the complexities of realistic models and the large parameter uncertainties that need to be handled by advanced numerical methods. This contribution reports on the TerraNeo project which delivered novel matrix-free geometric multigrid solvers for the Stokes system that forms the core of mantle convection models. In TerraNeo the hierarchical hybrid grids paradigm was employed to demonstrate that scalability can be achieved when solving the Stokes system with more than ten trillion (1.1{\thinspace}⋅{\thinspace}1013) degrees of freedom even on present-day peta-scale supercomputers. Novel concepts were developed to ensure resilience of algorithms even in case of hard faults and new scheduling algorithms proposed for ensemble runs arising in Multilevel Monte Carlo algorithms for uncertainty quantification. The prototype framework was used to investigate geodynamic questions such as high velocity asthenospheric channels and dynamic topography and to perform adjoint inversions. We also describe the redesign of our software to support more advanced discretizations, adaptivity, and highly asynchronous execution while ensuring sustainability and flexibility for future extensions.},
-  isbn={978-3-030-47956-5}
+  isbn={978-3-030-47956-5},
+  doi={10.1007/978-3-030-47956-5_19},
 }
 
 @article{citcom,
[…]
