+++ docs/paper/paper.bib
@@ -30,7 +30,7 @@
   author = {Polanco, J. I.},
   title = {WriteVTK.jl: a Julia package for writing VTK XML files (v1.18.0)},
   year = {2023},
-  doi = {doi.org/10.5281/zenodo.7804590},
+  doi = {10.5281/zenodo.7804590},
   URL = {https://doi.org/10.5281/zenodo.7804590}
 }
 @article{Byrne2021,
