+++ docs/paper/paper.bib
@@ -108,15 +108,6 @@
   doi = {10.1016/j.pepi.2008.08.005},
   publisher={Elsevier}
 }
-@article{lamem,
-  title={Forward and inverse modelling of lithospheric deformation on geological timescales},
-  author={Kaus, Boris JP and Popov, Anton A and Baumann, T and Pusok, A and Bauville, Arthur and Fernandez, Naiara and Collignon, Marine},
-  journal={Proceedings of nic symposium},
-  volume={48},
-  pages={978--983},
-  year={2016},
-  publisher={John von Neumann Institute for Computing (NIC), NIC Series}
-}
 @software{lamem,
   author = {Popov, Anton A. and Kaus, Boris J. P.},
   title = {LaMEM - Lithosphere and Mantle Evolution Model},
