### 2026-01-13 #9365 @albert-de-montserrat
Apologies for the spam, references are now all sorted.

@willgearty do I need to make a new version release and Zenodo upload due to the changes done to fix the `.bib` file?
