+++ docs/paper/paper.bib
@@ -14,6 +14,7 @@
   author = {Kaus, Boris and de Montserrat, Albert and Medinger, Nils and Aellig, Pascal and Dominguez, Hugo and Riel, Nicolas and Cosarinsky, Matias and Spang, Arne and Berlie, Nicolas and Kiss, Daniel and Ranocha, Hendrik and Fuchs, Lukas and Frasunkiewicz, Jacob and Räss, Ludovic and Lueder, Mona and Seiler, Alexander and Duretz, Thibault},
   title = {GeoParams.jl: v0.7.4},
   year = {2025},
+  doi = {10.5281/zenodo.8089230},
   URL = {https://doi.org/10.5281/zenodo.15719680}
 }
 
@@ -21,6 +22,7 @@
   author = {de Montserrat, Albert and Aellig, Pascal and Räss, Ludovic and Duretz, Thibault and Utkin, Ivan and Kaus, Boris and Ranocha, Hendrik},
   title = {JustPIC.jl: v0.5.10},
   year = {2025},
+  doi = {10.5281/zenodo.10212675},
   URL = {https://doi.org/10.5281/zenodo.15704602}
 }
 
@@ -28,20 +30,21 @@
   author = {Polanco, J. I.},
   title = {WriteVTK.jl: a Julia package for writing VTK XML files (v1.18.0)},
   year = {2023},
+  doi = {10.5281/zenodo.7804590},
   URL = {https://doi.org/10.5281/zenodo.7804590}
 }
-@article{Byrne2021, 
-  doi = {10.21105/jcon.00068}, 
-  url = {https://doi.org/10.21105/jcon.00068}, 
-  year = {2021}, 
-  publisher = {The Open Journal}, 
-  volume = {1}, 
-  number = {1}, 
-  pages = {68}, 
-  author = {Simon Byrne and Lucas C. Wilcox and Valentin Churavy}, 
-  title = {MPI.jl: Julia bindings for the Message Passing Interface}, 
-  journal = {Proceedings of the JuliaCon Conferences} 
-} 
+@article{Byrne2021,
+  doi = {10.21105/jcon.00068},
+  url = {https://doi.org/10.21105/jcon.00068},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {1},
+  number = {1},
+  pages = {68},
+  author = {Simon Byrne and Lucas C. Wilcox and Valentin Churavy},
+  title = {MPI.jl: Julia bindings for the Message Passing Interface},
+  journal = {Proceedings of the JuliaCon Conferences}
+}
 
 @article{besard2018juliagpu,
   author        = {Besard, Tim and Foket, Christophe and De Sutter, Bjorn},
@@ -55,59 +58,17 @@
   primaryClass  = {cs.PL},
 }
 
-@article{larionov2017variational,
-  author = {Larionov, Egor and Batty, Christopher and Bridson, Robert}, 
-  title = {Variational stokes: a unified pressure-viscosity solver for accurate viscous liquids}, 
-  year = {2017}, 
-  issue_date = {August 2017}, 
-  publisher = {Association for Computing Machinery}, 
-  address = {New York, NY, USA}, 
-  volume = {36}, 
-  number = {4},  
-  issn = {0730-0301}, 
-  url = {https://doi.org/10.1145/3072959.3073628}, 
-  doi = {10.1145/3072959.3073628},
-}
-
-@article{Harlow1965,
-    title = {{Numerical Calculation of Time‐Dependent Viscous Incompressible Flow of Fluid with Free Surface}},
-    year = {1965},
-    journal = {The Physics of Fluids},
-    author = {Harlow, Francis H. and Welch, J. Eddie},
[…]
