+++ Project.toml
@@ -1,6 +1,6 @@
 name = "JustRelax"
 uuid = "34418575-392d-4e26-8c6d-96b0910afa06"
-version = "0.4.4"
+version = "0.4.5"
 authors = ["Albert De Montserrat <albertdemontserratnavarro@erdw.ethz.ch>, Pascal Aellig <paellig@uni-mainz.de>"]
 
 [deps]
