+++ dev/assets/man_JustPIC.md.DrnJScjC.lean.js
@@ -1,0 +1,1 @@
+import{_ as o,C as p,c as r,o as d,j as e,a as l,aA as a,G as t,w as n}from"./chunks/framework.OnVvzU9V.js";const ti=JSON.parse('{"title":"List of JustPIC functions","description":"","frontmatter":{},"headers":[],"relativePath":"man/JustPIC.md","filePath":"man/JustPIC.md","lastUpdated":null}'),c={name:"man/JustPIC.md"},h={class:"jldocstring custom-block",open:""},u={class:"jldocstring custom-block",open:""},k={class:"jldocstring custom-block",open:""},g={class:"jldocstring custom-block",open:""},T={class:"jldocstring custom-block",open:""},y={class:"jldocstring custom-block",open:""},_={class:"jldocstring custom-block",open:""},b={class:"jldocstring custom-block",open:""},m={class:"jldocstring custom-block",open:""},C={class:"jldocstring custom-block",open:""},f={class:"jldocstring custom-block",open:""},v={class:"jldocstring custom-block",open:""},E={class:"jldocstring custom-block",open:""},A={class:"jldocstring custom-block",open:""},N={class:"jldocstring custom-block",open:""},j={class:"jldocstring custom-block",open:""},P={class:"jldocstring custom-block",open:""},I={class:"jldocstring custom-block",open:""},F={class:"jldocstring custom-block",open:""},D={class:"jldocstring custom-block",open:""},J={class:"jldocstring custom-block",open:""},x={class:"jldocstring custom-block",open:""},V={class:"jldocstring custom-block",open:""},B={class:"jldocstring custom-block",open:""},S={class:"jldocstring custom-block",open:""},w={class:"jldocstring custom-block",open:""},M={class:"jldocstring custom-block",open:""},L={class:"jldocstring custom-block",open:""},R={class:"jldocstring custom-block",open:""},U={class:"jldocstring custom-block",open:""},G={class:"jldocstring custom-block",open:""},O={class:"jldocstring custom-block",open:""},Q={class:"jldocstring custom-block",open:""},z={class:"jldocstring custom-block",open:""},K={class:"jldocstring custom-block",open:""},$={class:"jldocstring custom-block",open:""},H={class:"jldocstring custom-block",open:""},q={class:"jldocstring custom-block",open:""};function W(X,i,Y,Z,ii,ei){const s=p("Badge");return d(),r("div",null,[i[152]||(i[152]=e("h1",{id:"List-of-JustPIC-functions",tabindex:"-1"},[l("List of JustPIC functions "),e("a",{class:"header-anchor",href:"#List-of-JustPIC-functions","aria-label":'Permalink to "List of JustPIC functions {#List-of-JustPIC-functions}"'},"​")],-1)),i[153]||(i[153]=e("p",null,[l("Here an overview of all functions in JustPIC.jl, , for a complete list see "),e("a",{href:"https://juliageodynamics.github.io/JustPIC.jl/dev/",target:"_blank",rel:"nor
[…]
