+++ previews/PR419/assets/man_GeoParams.md.D4Dgfny6.lean.js
@@ -1,0 +1,1 @@
+import{_ as Q,C as h,c as n,o as T,j as s,a,aA as i,G as l,w as r}from"./chunks/framework.4wj8HieU.js";const w4=JSON.parse('{"title":"List of GeoParams functions","description":"","frontmatter":{},"headers":[],"relativePath":"man/GeoParams.md","filePath":"man/GeoParams.md","lastUpdated":null}'),p={name:"man/GeoParams.md"},o={class:"jldocstring custom-block",open:""},d={class:"jldocstring custom-block",open:""},k={class:"jldocstring custom-block",open:""},m={class:"jldocstring custom-block",open:""},g={class:"jldocstring custom-block",open:""},y={class:"jldocstring custom-block",open:""},u={class:"jldocstring custom-block",open:""},c={class:"jldocstring custom-block",open:""},E={class:"jldocstring custom-block",open:""},x={class:"jldocstring custom-block",open:""},w={class:"jldocstring custom-block",open:""},C={class:"jldocstring custom-block",open:""},b={class:"jldocstring custom-block",open:""},f={class:"jldocstring custom-block",open:""},F={class:"jldocstring custom-block",open:""},H={class:"jldocstring custom-block",open:""},_={class:"jldocstring custom-block",open:""},v={class:"jldocstring custom-block",open:""},D={class:"jldocstring custom-block",open:""},A={class:"jldocstring custom-block",open:""},L={class:"jldocstring custom-block",open:""},M={class:"jldocstring custom-block",open:""},j={class:"jldocstring custom-block",open:""},V={class:"jldocstring custom-block",open:""},P={class:"jldocstring custom-block",open:""},Z={class:"jldocstring custom-block",open:""},B={class:"jldocstring custom-block",open:""},S={class:"jldocstring custom-block",open:""},N={class:"jldocstring custom-block",open:""},I={class:"jldocstring custom-block",open:""},G={class:"jldocstring custom-block",open:""},R={class:"jldocstring custom-block",open:""},O={class:"jldocstring custom-block",open:""},J={class:"jldocstring custom-block",open:""},z={class:"jldocstring custom-block",open:""},U={class:"jldocstring custom-block",open:""},X={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},q={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"0"},xmlns:"http://www.w3.org/2000/svg",width:"3.832ex",height:"2.021ex",role:"img",focusable:"false",viewBox:"0 -893.3 1693.8 893.3","aria-hidden":"true"},K={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},W={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-3.531ex"},xmlns:"http://www.w3.org/2000/svg",width:"19.885ex",height:"6.567ex",role:"img",focusable:"false",viewBox:"0 -1342 8789.2 2902.7","aria-hidden":"true"},$={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},Y={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-3.119ex"},xmlns:"http://www.w3.org/2000/svg"
[…]

[…]
