+++ stable/assets/man_restart.md.Cnw8cdTq.js
@@ -8,7 +8,7 @@ import{_ as s,c as a,o as t,aA as e}from"./chunks/framework.DLzZTck7.js";const c
 <span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">particle_args </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> TA</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(backend).(Float64, data[</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;">&quot;particle_args&quot;</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">])</span></span>
 <span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">subgrid_arrays  </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> SubgridDiffusionCellArrays</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(particles)</span></span>
 <span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># velocity staggered grids</span></span>
-<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">grid_vxi        </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> velocity_grids</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(xci, xvi, di)</span></span></code></pre></div><h2 id="Load-Stokes-and-Thermal-arrays-from-checkpoint-file" tabindex="-1">Load Stokes and Thermal arrays from checkpoint file <a class="header-anchor" href="#Load-Stokes-and-Thermal-arrays-from-checkpoint-file" aria-label="Permalink to &quot;Load Stokes and Thermal arrays from checkpoint file {#Load-Stokes-and-Thermal-arrays-from-checkpoint-file}&quot;">​</a></h2><div class="vp-code-group vp-adaptive-theme"><div class="tabs"><input type="radio" name="group-OsZ5w" id="tab-6vRFN_l" checked><label data-title="Normal use" for="tab-6vRFN_l">Normal use</label><input type="radio" name="group-OsZ5w" id="tab-445m3CI"><label data-title="MPI" for="tab-445m3CI">MPI</label><input type="radio" name="group-OsZ5w" id="tab-5F5hPIX"><label data-title="Additional fields" for="tab-5F5hPIX">Additional fields</label></div><div class="blocks"><div class="language-julia vp-adaptive-theme active"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">dst </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> &quot;Your_checkpointing_directory&quot;</span></span>
[…]
