+++ v0.4/assets/chunks/VPLocalSearchBox.mau1sOXZ.js
@@ -1,4 +1,4 @@
-var Ot=Object.defineProperty;var At=(a,e,t)=>e in a?Ot(a,e,{enumerable:!0,configurable:!0,writable:!0,value:t}):a[e]=t;var Me=(a,e,t)=>At(a,typeof e!="symbol"?e+"":e,t);import{V as Ct,D as ce,h as be,aB as nt,aC as Mt,aD as Lt,aE as Dt,q as Be,aF as Pt,d as zt,aG as rt,p as he,aH as Vt,ac as jt,s as $t,aI as Bt,v as Le,P as ve,O as xe,aJ as Wt,ak as Kt,W as Jt,R as qt,$ as Ut,b as Ht,o as q,j as T,a0 as Gt,a2 as Qt,k as P,a5 as Yt,ad as Zt,c as Q,e as _e,n as st,B as it,F as at,a as pe,t as me,aa as Xt,aK as ot,aL as en,aq as tn,av as nn,aM as rn,_ as sn}from"./framework.B3_v1NGG.js";import{u as an,c as on}from"./theme.5GJJglph.js";const ln={root:()=>Ct(()=>import("./@localSearchIndexroot.DLEX_mcK.js"),[])};/*!
+var Ot=Object.defineProperty;var At=(a,e,t)=>e in a?Ot(a,e,{enumerable:!0,configurable:!0,writable:!0,value:t}):a[e]=t;var Me=(a,e,t)=>At(a,typeof e!="symbol"?e+"":e,t);import{V as Ct,D as ce,h as be,aB as nt,aC as Mt,aD as Lt,aE as Dt,q as Be,aF as Pt,d as zt,aG as rt,p as he,aH as Vt,ac as jt,s as $t,aI as Bt,v as Le,P as ve,O as xe,aJ as Wt,ak as Kt,W as Jt,R as qt,$ as Ut,b as Ht,o as q,j as T,a0 as Gt,a2 as Qt,k as P,a5 as Yt,ad as Zt,c as Q,e as _e,n as st,B as it,F as at,a as pe,t as me,aa as Xt,aK as ot,aL as en,aq as tn,av as nn,aM as rn,_ as sn}from"./framework.B3_v1NGG.js";import{u as an,c as on}from"./theme.DavLe2Jd.js";const ln={root:()=>Ct(()=>import("./@localSearchIndexroot.Cgg46N-Z.js"),[])};/*!
 * tabbable 6.4.0
 * @license MIT, https://github.com/focus-trap/tabbable/blob/master/LICENSE
[…]
