### 2026-01-16 #9365 @willgearty
@albert-de-montserrat looks like there's an issue with using `$\partial$` within the hyperlink in the acknowledgments:
https://github.com/PTsolvers/JustRelax.jl/blob/main/docs/paper/paper.md?plain=1#L95

I guess it generates invalid XML during the rendering process?
https://github.com/openjournals/joss-papers/blob/f9e1d49e92d3925ee20504bcf98f7463feba558b/joss.09365/paper.jats/10.21105.joss.09365.jats#L385-L390

This is a bit beyond my XML/LaTeX understanding. Perhaps you could change it to something like this?
`GPU4GEO & $\partial$GPU4GEO [projects](https://gpu4geo.org/)`

You could also try copying and pasting the unicode character ∂ into the markdown file?
