### 2026-01-16 #9365 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
