### 2026-01-19 e228154
fix grid generation when MPI is not used
