### 2026-01-19 #9365 @kthyng
@albert-de-montserrat A few paper comments:
* Would be a bit nicer to reference the figure with something like "Figure 1" instead of "1".
* What does "(in)compressible" mean? Do you solve incompressible or compressible Stokes equations? If you mean it to represent both, I would have thought "compressible" would include the "incompressible" case but is that not the case?
* One reference has "julia" lowercase but should it be capitalized? You can preserve capitalization by placing {} around characters/words in your .bib file.
