+++ docs/paper/paper.bib
@@ -67,7 +67,7 @@
   number = {65},
   pages = {137},
   author = {Samuel Omlin and Ludovic Räss and Ivan Utkin},
-  title = {Distributed Parallelization of xPU Stencil Computations in Julia},
+  title = {Distributed Parallelization of xPU Stencil Computations in {J}ulia},
   journal = {Proceedings of the JuliaCon Conferences}
 }
 
