### 2026-01-20 8c449f2
sync paper.md changes
