+++ docs/paper/paper.md
@@ -75,7 +75,7 @@ In short, `JustRelax.jl` delivers a high‑performance, portable alternative to 
 
 `JustRelax.jl` features:
 
-- **High-performance and scalable matrix-free solver**: `JustRelax.jl` implements the APT method for compressible Stokes and diffusion problems to circumvent the need for computationally expensive linear algebra operations and direct solvers, significantly improving computational efficiency for large-scale simulations. The embarrassingly parallel nature of the APT method makes it an excellent solver to exploit hardware accelerators. The weak scaling curve of the 3D Stokes solver is shown in Fig. \ref{fig:scaling}, where the parallel efficiency is the wall-time of any multi-process simulation normalized by the wall-time of a single-process simulation ($t_{\text{parallel}}/t_{\text{series}}$). Distributed parallelism across multiple CPU/GPU nodes is achieved with [ImplicitGlobalGrid.jl](https://github.com/eth-cscs/ImplicitGlobalGrid.jl) [@Omlin2024].
+- **High-performance and scalable matrix-free solver**: `JustRelax.jl` implements the APT method for compressible Stokes and diffusion problems to circumvent the need for computationally expensive linear algebra operations and direct solvers, significantly improving computational efficiency for large-scale simulations. The embarrassingly parallel nature of the APT method makes it an excellent solver to exploit hardware accelerators. The weak scaling curve of the 3D Stokes solver is shown in Figure \ref{fig:scaling}, where the parallel efficiency is the wall-time of any multi-process simulation normalized by the wall-time of a single-process simulation ($t_{\text{parallel}}/t_{\text{series}}$). Distributed parallelism across multiple CPU/GPU nodes is achieved with [ImplicitGlobalGrid.jl](https://github.com/eth-cscs/ImplicitGlobalGrid.jl) [@Omlin2024].
 
 ![GPU weak scaling performance of `JustRelax.jl` of the three-dimensional backend, demonstrating efficient parallelization and scalability of the Stokes solver. \label{fig:scaling}](figs/efficiency_Stokes3D.png){width=50%}
 
@@ -87,7 +87,7 @@ In short, `JustRelax.jl` delivers a high‑performance, portable alternative to 
   
 # Examples
 
-An extensive set of benchmarks and model examples are available in the GitHub repository of [`JustRelax.jl`](https://github.com/PTsolvers/JustRelax.jl). Some examples such as [shear band localization](https://ptsolvers.github.io/JustRelax.jl/dev/man/ShearBands), [2D subduction](https://ptsolvers.github.io/JustRelax.jl/dev/man/subduction2D/subduction2D), or the rise of a [3D plume](https://ptsolvers.github.io/JustRelax.jl/dev/man/plume3D/plume3D), are described in the [documentation](https://ptsolvers.github.io/JustRelax.jl/dev/). Here, we limit ourselves to showing some snapshots of the results of these examples in Fig. \ref{fig:examples}.
+An extensive set of benchmarks and model examples are available in the GitHub repository of [`JustRelax.jl`](https://github.com/PTsolvers/JustRelax.jl). Some examples such as [shear band localization](https://ptsolvers.github.io/JustRelax.jl/dev/man/ShearBands), [2D subduction](https://ptsolvers.github.io/JustRelax.jl/dev/man/subduction2D/subduction2D), or the rise of a [3D plume](https://ptsolvers.github.io/JustRelax.jl/dev/man/plume3D/plume3D), are described in the [documentation](https://ptsolvers.github.io/JustRelax.jl/dev/). Here, we limit ourselves to showing some snapshots of the results of these examples in Figure \ref{fig:examples}.
 
[…]
