### 2026-01-20 #9365 @albert-de-montserrat
@kthyng Thanks for the comments:

1. Thanks for spotting it, it now reads "Figure 1". The references to the other figures now also read as "Figure X" instead of "Fig. X".
2. I indeed meant that it solves both incompressible and compressible Stokes equations. The code has a compressible formulation that can indeed represent the incompressible limit by just changing some physical parameters. Following your suggestion, it now reads as just "compressible".
3. Fixed.

I have merged this corrections into the main branch of the code. If no more amends are needed, I will proceed to make a new GitHub and Zenodo version release.
