+++ ShearBand2D_PowerLaw_DYREL.jl
@@ -43,37 +43,19 @@ function main(igg; nx = 64, ny = 64, figdir = "model_figs")
 
     # Physical domain ------------------------------------
     ly = 1.0e0          # domain length in y
-    lx = ly           # domain length in x
-    ni = nx, ny       # number of cells
-    li = lx, ly       # domain length in x- and y-
-    di = @. li / ni   # grid step in x- and -y
-    origin = 0.0, 0.0     # origin coordinates
+    lx = ly             # domain length in x
+    ni = nx, ny         # number of cells
+    li = lx, ly         # domain length in x- and y-
+    di = @. li / ni     # grid step in x- and -y
+    origin = @. -li / 2  # origin coordinates
     grid = Geometry(ni, li; origin = origin)
     (; xci, xvi) = grid # nodes at the center and vertices of the cells
-    dt = Inf
-
+    dt  = Inf
+    εbg = 1
+    
     # Physical properties using GeoParams ----------------
-    τ_y = 1.6           # yield stress. If do_DP=true, τ_y stand for the cohesion: c*cos(ϕ)
-    ϕ = 30            # friction angle
-    C = τ_y           # Cohesion
-    η0 = 1.0           # viscosity
-    G0 = 1.0           # elastic shear modulus
-    Gi = G0 / (6.0 - 4.0)  # elastic shear modulus perturbation
-    εbg = 1.0           # background strain-rate
-    η_reg = 8.0e-3          # regularisation "viscosity"
-    dt = η0 / G0 / 4.0     # assumes Maxwell time of 4
-    el_bg = ConstantElasticity(; G = G0, Kb = 5)
-    el_inc = ConstantElasticity(; G = Gi, Kb = 5)
-    visc_bg  = PowerlawViscous(; η0 = 1e2,  n=3, ε0 = 1e0 )
-    visc_inc = PowerlawViscous(; η0 = 1e-1, n=1, ε0 = 1e0 )
-
-    pl = DruckerPrager_regularised(;
-        # non-regularized plasticity
-        C = C / cosd(ϕ),
-        ϕ = ϕ,
-        η_vp = η_reg,
-        Ψ = 0
-    )
+    visc_bg  = PowerlawViscous(; η0 = 1e2,  n=3, ε0 = 1e0)
+    visc_inc = PowerlawViscous(; η0 = 1e-1, n=3, ε0 = 1e0)
 
     rheology = (
         # Low density phase
@@ -82,7 +64,6 @@ function main(igg; nx = 64, ny = 64, figdir = "model_figs")
             Density = ConstantDensity(; ρ = 0.0),
             Gravity = ConstantGravity(; g = 0.0),
             CompositeRheology = CompositeRheology((visc_bg,)),
-            # Elasticity = el_bg,
         ),
         # High density phase
         SetMaterialParams(;
@@ -90,31 +71,28 @@ function main(igg; nx = 64, ny = 64, figdir = "model_figs")
             Density = ConstantDensity(; ρ = 0.0),
             Gravity = ConstantGravity(; g = 0.0),
             CompositeRheology = CompositeRheology((visc_inc,)),
-            # Elasticity = el_inc,
         ),
     )
-
-    # perturbation array for the cohesion
-    perturbation_C = @zeros(ni...)
 
     # Initialize phase ratios -------------------------------
     phase_ratios = PhaseRatios(backend_JP, length(rheology), ni)
     radius = 0.1
-    origin = 0.5, 0.5
[…]
