+++ Stokes2D_DYREL_simple_PowerLaw_v5(1).jl
@@ -6,16 +6,74 @@ using Plots, Printf, Statistics, LinearAlgebra
 @views av_ya(A) =  0.5*(A[:,1:end-1].+A[:,2:end])
 @views av4_harm(A) = 1.0./( 0.25.*(1.0./A[1:end-1,1:end-1].+1.0./A[2:end,1:end-1].+1.0./A[1:end-1,2:end].+1.0./A[2:end,2:end]) ) 
 
-function free_slip!(A)
-    @views A[:,1]   .= A[:,2] 
-    @views A[:,end] .= A[:,end-1]
-    @views A[1,:]   .= A[2,:]
-    @views A[end,:] .= A[end-1,:]
-    return nothing
+# can be replaced by AD
+function Gershgorin_Stokes2D_SchurComplement(ηc, ηv, γ, Δx, Δy, ncx, ncy)
+    
+    Dx = zeros(ncx-1, ncy)
+    Dy = zeros(ncx, ncy-1)
+    λmaxVx = zeros(ncx-1, ncy)
+    λmaxVy = zeros(ncx, ncy-1)
+    
+    # Compute Dx, Cxx, Cxy for x-component
+    for j in 1:ncy
+        for i in 1:ncx-1
+            ηN = (j > 1) ? ηv[i+1, j] : ηv[i+1, j+1]
+            ηS = (j < ncy) ? ηv[i+1, j+1] : ηv[i+1, j]
+            ηW = ηc[i, j]
+            ηE = ηc[i+1, j]
+            ebW = γ[i, j]
+            ebE = γ[i+1, j]
+            
+            Cxx = abs(ηN / Δy^2) + abs(ηS / Δy^2) + 
+                  abs(ebE / Δx^2 + (4/3) * ηE / Δx^2) + 
+                  abs(ebW / Δx^2 + (4/3) * ηW / Δx^2) + 
+                  abs(-(-ηN / Δy - ηS / Δy) / Δy + (ebE / Δx + ebW / Δx) / Δx + 
+                      ((4/3) * ηE / Δx + (4/3) * ηW / Δx) / Δx)
+            
+            Cxy = abs(ebE / (Δx * Δy) - 2/3 * ηE / (Δx * Δy) + ηN / (Δx * Δy)) + 
+                  abs(ebE / (Δx * Δy) - 2/3 * ηE / (Δx * Δy) + ηS / (Δx * Δy)) + 
+                  abs(ebW / (Δx * Δy) + ηN / (Δx * Δy) - 2/3 * ηW / (Δx * Δy)) + 
+                  abs(ebW / (Δx * Δy) + ηS / (Δx * Δy) - 2/3 * ηW / (Δx * Δy))
+            
+            Dx[i, j] = -(-ηN / Δy - ηS / Δy) / Δy + (ebE / Δx + ebW / Δx) / Δx + 
+                       ((4/3) * ηE / Δx + (4/3) * ηW / Δx) / Δx
+            
+            λmaxVx[i, j] = 1.0 / Dx[i, j] * (Cxx + Cxy)
+        end
+    end
+    
+    # Compute Dy, Cyy, Cyx for y-component
+    for j in 1:ncy-1
+        for i in 1:ncx
+            ηE = (i < ncx) ? ηv[i+1, j+1] : ηv[i, j+1]
+            ηW = (i > 1) ? ηv[i, j+1] : ηv[i+1, j+1]
[…]

[…]
