+++ src/DYREL/DYREL2D copy.jl
@@ -83,6 +83,7 @@ function _solve_DYREL!(
     stokes.λ  .= 0.0
     stokes.λv .= 0.0
 
+    ηc_true = similar(stokes.viscosity.η)
 
     # # compute buoyancy forces and viscosity
     # compute_ρg!(ρg, phase_ratios, rheology, args)
@@ -122,21 +123,31 @@ function _solve_DYREL!(
         @parallel (@idx ni .+ 1) compute_strain_rate!(
             @strain(stokes)..., stokes.∇V, @velocity(stokes)..., _di...
         )
-        vertex2center!(stokes.ε.xy_c, stokes.ε.xy)
 
         # compute deviatoric stress
-        compute_stress_DRYEL!(stokes, rheology, phase_ratios, λ_relaxation_PH, dt) # not resetting λ in every PH iteration seems to work better
+        # compute_stress_DRYEL!(stokes, rheology, phase_ratios, λ_relaxation_PH, dt) # not resetting λ in every PH iteration seems to work better
         # compute_stress_DRYEL!(stokes, rheology, phase_ratios, 1, dt) # λ_relaxation = 1 to reset λ
-        update_halo!(stokes.τ.xy)
+        # update_halo!(stokes.τ.xy)
 
-        update_viscosity_τII!(
-            stokes,
-            phase_ratios,
-            args,
-            rheology,
-            viscosity_cutoff;
-            relaxation = viscosity_relaxation,
-        )
+        # update_viscosity_τII!(
+        #     stokes,
+        #     phase_ratios,
+        #     args,
+        #     rheology,
+        #     viscosity_cutoff;
+        #     relaxation = viscosity_relaxation,
+        # )
+
+        vertex2center!(stokes.ε.xy_c, stokes.ε.xy)
+        stokes.τ.xx   .= 2.0.*stokes.viscosity.η  .* stokes.ε.xx 
+        stokes.τ.yy   .= 2.0.*stokes.viscosity.η  .* stokes.ε.yy 
+        stokes.τ.xy   .= 2.0.*stokes.viscosity.ηv .* stokes.ε.xy 
+        stokes.τ.xy_c .= 2.0.*stokes.viscosity.η  .* stokes.ε.xy_c
+        stokes.τ.II   .= sqrt.(0.5.*(stokes.τ.xx.^2  .+ stokes.τ.yy.^2  .+ (.-stokes.τ.xx.-stokes.τ.yy).^2)   .+ stokes.τ.xy_c.^2 )
+
+        ηc_true .= @. 2^(3-1) * args.ηc0^3 * stokes.τ.II^(1 - 3)
+        stokes.viscosity.η .= exp.(viscosity_relaxation*log.(ηc_true) .+ (1-viscosity_relaxation).*log.(stokes.viscosity.η))
+        center2vertex_harm!(stokes.viscosity.ηv, stokes.viscosity.η)
 
         # compute velocity residuals
         @parallel (@idx ni) compute_PH_residual_V!(
@@ -176,7 +187,6 @@ function _solve_DYREL!(
             (min(errVx/errVx0, errVx), min(errVy/errVy0, errVy), min(errPt/(errPt0 + eps()), errPt))
         )
         isnan(err) && error("NaN detected in outer loop")
-        err > 1e10 && error("Kaboom! Error > 1e10 in outer loop")
         # @printf("itPH = %02d iter = %06d iter/nx = %03d, err = %1.3e norm[Rx=%1.3e, Ry=%1.3e, Rp=%1.3e] \n", itPH, iter, iter/ni[1], err, min(errVx/errVx0, errVx), min(errVy/errVy0, errVy), min(errPt/errPt0, errPt))
[…]
