+++ miniapps/benchmarks/thermal_stress/Thermal_Stress_Magma_Chamber_nondim_VS.jl
@@ -1,0 +1,812 @@
+using CUDA
+using JustRelax, JustRelax.JustRelax2D, JustRelax.DataIO
+
+const backend_JR = CUDABackend
+
+using ParallelStencil, ParallelStencil.FiniteDifferences2D
+@init_parallel_stencil(CUDA, Float64, 2) #or (CUDA, Float64, 2) or (AMDGPU, Float64, 2)
+
+using JustPIC
+using JustPIC._2D
+# Threads is the default backend,
+# to run on a CUDA GPU load CUDA.jl (i.e. "using CUDA") at the beginning of the script,
+# and to run on an AMD GPU load AMDGPU.jl (i.e. "using AMDGPU") at the beginning of the script.
+const backend = CUDABackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+
+using Printf, Statistics, LinearAlgebra, GeoParams, GLMakie
+
+# -----------------------------------------------------------------------------------------
+## SET OF HELPER FUNCTIONS PARTICULAR FOR THIS SCRIPT --------------------------------
+function copyinn_x!(A, B)
+    @parallel function f_x(A, B)
+        @all(A) = @inn_x(B)
+        return nothing
+    end
+
+    return @parallel f_x(A, B)
+end
+
+import ParallelStencil.INDICES
+const idx_j = INDICES[2]
+macro all_j(A)
+    return esc(:($A[$idx_j]))
+end
+
+@parallel function init_P!(P, ρg, z, sticky_air)
+    @all(P) = abs(@all(ρg) * (@all_j(z) - sticky_air)) #* <(@all_j(z), 0.0)
+    return nothing
+end
+
+function init_phases!(phases, particles, xc_anomaly, yc_anomaly, r_anomaly, sticky_air, top, bottom)
+    ni = size(phases)
+
+    @parallel_indices (i, j) function init_phases!(
+            phases, px, py, index, xc_anomaly, yc_anomaly, r_anomaly, sticky_air, top, bottom
+        )
+        @inbounds for ip in cellaxes(phases)
+            # quick escape
+            @index(index[ip, i, j]) == 0 && continue
+
+            x = @index px[ip, i, j]
+            y = -(@index py[ip, i, j]) #- sticky_air
+            if top ≤ y ≤ bottom
+                @index phases[ip, i, j] = 1.0 # crust
+            end
+
+            # thermal anomaly - circular
+            if ((x - xc_anomaly)^2 + (y + yc_anomaly)^2 ≤ r_anomaly^2)
+                @index phases[ip, i, j] = 2.0
+            end
+
+            if y < top
+                @index phases[ip, i, j] = 3.0
+            end
+        end
+        return nothing
+    end
+
+    return @parallel (@idx ni) init_phases!(
+        phases,
+        particles.coords...,
+        particles.index,
+        xc_anomaly,
+        yc_anomaly,
+        r_anomaly,
+        sticky_air,
+        top,
+        bottom,
+    )
+end
+
+function heat_chamber_particles!(pT, phases, particles, T_chamber)
+    ni = size(phases)
+
+    @parallel_indices (i, j) function _heat_chamber_particles!(
+            pT, phases, index, T_chamber
+        )
+        @inbounds for ip in cellaxes(phases)
+            # quick escape
+            @index(index[ip, i, j]) == 0 && continue
+
+            # thermal anomaly 
+            if  @index(phases[ip, i, j]) == 2.0
+                @index pT[ip, i, j] = T_chamber
+            end
+        end
+        return nothing
+    end
+
[…]
