+++ src/topology/Topology.jl
@@ -107,7 +107,10 @@ function lazy_grid(
     # nodes at the center of the grid cells
     xci = ntuple(Val(N)) do i
         Base.@_inline_meta
-        @inbounds LinRange(origin[i] + di[i] / 2, origin[i] + Li[i] + di[i] / 2, ni[i])
+        if i == 1
+            @show origin[i] + Li[i] + di[i] / 2
+        end
+        @inbounds LinRange(origin[i] + di[i] / 2, origin[i] + Li[i] - di[i] / 2, ni[i])
     end
 
     # nodes at the vertices of the grid cells
