+++ src/DYREL/DYREL2D copy 2.jl
@@ -1,699 +1,0 @@
-## 2D VISCO-ELASTIC STOKES SOLVER
-
-# backend trait
-function solve_DYREL!(stokes::JustRelax.StokesArrays, args...; kwargs)
-    out = solve_DYREL!(backend(stokes), stokes, args...; kwargs)
-    return out
-end
-
-# entry point for extensions
-solve_DYREL!(::CPUBackendTrait, stokes, args...; kwargs) = _solve_DYREL!(stokes, args...; kwargs...)
-
-function _solve_DYREL!(
-        stokes::JustRelax.StokesArrays,
-        ρg,
-        dyrel,
-        flow_bcs::AbstractFlowBoundaryConditions,
-        phase_ratios::JustPIC.PhaseRatios,
-        rheology,
-        args,
-        di::NTuple{2, T},
-        dt,
-        igg::IGG;
-        viscosity_cutoff = (-Inf, Inf),
-        viscosity_relaxation = 1.0e-2,
-        λ_relaxation_DR = 1,
-        λ_relaxation_PH = 1,
-        iterMax = 50.0e3,
-        iterMin = 1.0e2,
-        free_surface = false,
-        nout = 100,
-        rel_drop = 1e-2,
-        b_width = (4, 4, 0),
-        verbose = true,
-        kwargs...,
-    ) where {T}
-
-    (;
-        γ_eff,
-        Dx,
-        Dy,
-        λmaxVx,
-        λmaxVy,
-        dVxdτ,
-        dVydτ,
-        dτVx,
-        dτVy,
-        dVx,
-        dVy,
-        βVx,
-        βVy,
-        cVx,
-        cVy,
-        αVx,
-        αVy,
-        c_fact,
-        ηb,
-    ) = dyrel
-    # unpack
-
-    _di   = inv.(di)
-    ni    = size(stokes.P)
-    γfact = 20
-
-    # errors
-    err      = 1.0
-    iter     = 0
-    err_evo1 = Float64[]
-    err_evo2 = Float64[]
-    norm_Rx  = Float64[]
-    norm_Ry  = Float64[]
-    norm_∇V  = Float64[]
-
-    # solver loop
-    @copy stokes.P0 stokes.P
-    Rx0 = similar(stokes.R.Rx)
-    Ry0 = similar(stokes.R.Ry)
-
-    for Aij in @tensor_center(stokes.ε_pl)
-        Aij .= 0.0
-    end
-
-    # reset plastic multiplier at the beginning of the time step
-    stokes.λ  .= 0.0
-    stokes.λv .= 0.0
-
-    ηc_true = similar(stokes.viscosity.η)
-
-    # # compute buoyancy forces and viscosity
-    # compute_ρg!(ρg, phase_ratios, rheology, args)
-    # # viscosity guess based on strain rate
-    # compute_viscosity!(stokes, phase_ratios, args, rheology, viscosity_cutoff)
-    # center2vertex!(stokes.viscosity.ηv, stokes.viscosity.η)
-
-    # rel_drop   = 1e-2 # relative drop of velocity residual per PH iteration
-    # Iteration loop
-    errVx0     = 1.0
-    errVy0     = 1.0
-    errPt0     = 1.0 
-    errVx00    = 1.0
-    errVy00    = 1.0 
-    iter       = 0
-    ϵ          = dyrel.ϵ
-    err        = 2 * ϵ
-    err_evo_V  = Float64[]
-    err_evo_P  = Float64[]
-    err_evo_it = Float64[]
-    itg        = 0
-    P_num      = similar(stokes.P)
-    # Powell-Hestenes iterations
-    for itPH in 1:250
-
[…]
