### 2026-01-23 #9365 @kthyng
@albert-de-montserrat The paper looks great. There is no need to make a new archive for paper changes. Since the compiled paper is part of the JOSS archive.
