### 2026-01-23 #9365 @willgearty
Huh, that error had been fixed by @aelligp in https://github.com/PTsolvers/JustRelax.jl/commit/fb4f8825ce90e14dceefbd8f1188514650b694dd but then that change was reverted by @albert-de-montserrat in https://github.com/PTsolvers/JustRelax.jl/commit/91940ec658ce4b33e9b3a9372cdbbbdaa6165f69
