### 2026-01-24 #9365 @aelligp
@kthyng @willgearty
It appeared that we changed this only in the JOSS branch and when switching to main it reoccurred. We have just fixed this and everything should be in order now.
