+++ dev/assets/man_security.md.BqRGWwJV.lean.js
@@ -1,1 +1,0 @@
-import{_ as t,c as i,o as r,aA as a}from"./chunks/framework.CNihevto.js";const y=JSON.parse('{"title":"Security Policy","description":"","frontmatter":{},"headers":[],"relativePath":"man/security.md","filePath":"man/security.md","lastUpdated":null}'),o={name:"man/security.md"};function s(n,e,l,u,p,c){return r(),i("div",null,[...e[0]||(e[0]=[a("",9)])])}const h=t(o,[["render",s]]);export{y as __pageData,h as default};
