+++ dev/assets/man_plume3D_plume3D.md.mAjuECQ0.js
@@ -1,0 +1,200 @@
+import{_ as t,o as n,c as h,aA as a,j as i,a as l}from"./chunks/framework.B8rbG-0R.js";const k="/JustRelax.jl/dev/assets/Plume3D.DrhDZzyC.png",D=JSON.parse('{"title":"3D Plume","description":"","frontmatter":{},"headers":[],"relativePath":"man/plume3D/plume3D.md","filePath":"man/plume3D/plume3D.md","lastUpdated":null}'),p={name:"man/plume3D/plume3D.md"},e={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},r={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.357ex"},xmlns:"http://www.w3.org/2000/svg",width:"3.05ex",height:"1.332ex",role:"img",focusable:"false",viewBox:"0 -431 1348.1 588.8","aria-hidden":"true"},E={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},d={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.357ex"},xmlns:"http://www.w3.org/2000/svg",width:"3.389ex",height:"1.332ex",role:"img",focusable:"false",viewBox:"0 -431 1498 588.8","aria-hidden":"true"};function g(y,s,o,F,c,C){return n(),h("div",null,[s[8]||(s[8]=a(`<h1 id="3D-Plume" tabindex="-1">3D Plume <a class="header-anchor" href="#3D-Plume" aria-label="Permalink to &quot;3D Plume {#3D-Plume}&quot;">​</a></h1><h1 id="Initialize-packages" tabindex="-1">Initialize packages <a class="header-anchor" href="#Initialize-packages" aria-label="Permalink to &quot;Initialize packages {#Initialize-packages}&quot;">​</a></h1><p>Load JustRelax necessary modules and define backend. In this case we will use the CUDA backend, but you can also use the CPU or AMD GPU backends.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> CUDA </span><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># comment this out if you are not using CUDA; or load AMDGPU.jl if you are using an AMD GPU</span></span>
+<span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> JustRelax, JustRelax</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">.</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">JustRelax3D, JustRelax</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">.</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">DataIO</span></span>
[…]

[…]
