+++ miniapps/benchmarks/stokes2D/shear_heating/Shearheating2D_DYREL.jl
@@ -1,0 +1,340 @@
+# Benchmark of Duretz et al. 2014
+# http://dx.doi.org/10.1002/2014GL060438
+using JustRelax, JustRelax.JustRelax2D, JustRelax.DataIO
+const backend_JR = CPUBackend
+
+using ParallelStencil, ParallelStencil.FiniteDifferences2D
+@init_parallel_stencil(Threads, Float64, 2)
+
+using JustPIC, JustPIC._2D
+# Threads is the default backend,
+# to run on a CUDA GPU load CUDA.jl (i.e. "using CUDA") at the beginning of the script,
+# and to run on an AMD GPU load AMDGPU.jl (i.e. "using AMDGPU") at the beginning of the script.
+const backend = JustPIC.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+
+# Load script dependencies
+using GeoParams, GLMakie
+
+# Load file with all the rheology configurations
+include("Shearheating_rheology.jl")
+
+## SET OF HELPER FUNCTIONS PARTICULAR FOR THIS SCRIPT --------------------------------
+
+function copyinn_x!(A, B)
+    @parallel function f_x(A, B)
+        @all(A) = @inn_x(B)
+        return nothing
+    end
+
+    return @parallel f_x(A, B)
+end
+
+import ParallelStencil.INDICES
+const idx_j = INDICES[2]
+macro all_j(A)
+    return esc(:($A[$idx_j]))
+end
+
+# Initial pressure profile - not accurate
+@parallel function init_P!(P, ρg, z)
+    @all(P) = abs(@all(ρg) * @all_j(z)) * <(@all_j(z), 0.0)
+    return nothing
+end
+
+
+## END OF HELPER FUNCTION ------------------------------------------------------------
+
+## BEGIN OF MAIN SCRIPT --------------------------------------------------------------
+function main2D(igg; ar = 8, ny = 16, nx = ny * 8, figdir = "figs2D", do_vtk = false)
+
+    # Physical domain ------------------------------------
+    ly = 40.0e3         # domain length in y
+    lx = 70.0e3         # domain length in x
+    ni = nx, ny         # number of cells
+    li = lx, ly         # domain length in x- and y-
+    di = @. li / ni     # grid step in x- and -y
+    origin = 0.0, -ly       # origin coordinates (15km f sticky air layer)
+    grid = Geometry(ni, li; origin = origin)
+    (; xci, xvi) = grid     # nodes at the center and vertices of the cells
+    # ----------------------------------------------------
+    # Physical properties using GeoParams ----------------
+    rheology = init_rheologies(; is_TP_Conductivity = false)
+    κ = (4 / (rheology[1].HeatCapacity[1].Cp * rheology[1].Density[1].ρ))
+    dt = dt_diff = 0.5 * min(di...)^2 / κ / 2.01 # diffusive CFL timestep limiter
+    # ----------------------------------------------------
+
+    # Initialize particles -------------------------------
+    nxcell, max_xcell, min_xcell = 20, 32, 12
+    particles = init_particles(
+        backend, nxcell, max_xcell, min_xcell, xvi...
+    )
+    # velocity grids
+    grid_vx, grid_vy = velocity_grids(xci, xvi, di)
+    # temperature
+    pT, pPhases = init_cell_arrays(particles, Val(3))
+    particle_args = (pT, pPhases)
+
+    # Elliptical temperature anomaly
+    xc_anomaly = lx / 2           # origin of thermal anomaly
+    yc_anomaly = 40.0e3           # origin of thermal anomaly
+    r_anomaly = 3.0e3            # radius of perturbation
[…]
