### 2026-01-27 fe089cb
update kernels & solver
