### 2026-01-28 adf03ab
initial version of the GMD benchmark files for Blankenbach and VanKeken
