+++ miniapps/benchmarks/stokes2D/VanKeken/VanKeken_GMD.jl
@@ -1,0 +1,287 @@
+using ParallelStencil
+@init_parallel_stencil(Threads, Float64, 2)
+
+using Printf, LinearAlgebra, GeoParams, CellArrays
+using JustRelax, JustRelax.JustRelax2D
+using MPI: MPI
+
+const backend_JR = CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+
+using GLMakie
+
+using JustPIC, JustPIC._2D
+# Threads is the default backend,
+# to run on a CUDA GPU load CUDA.jl (i.e. "using CUDA") at the beginning of the script,
+# and to run on an AMD GPU load AMDGPU.jl (i.e. "using AMDGPU") at the beginning of the script.
+const backend = JustPIC.CPUBackend # Options: CPUBackend, CUDABackend, AMDGPUBackend
+
+# x-length of the domain
+const λ = 0.9142
+
+# HELPER FUNCTIONS ---------------------------------------------------------------
+# Initialize phases on the particles
+function init_phases!(phases, particles)
+    ni = size(phases)
+
+    @parallel_indices (i, j) function init_phases!(phases, px, py, index)
+        @inbounds for ip in cellaxes(phases)
+            # quick escape
+            @index(index[ip, i, j]) == 0 && continue
+
+            x = @index px[ip, i, j]
+            y = @index py[ip, i, j]
+
+            # plume - rectangular
+            if y > 0.2 + 0.02 * cos(π * x / λ)
+                @index phases[ip, i, j] = 2.0
+            else
+                @index phases[ip, i, j] = 1.0
+            end
+        end
+        return nothing
+    end
+
+    return @parallel (@idx ni) init_phases!(phases, particles.coords..., particles.index)
+end
+# END OF HELPER FUNCTIONS --------------------------------------------------------
+
+# MAIN SCRIPT --------------------------------------------------------------------
+function VanKeken2D(igg; ny = 64, nx = 64, figdir = "model_figs", finalize_MPI = false)
+
+    # Physical domain ------------------------------------
+    ly = 1            # domain length in y
+    lx = ly           # domain length in x
+    ni = nx, ny       # number of cells
+    li = lx, ly       # domain length in x- and y-
+    di = @. li / ni   # grid step in x- and -y
+    origin = 0.0, 0.0     # origin coordinates
+    grid = Geometry(ni, li; origin = origin)
+    (; xci, xvi) = grid # nodes at the center and vertices of the cells
+    dt = 1.0e-10
+
+    # Physical properties using GeoParams ----------------
+    rheology = (
+        # Low density phase
+        SetMaterialParams(;
+            Phase = 1,
+            Density = ConstantDensity(; ρ = 1),
+            Gravity = ConstantGravity(; g = 1),
+            CompositeRheology = CompositeRheology((LinearViscous(; η = 1.0e0),)),
+
+        ),
+        # High density phase
+        SetMaterialParams(;
+            Density = ConstantDensity(; ρ = 2),
+            Gravity = ConstantGravity(; g = 1),
+            CompositeRheology = CompositeRheology((LinearViscous(; η = 1.0e0),)),
+        ),
+    )
+
+    # Initialize particles -------------------------------
+    nxcell, max_p, min_p = 20, 30, 10
+    particles = init_particles(
+        backend, nxcell, max_p, min_p, xvi...
+    )
+    # velocity grids
+    grid_vx, grid_vy = velocity_grids(xci, xvi, di)
[…]
