+++ miniapps/benchmarks/DYREL2D/shear_band/ShearBand2D_DYREL.jl
@@ -160,6 +160,7 @@ function main(igg; nx = 64, ny = 64, figdir = "model_figs")
                 verbose_PH = false,
                 verbose_DR = false,
                 viscosity_relaxation = 1/2,
+                linear_viscosity = true,
                 viscosity_cutoff = (-Inf, Inf),
             )
         );
