+++ miniapps/subduction/2D/Subduction2D_DYREL.jl
@@ -1,7 +1,7 @@
 # Load script dependencies
 using GeoParams, GLMakie
 
-const isCUDA = true
+const isCUDA = false
 
 @static if isCUDA
     using CUDA
@@ -206,12 +206,11 @@ function main(li, origin, phases_GMG, igg; nx = 16, ny = 16, figdir = "figs2D", 
                 dt,
                 igg;
                 kwargs = (;
-                    verbose = false,
-                    iterMax = 50.0e3,
-                    nout    = 200,
+                    verbose  = false,
+                    iterMax  = 50.0e3,
+                    rel_drop = 1e-2,
+                    nout     = 400,
                     λ_relaxation = 1,
-                    # λ_relaxation = 1e-2,
-                    # λ_relaxation = 1.075,
                     viscosity_relaxation = 1.0e-3,
                     viscosity_cutoff = (1e18, 1e23),
                 )
@@ -361,7 +360,7 @@ end
 do_vtk = true # set to true to generate VTK files for ParaView
 figdir = "Subduction2D_reset_lambda"
 # n = 128 * 2
-n = 128
+n = 200
 nx, ny = n * 2, n
 
 li, origin, phases_GMG, T_GMG = GMG_subduction_2D(nx + 1, ny + 1)
