+++ previews/PR428/assets/man_advection.md.kLWSew0K.lean.js
@@ -1,0 +1,1 @@
+import{_ as a,o as t,c as i,aA as l}from"./chunks/framework.DQz_OLHA.js";const h=JSON.parse('{"title":"Field advection","description":"","frontmatter":{},"headers":[],"relativePath":"man/advection.md","filePath":"man/advection.md","lastUpdated":null}'),r={name:"man/advection.md"};function n(o,e,s,c,d,_){return t(),i("div",null,[...e[0]||(e[0]=[l("",3)])])}const f=a(r,[["render",n]]);export{h as __pageData,f as default};
