+++ previews/PR428/assets/man_Blankenbach.md.D9NBDgRr.js
@@ -1,0 +1,209 @@
+import{_ as l,o as n,c as t,aA as a,j as s,a as h}from"./chunks/framework.DQz_OLHA.js";const k="/JustRelax.jl/previews/PR428/assets/Blankenbach_Temp.BkOD86kY.png",p="/JustRelax.jl/previews/PR428/assets/Blankenbach_Time_Series_V_Nu.Z1Dq9Kzc.png",b=JSON.parse('{"title":"Blankenbach benchmark","description":"","frontmatter":{},"headers":[],"relativePath":"man/Blankenbach.md","filePath":"man/Blankenbach.md","lastUpdated":null}'),e={name:"man/Blankenbach.md"},r={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},d={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.566ex"},xmlns:"http://www.w3.org/2000/svg",width:"11.478ex",height:"2.262ex",role:"img",focusable:"false",viewBox:"0 -750 5073.2 1000","aria-hidden":"true"},E={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},g={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.816ex"},xmlns:"http://www.w3.org/2000/svg",width:"23.306ex",height:"2.773ex",role:"img",focusable:"false",viewBox:"0 -864.9 10301.2 1225.5","aria-hidden":"true"},y={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},o={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-2.607ex"},xmlns:"http://www.w3.org/2000/svg",width:"25.594ex",height:"6.923ex",role:"img",focusable:"false",viewBox:"0 -1907.6 11312.6 3060","aria-hidden":"true"},Q={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},c={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-1.948ex"},xmlns:"http://www.w3.org/2000/svg",width:"17.518ex",height:"5.095ex",role:"img",focusable:"false",viewBox:"0 -1391 7743.1 2252","aria-hidden":"true"};function T(F,i,m,C,u,A){return n(),t("div",null,[i[10]||(i[10]=a(`<h1 id="Blankenbach-benchmark" tabindex="-1">Blankenbach benchmark <a class="header-anchor" href="#Blankenbach-benchmark" aria-label="Permalink to &quot;Blankenbach benchmark {#Blankenbach-benchmark}&quot;">​</a></h1><p>Thermal convection benchmark from <a href="https://academic.oup.com/gji/article/98/1/23/622167" target="_blank" rel="noreferrer">Blankenbach et al., 1989</a></p><h2 id="Initialize-packages" tabindex="-1">Initialize packages <a class="header-anchor" href="#Initialize-packages" aria-label="Permalink to &quot;Initialize packages {#Initialize-packages}&quot;">​</a></h2><p>Load <a href="https://github.com/PTsolvers/JustRelax.jl" target="_blank" rel="noreferrer">JustRelax.jl</a> necessary modules and define backend.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="
[…]
