+++ previews/PR428/assets/man_subduction2D_rheology.md.5xXls7GC.lean.js
@@ -1,0 +1,1 @@
+import{_ as e,o as t,c as n,j as s,a,aA as l}from"./chunks/framework.DQz_OLHA.js";const H=JSON.parse('{"title":"Using GeoParams.jl to define the rheology of the material phases","description":"","frontmatter":{},"headers":[],"relativePath":"man/subduction2D/rheology.md","filePath":"man/subduction2D/rheology.md","lastUpdated":null}'),h={name:"man/subduction2D/rheology.md"},p={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},k={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.025ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.448ex",height:"1.025ex",role:"img",focusable:"false",viewBox:"0 -442 640 453","aria-hidden":"true"},Q={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},r={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.65ex"},xmlns:"http://www.w3.org/2000/svg",width:"2.61ex",height:"2.245ex",role:"img",focusable:"false",viewBox:"0 -705 1153.7 992.2","aria-hidden":"true"},d={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},T={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.566ex"},xmlns:"http://www.w3.org/2000/svg",width:"34.965ex",height:"2.262ex",role:"img",focusable:"false",viewBox:"0 -750 15454.7 1000","aria-hidden":"true"},o={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},g={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.566ex"},xmlns:"http://www.w3.org/2000/svg",width:"33.732ex",height:"2.709ex",role:"img",focusable:"false",viewBox:"0 -947.3 14909.3 1197.3","aria-hidden":"true"};function E(m,i,y,c,u,C){return t(),n("div",null,[i[18]||(i[18]=s("h1",{id:"Using-GeoParams.jl-to-define-the-rheology-of-the-material-phases",tabindex:"-1"},[a("Using GeoParams.jl to define the rheology of the material phases "),s("a",{class:"header-anchor",href:"#Using-GeoParams.jl-to-define-the-rheology-of-the-material-phases","aria-label":'Permalink to "Using GeoParams.jl to define the rheology of the material phases {#Using-GeoParams.jl-to-define-the-rheology-of-the-material-phases}"'},"​")],-1)),i[19]||(i[19]=s("p",null,[a("We will use the same physical parameters as the ones defined in "),s("a",{href:"https://doi.org/10.5194/se-15-567-2024",target:"_blank",rel:"noreferrer"},"Hummel et al 2024"),a(".")],-1)),s("p",null,[i[4]||(i[4]=a("The thermal expansion coefficient ",-1)),s("mjx-container",p,[(t(),n("svg",k,[...i[0]||(i[0]=[s("g",{stroke:"currentColor",fill:"currentColor","stroke-width":"0",transform:"scale(1,-1)"},[s("g",{"data-mml-node":"math"},[s("g",{"data-mml-node":"mi"},[s("path",{"data-c":"1D6FC",d:"M34 156Q34 270 120 3
[…]
