+++ previews/PR428/assets/man_security.md.Dfmk9iPS.lean.js
@@ -1,0 +1,1 @@
+import{_ as t,o as i,c as r,aA as a}from"./chunks/framework.DQz_OLHA.js";const y=JSON.parse('{"title":"Security Policy","description":"","frontmatter":{},"headers":[],"relativePath":"man/security.md","filePath":"man/security.md","lastUpdated":null}'),o={name:"man/security.md"};function s(n,e,l,u,p,c){return i(),r("div",null,[...e[0]||(e[0]=[a("",9)])])}const h=t(o,[["render",s]]);export{y as __pageData,h as default};
