+++ previews/PR428/assets/man_material_physics.md.fFT2NFyQ.lean.js
@@ -1,0 +1,1 @@
+import{_ as r,o as s,c as i,j as a,a as e}from"./chunks/framework.DQz_OLHA.js";const f=JSON.parse('{"title":"Material physics","description":"","frontmatter":{},"headers":[],"relativePath":"man/material_physics.md","filePath":"man/material_physics.md","lastUpdated":null}'),l={name:"man/material_physics.md"};function o(n,t,c,p,m,h){return s(),i("div",null,[...t[0]||(t[0]=[a("h1",{id:"Material-physics",tabindex:"-1"},[e("Material physics "),a("a",{class:"header-anchor",href:"#Material-physics","aria-label":'Permalink to "Material physics {#Material-physics}"'},"​")],-1),a("p",null,[a("a",{href:"https://github.com/PTsolvers/JustRelax.jl",target:"_blank",rel:"noreferrer"},"JustRelax.jl"),e(" is fully integrated with "),a("a",{href:"https://github.com/JuliaGeodynamics/GeoParams.jl/tree/main",target:"_blank",rel:"noreferrer"},"GeoParams.jl"),e(" to perform all the material physics computations.")],-1)])])}const _=r(l,[["render",o]]);export{f as __pageData,_ as default};

[…]
