+++ src/DYREL/constructors.jl
@@ -1,83 +1,85 @@
-function DYREL(ni::NTuple{2}; ϵ=1e-6, ϵ_vel=1e-6, CFL= 0.99, c_fat = 0.5)
+function DYREL(ni::NTuple{2}; ϵ = 1.0e-6, ϵ_vel = 1.0e-6, CFL = 0.99, c_fat = 0.5)
     nx, ny = ni
     # penalty parameter
-    γ_eff  = @zeros(nx, ny)
+    γ_eff = @zeros(nx, ny)
     # bulk viscosity
-    ηb     = @zeros(nx, ny)
+    ηb = @zeros(nx, ny)
     # Diagonal preconditioner arrays
-    Dx     = @zeros(nx-1, ny)
-    Dy     = @zeros(nx, ny-1)
+    Dx = @zeros(nx - 1, ny)
+    Dy = @zeros(nx, ny - 1)
     # maximum eigenvalue estimates
-    λmaxVx = @zeros(nx-1, ny)
-    λmaxVy = @zeros(nx, ny-1)
-    dVxdτ  = @zeros(nx-1, ny)
-    dVydτ  = @zeros(nx, ny-1)
-    dτVx   = @zeros(nx-1, ny)
-    dτVy   = @zeros(nx, ny-1)
-    dVx    = @zeros(nx-1, ny)
-    dVy    = @zeros(nx, ny-1)
-    βVx    = @zeros(nx-1, ny)
-    βVy    = @zeros(nx, ny-1)
-    cVx    = @zeros(nx-1, ny)
-    cVy    = @zeros(nx, ny-1)
-    αVx    = @zeros(nx-1, ny)
-    αVy    = @zeros(nx, ny-1)
-    
+    λmaxVx = @zeros(nx - 1, ny)
+    λmaxVy = @zeros(nx, ny - 1)
+    dVxdτ = @zeros(nx - 1, ny)
+    dVydτ = @zeros(nx, ny - 1)
+    dτVx = @zeros(nx - 1, ny)
+    dτVy = @zeros(nx, ny - 1)
+    dVx = @zeros(nx - 1, ny)
+    dVy = @zeros(nx, ny - 1)
+    βVx = @zeros(nx - 1, ny)
+    βVy = @zeros(nx, ny - 1)
+    cVx = @zeros(nx - 1, ny)
+    cVy = @zeros(nx, ny - 1)
+    αVx = @zeros(nx - 1, ny)
+    αVy = @zeros(nx, ny - 1)
+
     T = typeof(γ_eff)
     F = typeof(CFL)
-    JustRelax.DYREL{T, F}(γ_eff, Dx, Dy, λmaxVx, λmaxVy, dVxdτ, dVydτ, dτVx, dτVy,
-                     dVx, dVy, βVx, βVy, cVx, cVy, αVx, αVy, ηb, CFL, ϵ, ϵ_vel, c_fat)
+    return JustRelax.DYREL{T, F}(
+        γ_eff, Dx, Dy, λmaxVx, λmaxVy, dVxdτ, dVydτ, dτVx, dτVy,
+        dVx, dVy, βVx, βVy, cVx, cVy, αVx, αVy, ηb, CFL, ϵ, ϵ_vel, c_fat
+    )
 end
 
-DYREL(nx::Integer, ny::Integer; ϵ=1e-6, ϵ_vel=1e-6, CFL= 0.99, c_fat = 0.5) = DYREL((nx, ny); ϵ=ϵ, ϵ_vel=ϵ_vel, CFL=CFL, c_fat=c_fat)
[…]
