+++ src/DYREL/pressure_kernels.jl
@@ -1,4 +1,3 @@
-
 function compute_residual_P!(
         RP,
         P,
@@ -110,4 +109,4 @@ end
 end
 
 @inline _compute_RP!(P, P0, ∇V, Q, ηb, dt) = -∇V - (P - P0) / ηb + (Q / dt)
-@inline _compute_RP!(P, P0, ∇V, Q, ΔTc, α, ηb, dt) = -∇V - (P - P0) / ηb  + α * (ΔTc / dt) + (Q / dt) 
+@inline _compute_RP!(P, P0, ∇V, Q, ΔTc, α, ηb, dt) = -∇V - (P - P0) / ηb + α * (ΔTc / dt) + (Q / dt)

[…]
