+++ previews/PR428/assets/chunks/@localSearchIndexroot.IPF6qHaZ.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":153,"nextId":153,"documentIds":{"0":"/JustRelax.jl/previews/PR428/#What-is-JustRelax.jl?","1":"/JustRelax.jl/previews/PR428/#How-to-Install-JustRelax.jl?","2":"/JustRelax.jl/previews/PR428/#Example-miniapps","3":"/JustRelax.jl/previews/PR428/#Funding","4":"/JustRelax.jl/previews/PR428/man/Blankenbach#Blankenbach-benchmark","5":"/JustRelax.jl/previews/PR428/man/Blankenbach#Initialize-packages","6":"/JustRelax.jl/previews/PR428/man/Blankenbach#Script","7":"/JustRelax.jl/previews/PR428/man/Blankenbach#Model-domain","8":"/JustRelax.jl/previews/PR428/man/Blankenbach#Rheology","9":"/JustRelax.jl/previews/PR428/man/Blankenbach#Initialize-particles","10":"/JustRelax.jl/previews/PR428/man/Blankenbach#Temperature-anomaly","11":"/JustRelax.jl/previews/PR428/man/Blankenbach#Stokes-and-heat-diffusion-arrays","12":"/JustRelax.jl/previews/PR428/man/Blankenbach#Initialize-thermal-profile-and-viscosity-fields","13":"/JustRelax.jl/previews/PR428/man/Blankenbach#Boundary-conditions","14":"/JustRelax.jl/previews/PR428/man/Blankenbach#Pseuo-transient-coefficients","15":"/JustRelax.jl/previews/PR428/man/Blankenbach#Just-before-solving-the-problem...","16":"/JustRelax.jl/previews/PR428/man/Blankenbach#Advancing-one-time-step","17":"/JustRelax.jl/previews/PR428/man/Blankenbach#Visualization","18":"/JustRelax.jl/previews/PR428/man/Blankenbach#Fields","19":"/JustRelax.jl/previews/PR428/man/Blankenbach#Final-model","20":"/JustRelax.jl/previews/PR428/man/GeoParams#List-of-GeoParams-functions","21":"/JustRelax.jl/previews/PR428/man/JustPIC#List-of-JustPIC-functions","22":"/JustRelax.jl/previews/PR428/man/ShearBands#ShearBand-benchmark","23":"/JustRelax.jl/previews/PR428/man/ShearBands#Initialize-packages","24":"/JustRelax.jl/previews/PR428/man/ShearBands#Script","25":"/JustRelax.jl/previews/PR428/man/ShearBands#Model-domain","26":"/JustRelax.jl/previews/PR428/man/ShearBands#Physical-properties-using-GeoParams","27":"/JustRelax.jl/previews/PR428/man/ShearBands#Rheology","28":"/JustRelax.jl/previews/PR428/man/ShearBands#Phase-anomaly","29":"/JustRelax.jl/previews/PR428/man/ShearBands#Stokes-arrays","30":"/JustRelax.jl/previews/PR428/man/ShearBands#Pseuo-transient-coefficients","31":"/JustRelax.jl/previews/PR428/man/ShearBands#Initialize-viscosity-fields","32":"/JustRelax.jl/previews/PR428/man/ShearBands#Boundary-conditions","33":"/JustRelax.jl/previews/PR428/man/ShearBands#Just-before-solving-the-problem...","34":"/JustRelax.jl/previews/PR428/man/ShearBands#Solving-Stokes","35":"/JustRelax.jl/previews/PR428/man/ShearBands#Visualization","36":"/JustRelax.jl/previews/PR428/man/ShearBands#Fields","37":"/JustRelax.jl/previews/PR428/man/ShearBands#Final-model",
[…]
