### 2026-01-30 c88e38f
MPI dyrel & miniapp
