### 2026-02-02 e4c3f27
remove old miniapps
