+++ ShearBand2D_PowerLaw_APT.jl
@@ -1,195 +1,0 @@
-using GeoParams, GLMakie, CellArrays
-using JustRelax, JustRelax.JustRelax2D
-using ParallelStencil
-@init_parallel_stencil(Threads, Float64, 2)
-
-const backend = CPUBackend
-
-using JustPIC, JustPIC._2D
-import JustPIC._2D.GridGeometryUtils as GGU
-
-const backend_JP = JustPIC.CPUBackend
-
-# HELPER FUNCTIONS ----------------------------------- ----------------------------
-solution(ε, t, G, η) = 2 * ε * η * (1 - exp(-G * t / η))
-
-# Initialize phases on the particles
-function init_phases!(phase_ratios, xci, xvi, circle)
-    ni = size(phase_ratios.center)
-
-    @parallel_indices (i, j) function init_phases!(phases, xc, yc, circle)
-        x, y = xc[i], yc[j]
-        p = GGU.Point(x, y)
-        if GGU.inside(p, circle)
-            @index phases[1, i, j] = 0.0
-            @index phases[2, i, j] = 1.0
-
-        else
-            @index phases[1, i, j] = 1.0
-            @index phases[2, i, j] = 0.0
-
-        end
-        return nothing
-    end
-
-    @parallel (@idx ni) init_phases!(phase_ratios.center, xci..., circle)
-    @parallel (@idx ni .+ 1) init_phases!(phase_ratios.vertex, xvi..., circle)
-    return nothing
-end
-
-
-# MAIN SCRIPT --------------------------------------------------------------------
-function main(igg; nx = 64, ny = 64, figdir = "model_figs")
-
-    # Physical domain ------------------------------------
-    ly = 1.0e0          # domain length in y
-    lx = ly           # domain length in x
-    ni = nx, ny       # number of cells
-    li = lx, ly       # domain length in x- and y-
-    di = @. li / ni   # grid step in x- and -y
-    origin = 0.0, 0.0     # origin coordinates
-    grid = Geometry(ni, li; origin = origin)
-    (; xci, xvi) = grid # nodes at the center and vertices of the cells
-    dt = Inf
-
-    # Physical properties using GeoParams ----------------
-    visc_bg  = PowerlawViscous(; η0 = 1e2,  n=3, ε0 = 1e0 )
-    visc_inc = PowerlawViscous(; η0 = 1e-1, n=3, ε0 = 1e0 )
-
-    rheology = (
-        # Low density phase
-        SetMaterialParams(;
-            Phase = 1,
-            Density = ConstantDensity(; ρ = 0.0),
-            Gravity = ConstantGravity(; g = 0.0),
-            CompositeRheology = CompositeRheology((visc_bg,)),
-        ),
-        # High density phase
-        SetMaterialParams(;
-            Phase = 2,
-            Density = ConstantDensity(; ρ = 0.0),
-            Gravity = ConstantGravity(; g = 0.0),
-            CompositeRheology = CompositeRheology((visc_inc,)),
-        ),
-    )
-
-    # Initialize phase ratios -------------------------------
-    phase_ratios = PhaseRatios(backend_JP, length(rheology), ni)
-    radius = 0.1
-    origin = 0.5, 0.5
-    circle = GGU.Circle(origin, radius)
-    init_phases!(phase_ratios, xci, xvi, circle)
-
-    # STOKES ---------------------------------------------
-    # Allocate arrays needed for every Stokes problem
-    stokes = StokesArrays(backend, ni)
[…]
