+++ Stokes2D_DYREL_simple_PowerLaw_v5(1) copy.jl
@@ -1,897 +1,0 @@
-# Initialisation
-# Initialisation
-using GeoParams
-using JustRelax, JustRelax.JustRelax2D
-using ParallelStencil
-@init_parallel_stencil(Threads, Float64, 2)
-
-const backend = CPUBackend
-
-using JustPIC, JustPIC._2D
-import JustPIC._2D.GridGeometryUtils as GGU
-
-const backend_JP = JustPIC.CPUBackend
-
-import JustRelax.JustRelax2D as JR
-
-using Plots, Printf, Statistics, LinearAlgebra
-# Macros
-@views    av(A) = 0.25*(A[1:end-1,1:end-1].+A[2:end,1:end-1].+A[1:end-1,2:end].+A[2:end,2:end])
-@views av_xa(A) =  0.5*(A[1:end-1,:].+A[2:end,:])
-@views av_ya(A) =  0.5*(A[:,1:end-1].+A[:,2:end])
-@views av4_harm(A) = 1.0./( 0.25.*(1.0./A[1:end-1,1:end-1].+1.0./A[2:end,1:end-1].+1.0./A[1:end-1,2:end].+1.0./A[2:end,2:end]) ) 
-
-# Initialize phases on the particles
-function init_phases!(phase_ratios, xci, xvi, circle)
-    ni = size(phase_ratios.center)
-
-    @parallel_indices (i, j) function init_phases!(phases, xc, yc, circle)
-        x, y = xc[i], yc[j]
-        p = GGU.Point(x, y)
-        # if GGU.inside(p, circle)
-        #     @index phases[1, i, j] = 0.0
-        #     @index phases[2, i, j] = 1.0
-
-        # else
-        #     @index phases[1, i, j] = 1.0
-        #     @index phases[2, i, j] = 0.0
-
-        # end
-        r = 0.1
-        if x^2 + y^2 < r^2 
-            @index phases[1, i, j] = 0.0
-            @index phases[2, i, j] = 1.0
-
-        else
-            @index phases[1, i, j] = 1.0
-            @index phases[2, i, j] = 0.0
-
-        end
-        return nothing
-    end
-
-    @parallel (@idx ni) init_phases!(phase_ratios.center, xci..., circle)
-    @parallel (@idx ni .+ 1) init_phases!(phase_ratios.vertex, xvi..., circle)
-    return nothing
-end
-
-# can be replaced by AD
-function Gershgorin_Stokes2D_SchurComplement(ηc, ηv, γ, Δx, Δy, ncx, ncy)
-    
-    Dx = zeros(ncx-1, ncy)
-    Dy = zeros(ncx, ncy-1)
-    λmaxVx = zeros(ncx-1, ncy)
-    λmaxVy = zeros(ncx, ncy-1)
-    
-    # Compute Dx, Cxx, Cxy for x-component
-    for j in 1:ncy
-        for i in 1:ncx-1
-            ηN = (j > 1) ? ηv[i+1, j] : ηv[i+1, j+1]
-            ηS = (j < ncy) ? ηv[i+1, j+1] : ηv[i+1, j]
-            ηW = ηc[i, j]
-            ηE = ηc[i+1, j]
-            ebW = γ[i, j]
-            ebE = γ[i+1, j]
-            
[…]

[…]
