### 2026-02-02 5714ddd
fix stress rotation
