### 2026-02-03 7b6789c
update minapp project.toml
