### 2026-02-03 0fe626c
update miniapp to (mostly) consistent resolution
