+++ miniapps/benchmarks/stokes2D/sinking_block/SinkingBlock2D_WENO5.jl
@@ -278,7 +278,7 @@ function sinking_block2D(igg; ar = 8, ny = 16, nx = ny * 8, figdir = "figs2D", t
 end
 
 ar = 1 # aspect ratio
-n = 128
+n = 64
 nx = n * ar - 2
 ny = n - 2
 igg = if !(JustRelax.MPI.Initialized()) # initialize (or not) MPI grid
