+++ miniapps/convection/Particles2D/Layered_convection2D.jl
@@ -379,7 +379,7 @@ end
 figdir = "Plume2D"
 do_vtk = false # set to true to generate VTK files for ParaView
 ar = 1 # aspect ratio
-n = 128
+n = 64
 nx = n * ar - 2
 ny = n - 2
 igg = if !(JustRelax.MPI.Initialized()) # initialize (or not) MPI grid
