+++ miniapps/convection/GlobalConvection2D_WENO5.jl
@@ -3,7 +3,7 @@ using JustRelax, JustRelax.JustRelax2D, JustRelax.DataIO
 
 const backend = CPUBackend
 
-using ParallelStencil
+using ParallelStencil, ParallelStencil.FiniteDifferences2D
 @init_parallel_stencil(Threads, Float64, 2) #or (CUDA, Float64, 2) or (AMDGPU, Float64, 2)
 
 using GeoParams, CairoMakie, SpecialFunctions
