+++ miniapps/benchmarks/stokes2D/shear_band/ShearBand2D.jl
@@ -192,7 +192,7 @@ function main(igg; nx = 64, ny = 64, figdir = "model_figs")
     return nothing
 end
 
-n  = 64
+n = 64
 nx = n
 ny = n
 figdir = "ShearBands2D"
