+++ miniapps/benchmarks/thermal_stress/Thermal_Stress_Magma_Chamber_nondim.jl
@@ -745,8 +745,8 @@ figdir = "Thermal_stresses_around_cooling_magma_NonLinear"
 do_vtk = true # set to true to generate VTK files for ParaView
 n = 64
 ar = 1
-nx = n * ar - 2
-ny = n - 2
+nx = n * ar
+ny = n
 igg = if !(JustRelax.MPI.Initialized()) # initialize (or not) MPI grid
     IGG(init_global_grid(nx, ny, 1; init_MPI = true)...)
 else
