+++ miniapps/benchmarks/stokes2D/sinking_block/SinkingBlock2D.jl
@@ -197,8 +197,8 @@ end
 
 ar = 1 # aspect ratio
 n = 64
-nx = n * ar - 2
-ny = n - 2
+nx = n * ar
+ny = n
 igg = if !(JustRelax.MPI.Initialized()) # initialize (or not) MPI grid
     IGG(init_global_grid(nx, ny, 1; init_MPI = true)...)
 else
