### 2026-02-03 de8a44d
update and remove old convection codes
