+++ miniapps/deprecated/GlobalConvection2D_Upwind.jl
@@ -1,296 +1,0 @@
-using JustRelax, JustRelax.JustRelax2D, JustRelax.DataIO
-const backend_JR = CPUBackend
-
-using ParallelStencil, ParallelStencil.FiniteDifferences2D
-@init_parallel_stencil(Threads, Float64, 2) #or (CUDA, Float64, 2) or (AMDGPU, Float64, 2)
-
-using Printf, LinearAlgebra, GeoParams, CairoMakie, SpecialFunctions
-
-# function to compute strain rate (compulsory)
-@inline function custom_εII(a::CustomRheology, TauII; args...)
-    η = custom_viscosity(a; args...)
-    return TauII / η * 0.5
-end
-
-# function to compute deviatoric stress (compulsory)
-@inline function custom_τII(a::CustomRheology, EpsII; args...)
-    η = custom_viscosity(a; args...)
-    return 2.0 * η * EpsII
-end
-
-# helper function (optional)
-@inline function custom_viscosity(a::CustomRheology; P = 0.0, T = 273.0, depth = 0.0, kwargs...)
-    (; η0, Ea, Va, T0, R, cutoff) = a.args
-    η = η0 * exp((Ea + P * Va) / (R * T) - Ea / (R * T0))
-    correction = (depth ≤ 660.0e3) + (2740.0e3 ≥ depth > 660.0e3) * 1.0e1 + (depth > 2700.0e3) * 1.0e-1
-    return η = clamp(η * correction, cutoff...)
-end
-
-# HELPER FUNCTIONS ---------------------------------------------------------------
-
-import ParallelStencil.INDICES
-const idx_j = INDICES[2]
-macro all_j(A)
-    return esc(:($A[$idx_j]))
-end
-
-@parallel function init_P!(P, ρg, z)
-    @all(P) = @all(ρg) * abs(@all_j(z))
-    return nothing
-end
-
-# Half-space-cooling model
-@parallel_indices (i, j) function init_T!(T, z, κ, Tm, Tp, Tmin, Tmax)
-    yr = 3600 * 24 * 365.25
-    dTdz = (Tm - Tp) / 2890.0e3
-    zᵢ = abs(z[j])
-    Tᵢ = Tp + dTdz * (zᵢ)
-    time = 100.0e6 * yr
-    Ths = Tmin + (Tm - Tmin) * erf((zᵢ) * 0.5 / (κ * time)^0.5)
-    T[i, j] = min(Tᵢ, Ths)
-    return
-end
-
-function circular_perturbation!(T, δT, xc, yc, r, xvi)
-
-    @parallel_indices (i, j) function _circular_perturbation!(T, δT, xc, yc, r, x, y)
-        if (((x[i] - xc))^2 + ((y[j] - yc))^2) ≤ r^2
-            T[i + 1, j] *= δT / 100 + 1
-        end
-        return nothing
-    end
-    nx, ny = size(T)
-    return @parallel (1:(nx - 2), 1:ny) _circular_perturbation!(T, δT, xc, yc, r, xvi...)
-end
-
-function random_perturbation!(T, δT, xbox, ybox, xvi)
-
-    @parallel_indices (i, j) function _random_perturbation!(T, δT, xbox, ybox, x, y)
-        if (xbox[1] ≤ x[i] ≤ xbox[2]) && (abs(ybox[1]) ≤ abs(y[j]) ≤ abs(ybox[2]))
-            δTi = δT * (rand() - 0.5) # random perturbation within ±δT [%]
-            T[i + 1, j] *= δTi / 100 + 1
-        end
-        return nothing
-    end
-    nx, ny = size(T)
-    return @parallel (1:(nx - 2), 1:ny) _random_perturbation!(T, δT, xbox, ybox, xvi...)
-end
-
[…]
