### 2026-02-03 0675885
Update macOS version in CI workflow
